==> R/02-mordor-ab-format-data.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Format data for analysis"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Preamble

```{r preamble}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))
```

# Format the MBA antibody data

```{r load antibody data}
#-------------------------------
# load the raw MBA data sent from
# Diana Martin at CDC
#-------------------------------
d <- read_csv("~/Library/CloudStorage/Box-Box/mordor/mordor data/data/untouched-dbs/MORDOR MBA Raw Data Compiled 100219.csv")


#-------------------------------
# a couple of the DBS IDs include
# "Rerun" in the field. Cut this
# out of the ID
# trim all whitespace from the IDs
# just in case
#-------------------------------
d$description[grep(" Rerun", d$description)]
d$description <- gsub(" Rerun","",d$description)
d$description <- str_trim(d$description)

#-------------------------------
# pivot longer and create formatted
# antigen names and pathogen names
# careful / the antigen list
# and labels are in a specific order
#-------------------------------
antigens <- c("p18","p39",
              "etec","ctb",
              "salb","sald",
              "cp17","cp23",
              "vsp3","vsp5",
              "speb",
              "pfmsp1","ama","glurp","lsa1","csp","hrp2",
              "pmmsp1","pomsp1","pvmsp1",
              "pgp3","ct694",
              "nie",
              "gst"
              )
antigen_labs <- c(
  "Campylobacter p18",
  "Campylobacter p39",
  "ETEC LTB",
  "Cholera CTB",
  "Salmonella group B LPS",
  "Salmonella group D LPS",
  "Cryptosporidium Cp17",
  "Cryptosporidium Cp23",
  "Giardia VSP-3",
  "Giardia VSP-5",
  "Streptococcus group A SPEB",
  "P. falciparum MSP-1",
  "P. falciparum AMA1",
  "P. falciparum GLURP-Ro",
  "P. falciparum LSA1",
  "P. falciparum CSP",
  "P. falciparum HRP2",
  "P. malariae MSP-1",
  "P. ovale MSP-1",
  "P. vivax MSP-1",
  "Trachoma Pgp3",
  "Trachoma CT694",
  "Strongyloides NIE",
  "GST negative control"
)

d2 <- d %>%
  rename(dbsid=description) %>%
  pivot_longer(cols = c(-type,-well,-dbsid),
               names_to = "antigen", 
               values_to = "mfi") %>%
  mutate(antigenf = factor(antigen, levels=antigens, labels = antigen_labs)) %>%
  mutate(pathogen = case_when(
    antigen %in% c("p18","p39") ~ "Campylobacter",
    antigen %in% c("etec") ~ "ETEC",
    antigen %in% c("ctb") ~ "Cholera",
    antigen %in% c("salb","sald") ~ "Salmonella",
    antigen %in% c("cp17","cp23") ~ "Cryptosporidium",
    antigen %in% c("vsp3","vsp5") ~ "Giardia",
    antigen %in% c("speb") ~ "Streptococcus",
    antigen %in% c("pfmsp1","ama","glurp","lsa1","csp","hrp2") ~ "P. falciparum",
    antigen %in% c("pmmsp1") ~ "P. malariae",
    antigen %in% c("pomsp1") ~ "P. ovale",
    antigen %in% c("pvmsp1") ~ "P. vivax",
    antigen %in% c("nie") ~ "Strongyloides",
    antigen %in% c("pgp3","ct694") ~ "Trachoma",
    antigen %in% c("gst") ~ "Neg control"
  ) ) %>%
    mutate(pathogen = factor(pathogen,levels=c("Campylobacter","ETEC","Cholera","Salmonella","Cryptosporidium","Giardia","Streptococcus","P. falciparum","P. malariae","P. ovale","P. vivax","Strongyloides","Trachoma","Neg control"))
           ) %>%
  dplyr::select(dbsid,type,well,pathogen,antigen,antigenf,mfi)
        
```

Summarize the data
```{r summarize mba data}
summary(d2)

```

# Format morbidity data

Read in the final morbidity dataset shared by Jermey Keenan in December 2019.  Victoria Li originally wrote the code below, spliced into this Rmd file.

```{r format morbidity}
#-------------------------------
# read in the full morbidity file
#-------------------------------
md <- read_csv("~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-morbidity/MORDOR-Morbidity-merged-20Mar2019.csv",
                      col_types = cols(code_bspot_ = col_character(),
                                       arm = col_character(),
                                       txlabelgwu = col_character(),
                                       txdate = col_character(),
                                       dosex = col_character()))

#-------------------------------
# merge on the public IDs
# and drop study IDs to protect
# PII for public-facing data
#
# doing a separate match against
# cluster IDs and child IDs 
# because there are a small number
# of DBS specimens that are linked
# to cluster but not child IDs
# and can still be included in 
# the analysis
#-------------------------------
public_clids <- read_csv("~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-morbidity/public-IDs/mordor-morbidity-public-IDs-cluster.csv")

public_ids <- read_csv("~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-morbidity/public-IDs/mordor-morbidity-public-IDs-child.csv") %>%
  select(-gwu, -clusterid_public)

md2 <- public_clids %>%
  left_join(md, by = c("gwu")) %>%
  left_join(public_ids, by = c("masterperson_mcd"))


#-------------------------------
# Select relevant variables 
# rename variables to be more
# convenient
#-------------------------------
md3 <- md2 %>% 
  dplyr::select(dbsid=code_bspot_,
         childid_public,
         clusterid_public,
         gwu,
         masterperson_mcd,
         arm,
         txlabelgwu,
         studypop = studypop_,
         phase = phaseidmep,
         testdate = testdatets_bspot_,
         trdate = txdate,
         dose = dosex, 
         sex = gendermcd_,
         malaria_thicksmear = thicksmear_cons_bsmear_,
         malaria_density = officialdensity_bsmear_,
         malaria_gameto = gameto_bsmear_,
         contains("age")
         ) %>% 
  mutate(sex = factor(sex,levels=c("FEMALE","MALE"),labels = c("female","male")),
         dbsid = str_trim(dbsid),
         testdate = as.Date(testdate, "%d%b%Y"),
         trdate = as.Date(trdate, "%d%b%Y")
         )
  
#-------------------------------
# for each community and phase
# create a median treatment
# date
#-------------------------------
md3 <- md3 %>%
  group_by(clusterid_public,phase) %>%
  mutate(med_trdate = median(trdate,na.rm = TRUE))

#-------------------------------
# there are many age variables
# do some reconciliation to
# harmonize and create a single
# child age
#
# use agecc_mcd as the reference
# age, but for children with
# (a) missing values
# (b) values < 0
# (c) values > 59
# replace it with values from
# agemosblood_
#-------------------------------
md4 <- md3 %>%
  mutate(agem = agecc_mcd,
         agem = ifelse(is.na(agem),agemosblood_,agem),
         agem = ifelse(agem<0 & agemosblood_ >=0,agemosblood_, agem),
         agem = ifelse(agem>59 & agemosblood_ <=agem,agemosblood_, agem))
```

# Merge DBS with morbidity data

Merge the two datasets

```{r merge antibody and morbidity data}
#-------------------------------
# join to the DBS measures
#-------------------------------
d3 <- d2 %>%
  left_join(md4,by = "dbsid")

#-------------------------------
# There are 53 / 5699 (0.9%) 
# of the DBS specimen results
# that do not match to the
# morbidity file
# print the DBS IDs and 
# then put to the side for now
#-------------------------------
dbs_nomatch <- d3 %>%
  filter(is.na(clusterid_public)) %>%
  dplyr::select(dbsid,antigen,mfi) %>%
  pivot_wider(id_cols = dbsid, names_from = "antigen", values_from = "mfi")

d4 <- d3 %>%
  filter(!is.na(clusterid_public))

# do not save a file for now.
# instead, print to HTML notebook
# write_csv(dbs_nomatch,"~/Library/CloudStorage/Box-Box/mordor-ab/data/temp/mordor-ab-dbs-nomatch.csv")
dbs_nomatch


#-------------------------------
# There are 4 children who have
# ages > 60 months
# exclude them
#-------------------------------
table(d4$agem[d4$antigen=="vsp3"]>60)
d4 <- d4 %>%
  filter(agem <= 60)

#-------------------------------
# there are 19 children who
# have matches between the
# morbidity file and the DBS
# data, but do not have a unique
# childid (masterperson_mcd is missing)
#
# for these children assign
# a random ID. Also ensure
# the age used is the agemosblod_
# variable (others are missing)
#-------------------------------
table(is.na(d4$childid_public[d4$antigen=="vsp3"]))
newids <- data.frame(dbsid = d4$dbsid[is.na(d4$childid_public) & d4$antigen=="vsp3"],childid_recode = paste0("childid recode ",1:19),stringsAsFactors = FALSE)
d4 <- d4 %>%
  left_join(newids, by = "dbsid") %>%
  mutate(
    agem = ifelse(is.na(childid_public), agemosblood_, agem),
    childid_public = ifelse(is.na(childid_public),childid_recode,childid_public)
    )

#-------------------------------
# For each study cluster, identify
# the earliest treatment date in
# each phase, and calculate
# the days between treatments
# merge back to the main dataset
#-------------------------------
d_firsttr <- d4 %>%
  group_by(clusterid_public, phase) %>%
  summarize(mintrdate = min(trdate,na.rm = TRUE), .groups = "keep")
# community-24 (public ID) is missing 
# treatment information in phase 6
# assume it was 183 days later than phase 0
d_firsttr$mintrdate[d_firsttr$clusterid_public=="community-24" & d_firsttr$phase == 6] <- as.Date("2015-07-06")+183
d_firsttr <- d_firsttr %>%
  group_by(clusterid_public) %>%
  mutate(firsttr = ifelse(phase == 0, mintrdate, NA),
         firsttr = as.Date(max(firsttr,na.rm = TRUE),origin ="1970-01-01"),
         diff_phase0 = mintrdate - firsttr,
         diff_phase = mintrdate - lag(mintrdate),
         diff_phase = ifelse(phase == 0, 0, diff_phase)) 

d4 <- left_join(d4,d_firsttr, by = c("clusterid_public","phase"))


#-------------------------------
# Identify children who were
# less than 6 months old
# at the beginning of the trial
# defined as the first treatment
# date for their cluster
#-------------------------------
d4 <- d4 %>%
  mutate(agem_start = as.numeric(agem - diff_phase0/30.4167),
         agem_start_under6 = ifelse(agem_start <=6, "Yes","No"),
         agem_start_under6 = factor(agem_start_under6))

```

# Merge in wet season information

From the monthly total rainfall data, merge by clusterid and year the start and end dates of the wet season.

```{r merge in wet season}
#-------------------------------
# load monthly precipitation
# data, created by
# mordor-ab-format-data-gps-climate.Rmd
#-------------------------------
d_pre <- read_csv(here("data","mordor-ab-monthly-precip.csv"))

d_pre2 <- d_pre %>%
  group_by(clusterid_public,pyear) %>%
  dplyr::select(clusterid_public,pyear,wet_start,wet_end) %>%
  slice(1) %>%
  ungroup()

#-------------------------------
# for the MORDOR data, use each cluster's median
# test date.  merge by year 
# (since wet seasons do not straddle calendar years)
#-------------------------------
d4 <- d4 %>%
  group_by(clusterid_public,phase) %>%
  mutate(medtestdate = as_date(median(testdate)),
         pyear = year(medtestdate)
  ) %>%
  left_join(d_pre2, by = c("clusterid_public","pyear")) %>%
  dplyr::select(-pyear)

#-------------------------------
# summarize wet season begin and 
# end dates for each phase
#-------------------------------
d4 %>% group_by(phase) %>% summarize(wet_start = as_date(min(wet_start), .groups = "keep"), wet_end = as_date(max(wet_end)))

#-------------------------------
# identify wet season observations
#-------------------------------
d4 <- d4 %>%
  mutate(season = ifelse(testdate >= wet_start & testdate <= wet_end, "wet season","dry season"),
         season = factor(season))

```

# Re-randomize for masking

Re-randomize community assignments. The morbidity file includes masked assignments (sun, moon) but they are not re-randomized. Save the real assignments on the side and use re-randomized labels until the main analyses are all finished. Then, we will unmask by updating this script to merge in the real assignments.

As of September 3, 2020, this code is no longer evaluated. The analysis was unmasked 2020-09-03 at 16:45.

```{r re-randomize community assignments, eval = FALSE}
#-------------------------------
# store the real treatment
# assignments, put to the side
#-------------------------------
comd <- d4 %>%
  ungroup() %>%
  dplyr::select(clusterid_public,arm) %>%
  group_by(clusterid_public) %>%
  slice(1)
write_csv(comd,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-tr.csv")

#-------------------------------
# create a set of re-randomized
# treatment labels
#-------------------------------
set.seed(12345)
comd_rr <- comd %>%
  dplyr::select(clusterid_public) %>%
  ungroup() %>%
  mutate(arm = sample(rep(c("sun","moon"),15),size=30,replace=FALSE))

#-------------------------------
# merge re-randomized (shuffled)
# labels to the antibody data
#-------------------------------
d5 <- d4 %>%
  dplyr::select(-arm) %>%
  left_join(comd_rr,by="clusterid_public") %>%
  mutate(arm = factor(arm,levels=c("sun","moon")))

```

```{r unmaske community assignments}
d5 <- d4 %>%
  mutate(arm = ifelse(arm == "sun", "azithro","placebo"),
         arm = factor(arm, levels = c("placebo","azithro"))
         )

```


# Seropositivity cutoffs

Create indicators of seropositivity for each antigen.  For antigens without cutoffs derived from external specimens (ROC, unexposed populations) use Gaussian mixture models.  For enterics, restrict the population to very young children to ensure sufficient differentiation between seronegative and seropositive subpopulations, following our previous work (https://www.ncbi.nlm.nih.gov/pubmed/31424386).


## ROC-based cutoffs
```{r external seropositivity cutoffs}
#-----------------------------
# add external seropositivity
# cutoffs, shared by CDC in:
# MORDOR Antigen Info 110419.docx
#-----------------------------
d6 <- d5 %>%
  mutate(excut = case_when(
    antigen == "cp17" ~ 290,
    antigen == "cp23" ~ 945,
    antigen == "vsp3" ~ 154,
    antigen == "vsp5" ~ 147,
    antigen == "hrp2" ~ 5336,
    antigen == "csp"  ~ 926,
    antigen == "glurp"~ 123,
    antigen == "lsa1" ~ 447,
    antigen == "ama"  ~ 646,
    antigen == "pfmsp1" ~ 221,
    antigen == "pvmsp1" ~ 787,
    antigen == "pomsp1" ~ 462,
    antigen == "pmmsp1" ~ 431,
    antigen == "pgp3" ~ 1613,
    antigen == "ct694" ~ 457,
    antigen == "nie" ~ 3523,
    TRUE ~ NA_real_
  )
         
         )
```

## Mixture model cutoffs


Work with log10 values
```{r log10 mfi, warning=FALSE}
#-----------------------------
# work with log10 MFI values
# for MFI values <=0, set to 1
# before the log transform
#-----------------------------
table(d6$antigenf,d6$mfi<=0)
d7 <- d6 %>%
  mutate(logmfi = ifelse(mfi<=0,log10(1),log10(mfi)))

```


Plot antibody distributions

```{r antibody distributions, fig.width=12, fig.height=12}

pcols <- cbpal


# custom log labels
log10labs <- c( 
  expression(10^0),
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)

p <- ggplot(data=d7 %>% filter(agem<=12 & logmfi>0),aes(x=logmfi)) +
  facet_wrap(~antigenf,nrow=6,ncol=4,scales="free_y") +
  geom_line(stat = "density")+
  scale_x_continuous(limits = c(0,4.5),breaks = 0:4,labels=log10labs) +
  # scale_fill_manual(values=pcols) +
  # scale_color_manual(values=pcols) +
  labs(x="Luminex Response (MFI-bg)") +
  theme_minimal(base_size=16) +
  theme(legend.position = "none")

p

```


```{r mixture fit function}
#-----------------------------
# mixture model fit function
#-----------------------------
fitmix <- function(x,lambda,k) {
  mmfit <- normalmixEM(x,lambda=lambda,k=k)
  mmcut <- mmfit$mu[order(mmfit$mu)][1]+3*mmfit$sigma[order(mmfit$mu)][1]
  cat(summary(mmfit),"\nCutoff value:",mmcut,"(log10 MFI), or:",
      round(10^mmcut),"(MFI)\n\n")
  # pull out fitted densities
  denmat <- matrix(NA,nrow=length(x),ncol=k)
  denord <- order(mmfit$mu)
  for(i in 1:k) {
    j <- denord[i]
    denmat[,i] <- mmfit$lambda[j] * dnorm(x,mean=mmfit$mu[j],sd=mmfit$sigma[j])
  }
  denmat <- data.frame(denmat)
  colnames(denmat) <- paste("den",1:k,sep="")
  # return original values plus fitted densities in a dataframe 
  # also return the cutoff value and normalmixEM object
  xden <- data.frame(x=x,denmat)
  list(xden=xden,mmcut=mmcut,mmfit=mmfit)
  
}

```



```{r mixture model seropositivity cutoffs}
#-----------------------------
# mixture model fits
# for enterics
#
# limited to children <12 mos
#
# store densities of mixture distributions
# and estimated cutoffs
#
# the NIE model does not converge
# so exclude
#-----------------------------
dmmcut <- d7 %>%
  filter(agem <= 12 & logmfi>0) %>%
  filter(antigen %in% c("p18","p39","etec","ctb","sald","salb","cp17","cp23","vsp3","vsp5","speb")) %>%
  mutate(antigenf = factor(antigenf))


set.seed(1234)
mixdens <- foreach(ab=levels(dmmcut$antigenf),.combine=rbind) %do% {
  cat("\n\n Mixture model fit for ",ab,"\n\n")
  mixf <- fitmix(x=dmmcut$logmfi[dmmcut$antigenf==ab],lambda=0.5,k=2)
  di <- mixf$xden
  di$mmcut <- mixf$mmcut
  di$antigenf=ab
  di <- di %>% arrange(x) %>% dplyr::select(antigenf,mmcut,everything())
  di
}

mmcuts <- mixdens %>%
  group_by(antigenf) %>% 
  filter(row_number()==1) %>% dplyr::select(antigenf,mmcut) %>% ungroup() %>%
  mutate(antigenf = factor(antigenf,levels=levels(d7$antigenf)))


#-----------------------------
# merge in the mixture cutoffs
# to the main dataset
#-----------------------------
d8 <- left_join(d7,mmcuts,by=c("antigenf"))


```



Figure of antibody mixture distributions

```{r Ab mixture distribution figure, fig.height=12, fig.width=12}
#-----------------------------
# Get the mixture density distributions
# set the pathogen field to an arbitrary value (Giardia) to 
# mesh with the global ggplot aesthetics
#-----------------------------
plotmixdens <- mixdens %>%
  distinct() %>%
  mutate(pathogen="Giardia",
         antigenf = factor(antigenf,levels=levels(d8$antigenf)))

#-----------------------------
# and custom log10 labels (log10labs)
# defined in a previous code chunk
#-----------------------------
# pcols <- c(cred,corange,cgreen,cteal,cblue,cgrey,cmagent)
# pcols <- c(corange,cred,cmagent,cblue,cteal,cgreen,cgrey)
# pcols <- vircols

pmix <- ggplot(data = d8 %>% filter(agem <= 12 & logmfi>0), aes(x=logmfi)) +
  facet_wrap(~antigenf,nrow=6,ncol=4,scales="free_y") +
  # plot empirical smoothed distribution
  geom_density(color=NA,fill=vircols[3],alpha=0.4) +
  geom_rug(color = "gray20",alpha = 0.2) +
  # plot fitted mixture distributions
  geom_line(data=plotmixdens,aes(x=x,y=den1), color = "gray40") +
  # add vertical lines for the cutoff values
  geom_vline(aes(xintercept=log10(excut))) +
  geom_vline(aes(xintercept=mmcut),linetype=2) +
  # labels and formatting
  scale_x_continuous(limits = c(0,4.5),breaks = 0:4,labels=log10labs) +
  # coord_cartesian(ylim=c(0,2)) +
  labs(x="Luminex Response (MFI-bg)", title = "Antibody distributions with mixture model cutoffs") +
  # scale_fill_manual(values=pcols) +
  # scale_color_manual(values=pcols) +
  theme_minimal(base_size = 16) +
  theme(legend.position = "none")

pmix


```

The mixture model fits look very reasonable for _Cryptosporidium_, _Giardia_, and _Salmonella_ but do not look reasonable for the other bacterial pathogens. 


## Cutoffs based on presumed unexposed

Try something more complicated.  For children with longidudinal measurements, identify those who had a 100-fold increase in MFI. Fit the distribution among their original MFI levels.

```{r presumed unexposed}
#-------------------------------
# identify children with
# longitudinal measurements
# subset to them
#-------------------------------
dl <- d8 %>%
  ungroup() %>%
  group_by(childid_public,antigen) %>%
  mutate(nobs = n(),
         obsnum = row_number()) %>%
  filter(nobs>1 & nobs < 6)

#-------------------------------
# further subset to children 
# who were < 12 months at their
# first measurement
#-------------------------------
dl2 <- dl %>%
  group_by(childid_public,antigen) %>%
  mutate(first_young = ifelse(agem < 12 & obsnum < nobs, 1,0),
         first_young = max(first_young)) %>%
  filter(first_young == 1) 


#-------------------------------
# calculate the change in MFI
# between measurements
#-------------------------------
dl3 <- dl2 %>%
  group_by(childid_public,antigen) %>%
  arrange(antigen,childid_public,phase) %>%
  mutate(dlogmfi = lead(logmfi)-logmfi)

#-------------------------------
# restrict to observations where
# the MFI subsequently increased
# by at least 2 (on log10 scale)
# equivalent to a 100-fold increase
#-------------------------------
dl4 <- dl3 %>%
  filter(dlogmfi>2) %>%
  filter(antigen %in% c("p18","p39","etec","ctb","salb","sald","speb")) %>%
  ungroup() %>%
  group_by(antigen) %>%
  mutate(meanmfi = mean(logmfi),
         sdmfi = sd(logmfi),
         pucut = meanmfi + 3*sdmfi) %>%
  dplyr::select(-meanmfi,sdmfi)

table(dl4$antigen)

#-------------------------------
# merge the presumed unexposed
# cutoffs back to the main data
#-------------------------------
pucuts <- dl4 %>%
  dplyr::select(antigen,pucut) %>%
  group_by(antigen,pucut) %>%
  slice(1)

d9 <- d8 %>%
  left_join(pucuts,by="antigen")
  
```

Figure of distributions among presumed unexposed
```{r plot presumed unexposed distribution}
#-------------------------------
# plot distribution
# among presumed unexposed
#-------------------------------

punexp <- ggplot(data=dl4, aes(x = logmfi)) + 
  facet_wrap(~antigenf,nrow=4,ncol=2,scales="free_y") +
  # plot empirical smoothed distribution
  geom_density(color=NA,fill=vircols[3],alpha=0.4) +
  geom_rug(color = "gray20",alpha = 0.5) +

  # add vertical lines for the cutoff values
  geom_vline(aes(xintercept=pucut)) +
  # geom_vline(aes(xintercept=mmcut),linetype=2) +
  # labels and formatting
  scale_x_continuous(limits = c(0,4.5),breaks = 0:4,labels=log10labs) +
  # coord_cartesian(ylim=c(0,2)) +
  labs(x="Luminex Response (MFI-bg)",title="Distribution among presumed unexposed") +
  # scale_fill_manual(values=pcols) +
  # scale_color_manual(values=pcols) +
  theme_minimal(base_size = 16) +
  theme(legend.position = "none")

punexp

```

Figure of cutoff values selected for each antigen (based on external samples or presumed unexposed)
```{r cutoff distribution figure, fig.height=12, fig.width=12}

pdens <- ggplot(data = d9 %>% filter(agem <= 12 & logmfi>0), aes(x=logmfi)) +
  facet_wrap(~antigenf,nrow=6,ncol=4,scales="free_y") +
  # plot empirical smoothed distribution
  geom_density(color=NA,fill=vircols[3],alpha=0.4) +
  geom_rug(color = "gray20",alpha = 0.2) +
  # add vertical lines for the cutoff values
  geom_vline(aes(xintercept=log10(excut))) +
  geom_vline(aes(xintercept=pucut),linetype=2) +
  # labels and formatting
  scale_x_continuous(limits = c(0,4.5),breaks = 0:4,labels=log10labs) +
  # coord_cartesian(ylim=c(0,2)) +
  labs(x="Luminex Response (MFI-bg)", title = "Antibody distributions with cutoffs") +
  theme_minimal(base_size = 16) +
  theme(legend.position = "none")

pdens


```

```{r final seropositivity cutoffs}
d10 <- d9 %>%
  mutate(serocut = ifelse(is.na(excut),round(10^pucut,0),excut),
         serocut_desc = case_when(
           !is.na(excut) ~ "external",
           is.na(excut) & antigen != "gst" ~ "presumed unexp",
           TRUE ~ as.character(NA)
         ),
         seropos = ifelse(mfi>serocut,1,0))
```


# Save analysis data

```{r save analysis data}


#-------------------------------
# select variables, summarize
# and save data
#-------------------------------
d11 <- d10 %>%
  dplyr::select(gwu,clusterid_public,phase,arm,
                masterperson_mcd,childid_public,agem,agem_start, agem_start_under6, sex, 
         testdate,
         wet_start,wet_end,season,
         malaria_thicksmear,malaria_density,malaria_gameto,
         dbsid,type,well,pathogen,antigen,antigenf,
         mfi,logmfi,serocut,serocut_desc,seropos)
write_rds(d11,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-analysis.rds")
write_csv(d11,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-analysis.csv")

#-------------------------------
# save a public version of the 
# data to share
# dropping study IDs 
# 
# recode dates to be accurate
# only to year and month by
# setting the day to 15 of 
# each month for every date
#-------------------------------
d12 <- d11
day(d12$testdate)    <- 15
day(d12$wet_start) <- 15
day(d12$wet_end)   <- 15
d12 <- d12 %>%
  select(-gwu, -masterperson_mcd) %>%
  mutate(antigenf = factor(antigenf))
summary(d12)

write_rds(d12,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-analysis-public.rds")
write_csv(d12,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-analysis-public.csv")

```

# Format treatment data (timing, coverage)

Treatment coverage data are stored in a separate file from the morbidity data. Here we read in the treatment data and format it for analysis.

```{r read in treatment data}
#-------------------------------
# read in the raw file
#-------------------------------
dtr <-read_csv("~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-morbidity/Stata-mordor-Niger-long-dedupe-forYing.csv") %>%
  rename(masterperson_mcd = masterperson)

#-------------------------------
# merge on the public IDs
# and drop study IDs to protect
# PII for public-facing data
#-------------------------------
public_ids <- read_csv("~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-morbidity/public-IDs/mordor-morbidity-public-IDs-child.csv")
dtr2 <- dtr %>%
  left_join(public_ids, by = c("masterperson_mcd"))

#-------------------------------
# limit data to morbidity trial
# communities (mortalityonly999)
# and to children eligible for
# treatment (eligx)
# 
# format a few variables then
# limit to relevant vars
#-------------------------------
dtr2 <- dtr2 %>%
  filter(mortalityonly999 == 0) %>%
  filter(eligx == 1) %>%
  # identify children who were treated based on txstatusx
  mutate(treated = ifelse(txstatusx == 1, 1, 0)) %>%
  # convert treatment date from Stata dates into R
  mutate(trdate = as.Date(newccdatex, origin = "1960-01-01")) %>%
  # limit to relevant vars
  dplyr::select(gwu = randomunit999, clusterid_public,childid_public,dupwpid_a, phase, agem = agex, female = femalex, trdate, treated, dose=dosex)

#-------------------------------
# then fill in cluster level
# public IDs for any children
# not in the morbidity dataset
#-------------------------------
dtr3 <- dtr2 %>%
  group_by(gwu) %>%
  tidyr::fill(clusterid_public, .direction = "downup") %>%
  ungroup() 

#-------------------------------
# merge on treatment assignments
#-------------------------------
dtrarm <- d11 %>%
  ungroup() %>%
  select(clusterid_public, arm) %>%
  group_by(clusterid_public, arm) %>%
  slice(1)
dtr4 <- dtr3 %>%
  left_join(dtrarm, by = "clusterid_public")

#-------------------------------
# restrict to relevant variables
#
# recode dates to be accurate
# only to year and month by
# setting the day to 15 of 
# each month for every date
#-------------------------------
dtr4 <- dtr4 %>%
  dplyr::select(clusterid_public,arm,phase,childid_public,agem,female,trdate,treated)
day(dtr4$trdate)    <- 15
```

## Save the treatment dataset

```{r save treatment data}
#-------------------------------
# summarize and save the treatment data
#-------------------------------
summary(dtr4)
write_rds(dtr4,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-treatment-public.rds")
write_csv(dtr4,"~/Library/CloudStorage/Box-Box/mordor/mordor data/data/final-dbs/mordor-ab-treatment-public.csv")

```

# Session Info
```{r session info}
sessionInfo()
```
