==> R/05-mordor-ab-descriptive-summary.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Descriptive summary of antibody distributions and age patterns"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Summarize age-dependent mean antibody levels and distributions. 

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
# drop the GST negative control values
# remove trachoma and strongy
# because no transmission in this
# population
#-------------------------------
d2 <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public) %>%
  filter(!antigen %in% c("gst","pgp3","ct694","nie")) %>%
  mutate(antigenf = factor(antigenf))
```


# Estimate age-dependent means

```{r age-dependent means}
#-----------------------------
# create a dummy indicator equal
# to 1 for all communities to zero
# out the marginal predictions
# over all community random effects
#
# create a factor variable
# for clusterid (for model)
#-----------------------------
d3 <- d2 %>%
  mutate(dummy = 1,
         clusteridf = factor(clusterid)
         )

#----------------------------------------
# fit a GAM model fitting a separate
# spline by age for each group
# estimate simulatenous 95% confidence 
# intervals
#----------------------------------------

fit_age <- foreach(abi = levels(d3$antigenf), .combine = rbind) %do% {
  cat("\n Estimating splines for",abi)
  di <- d3 %>% filter(antigenf == abi)
  fit_age <- mgcv::gam(logmfi ~ s(agem,bs="cr") + 
                         s(clusteridf,bs="re",by=dummy), 
                       data=di)
  newd <- di %>%  mutate(dummy=0)
  fit_ci <- gamCI(m=fit_age,newdata=newd,nreps=1000)
}

```


# Bacteria and protozoa figure

```{r age-dependent means figure enterics, fig.width = 6, fig.height = 12}
#-----------------------------
# plot age-dependent means
#-----------------------------

# subset to enterics and strep
fit_age_enterics <- fit_age %>%
  filter(antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","salb","sald","etec","ctb","speb"))


# cutoffs 
dcuts <- fit_age_enterics %>%
  select(pathogen, antigen, antigenf,serocut) %>%
  group_by(pathogen, antigen, antigenf) %>%
  slice(1) %>%
  mutate(serocut = log10(serocut))

# custom log labels
log10labs <- c( 
  expression(10^0),
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)
pcol <- vircols[c(3)]
plot_age_mfi_enterics <- ggplot(data = fit_age_enterics, aes(x = agem, y = fit)) +
  # facet by pathogen and antigen
  facet_grid(antigenf~.) +
  # seropositivity cutoffs
  geom_hline(data = dcuts, aes(yintercept = serocut), lwd = 0.25,col = "gray20") +
  # MFI values
  geom_jitter(aes(y = logmfi), width=0.3,height=0, color = pcol, alpha=0.1, size =0.02) +
  # 95% CI bands for spline fit
  geom_ribbon(aes(ymin = lwrS, ymax = uprS), fill = vircols[4], color = NA, alpha = 0.4) +
  # spline fit
  geom_line(color = vircols[4]) +
  
  # plot aesthetics
  scale_y_continuous(breaks = 0:4, labels = log10labs) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=6)) +
  coord_cartesian(ylim = c(0,4.5)) +
  labs(x = "Age, months", y = "Luminex response (MFI-bg)", title = "Age-dependent mean response", tag = "a") +
  theme(
    strip.text.y = element_text(angle = 0),
    plot.tag = element_text(face = "bold")
  )


```


```{r distribution enterics ages 0-12 mo}
#-----------------------------
# plot distribution among
# ages 0-6 months
#-----------------------------
hbreaks <- seq(0,4.6,by=0.1)
plot_dist_under6mo <- ggplot(data = fit_age_enterics %>% filter(agem<6), aes(x = logmfi)) +
  facet_grid(antigenf~.) +
  geom_histogram(aes(x = logmfi), breaks = hbreaks, color = NA, fill=pcol, alpha = 0.8)+
  geom_vline(data = dcuts, aes(xintercept = serocut),lwd = 0.25,col = "gray20") +
  scale_x_continuous(breaks = 0:4, labels = log10labs) +
  scale_y_continuous(expand = c(0,0), breaks = seq(0,700,by=100))+
  labs(x = NULL, y = "N observations", title = "Ages 0-6 mo", tag = "b") +
  coord_flip(ylim = c(0,400), xlim = c(0,4.5))+
  theme(
    legend.position = "none", 
    strip.text.y = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r distribution enterics ages 12-24 mo}
#-----------------------------
# plot distribution among
# ages 6-24 months
#-----------------------------
hbreaks <- seq(0,4.6,by=0.1)
plot_dist_6to24 <- ggplot(data = fit_age_enterics %>% filter(agem>=6 & agem<=24), aes(x = logmfi)) +
  facet_grid(antigenf~.) +
  geom_histogram(aes(x = logmfi), breaks = hbreaks, color = NA, fill=pcol, alpha = 0.8)+
  geom_vline(data = dcuts, aes(xintercept = serocut),lwd = 0.25,col = "gray20") +
  scale_x_continuous(breaks = 0:4, labels = log10labs) +
  scale_y_continuous(expand = c(0,0), breaks = seq(0,700,by=100))+
  labs(x = NULL, y = "N observations", title = "Ages 6-24 mo", tag = "c") +
  coord_flip(ylim = c(0,700), xlim = c(0,4.5))+
  theme(
    legend.position = "none", 
    strip.text.y = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```


```{r distribution enterics 24-59}
#-----------------------------
# plot distribution among
# ages 24-59 months
#-----------------------------
hbreaks <- seq(0,4.6,by=0.1)
plot_dist_24to59 <- ggplot(data = fit_age_enterics %>% filter(agem>24), aes(x = logmfi)) +
  facet_grid(antigenf~.) +
  geom_histogram(aes(x = logmfi), breaks = hbreaks, color = NA, fill=pcol, alpha = 0.8)+
  geom_vline(data = dcuts, aes(xintercept = serocut),lwd = 0.25,col = "gray20") +
  scale_x_continuous(breaks = 0:4, labels = log10labs) +
  scale_y_continuous(expand = c(0,0), breaks = seq(0,700,by=100))+
  labs(x = NULL, y = "N observations", title = "Ages 24-59 mo", tag = "d") +
  coord_flip(ylim = c(0,700), xlim = c(0,4.5))+
  theme(
    legend.position = "none", 
    strip.text.y = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r enterics age plot composite, warning=FALSE, fig.width = 12, fig.height = 16}
#-----------------------------
# stitch the panels together
# into a composite figure
#-----------------------------
pcurve_composite_enterics <- grid.arrange(plot_age_mfi_enterics,plot_dist_under6mo,plot_dist_6to24,plot_dist_24to59,ncol=4,nrow=1,widths=c(2.5,1.5,1.5,1.5))

ggsave(here("figures","mordor-ab-antibody-distributions-bacteria-protozoa.pdf"),pcurve_composite_enterics, device = "pdf", width = 12, height = 20)

```

# Malaria figure


```{r age-dependent means figure malaria}
#-----------------------------
# plot age-dependent means
#-----------------------------

# subset to malaria antigens
fit_age_malaria <- fit_age %>%
  filter(!antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","salb","sald","etec","ctb","speb")) %>%
  mutate(antigenf = factor(antigenf))


# cutoffs 
dcuts_malaria <- fit_age_malaria %>%
  select(pathogen, antigen, antigenf,serocut) %>%
  group_by(pathogen, antigen, antigenf) %>%
  slice(1) %>%
  mutate(serocut = log10(serocut))

# custom log labels
log10labs <- c( 
  expression(10^0),
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)
pcol <- vircols[c(3)]
plot_age_mfi_malaria <- ggplot(data = fit_age_malaria, aes(x = agem, y = fit)) +
  # facet by pathogen and antigen
  facet_grid(antigenf~.) +
  # seropositivity cutoffs
  geom_hline(data = dcuts_malaria, aes(yintercept = serocut), lwd = 0.25,col = "gray20") +
  # MFI values
  geom_jitter(aes(y = logmfi), width=0.3,height=0, color = pcol, alpha=0.1, size =0.02) +
  # 95% CI bands for spline fit
  geom_ribbon(aes(ymin = lwrS, ymax = uprS), fill = vircols[4], color = NA, alpha = 0.4) +
  # spline fit
  geom_line(color = vircols[4]) +
  
  # plot aesthetics
  scale_y_continuous(breaks = 0:4, labels = log10labs) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=6)) +
  coord_cartesian(ylim = c(0,4.5)) +
  labs(x = "Age, months", y = "Luminex response (MFI-bg)", title = "Age-dependent mean response", tag = "a") +
  theme(
    strip.text.y = element_text(angle = 0),
    plot.tag = element_text(face = "bold")
  )


```


```{r distribution malaria ages 0-12 mo}
#-----------------------------
# plot distribution among
# ages 0-12 months
#-----------------------------
hbreaks <- seq(0,4.6,by=0.1)
plot_dist_malaria_under12mo <- ggplot(data = fit_age_malaria %>% filter(agem<12), aes(x = logmfi)) +
  facet_grid(antigenf~.) +
  geom_histogram(aes(x = logmfi), breaks = hbreaks, color = NA, fill=pcol, alpha = 0.8)+
  geom_vline(data = dcuts_malaria, aes(xintercept = serocut),lwd = 0.25,col = "gray20") +
  scale_x_continuous(breaks = 0:4, labels = log10labs) +
  scale_y_continuous(expand = c(0,0), breaks = seq(0,700,by=100))+
  labs(x = NULL, y = "N observations", title = "Ages 0-12 mo", tag = "b") +
  coord_flip(ylim = c(0,500), xlim = c(0,4.5))+
  theme(
    legend.position = "none", 
    strip.text.y = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```


```{r distribution malaria 12-59 mo}
#-----------------------------
# plot distribution among
# ages 24-59 months
#-----------------------------
hbreaks <- seq(0,4.6,by=0.1)
plot_dist_malaria_12to59 <- ggplot(data = fit_age_malaria %>% filter(agem>=12), aes(x = logmfi)) +
  facet_grid(antigenf~.) +
  geom_histogram(aes(x = logmfi), breaks = hbreaks, color = NA, fill=pcol, alpha = 0.8)+
  geom_vline(data = dcuts_malaria, aes(xintercept = serocut),lwd = 0.25,col = "gray20") +
  scale_x_continuous(breaks = 0:4, labels = log10labs) +
  scale_y_continuous(expand = c(0,0), breaks = seq(0,700,by=100))+
  labs(x = NULL, y = "N observations", title = "Ages 12-59 mo", tag = "c") +
  coord_flip(ylim = c(0,700), xlim = c(0,4.5))+
  theme(
    legend.position = "none", 
    strip.text.y = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r malaria age plot composite, warning=FALSE, fig.width = 12, fig.height = 16}
#-----------------------------
# stitch the panels together
# into a composite figure
#-----------------------------
pcurve_composite_malaria <- grid.arrange(plot_age_mfi_malaria,plot_dist_malaria_under12mo,plot_dist_malaria_12to59,ncol=3,nrow=1,widths=c(2.5,1.5,1.5))

ggsave(here("figures","mordor-ab-antibody-distributions-malaria.pdf"),pcurve_composite_malaria, device = "pdf", width = 10, height = 16)

```

# Session Info
```{r session info}
sessionInfo()
```
