==> R/07-mordor-ab-pairwise-ab-correlation.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Assess pairwise correlation between antibody responses"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Antibody responses to ETEC labile toxin B subnit (LTB) and cholera toxin B subunit are known to be cross-reactive. Confirm this is the case in these measurements. Based on the results, exclude cholera CTB from the main results because it's difficult to interpret without additional antibody markers.

Arnold BF, et al. Enteropathogen antibody dynamics and force of infection among children in low-resource settings. _eLife_ 8, (2019). https://pubmed.ncbi.nlm.nih.gov/31424386/

Smith NW, Sack RB. 1973. Immunologic cross-reactions of enterotoxins from Escherichia coli and Vibrio cholerae. _Journal of Infectious Diseases_ 127:164–170. DOI: https://doi.org/10.1093/infdis/127.2.164, PMID: 4570128

Additionally, provide a summary of pairwise correlation across all antibody responses included in the multiplex.

# Preamble

```{r preamble, message=FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))


#-------------------------------
# drop "_public" suffix from IDs
#-------------------------------
d <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public)


#-------------------------------
# limit to ETEC and cholera
#-------------------------------
d1 <- d %>%
  filter(antigen %in% c("etec","ctb")) %>%
  mutate(antigenf = factor(antigenf))



#-------------------------------
# reshape the data wide
# to make it easier to plot
#-------------------------------
dmfi <- d1 %>%
  ungroup() %>%
  select(childid,phase,agem,antigen,logmfi) %>%
  pivot_wider(id_cols = c("childid","phase","agem"), names_from = antigen, values_from = logmfi)

```


# Correlation between ETEC LTB and cholera CTB

```{r ETEC v CTB figure, message=FALSE}

# estimate pearson correlation for printing
sp_corr <- cor(dmfi$etec,dmfi$ctb, method = "pearson")
cor_lab <- paste(sprintf("%1.2f",sp_corr))

# custom log labels
log10labs <- c( 
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)

plot_etec_ctb <- ggplot(data=dmfi, aes(x = ctb, y = etec)) +
  geom_point(alpha = 0.2, size = 0.3, color = vircols[5]) +
  geom_smooth(color = "black",lwd=0.75,se=FALSE, method = "loess") +
  annotate("text",x=1.1,y=4.25,label = paste("r ==",deparse(cor_lab)),parse=TRUE, hjust =0) +
  scale_x_continuous(breaks=1:4,labels = log10labs) +
  scale_y_continuous(breaks=1:4,labels = log10labs) +
  labs(x = "Cholera CTB  IgG Luminex Response (MFI-bg)", y = "ETEC LTB IgG Luminex Response (MFI-bg)") +
  theme_minimal()

plot_etec_ctb

ggsave(filename = here("figures","mordor-ab-etec-v-ctb.png"),plot_etec_ctb, device = "png",width=4, height = 4)

```

# Correlation between all antigens

Estimate the correlation in log10 MFI values for all antigen pairs

```{r pairwise correlation plot, fig.width = 10,fig.height=10}

# pivot the MFI values wider to estimate correlation
d_corr <- d %>%
  filter(!antigenf %in% c("GST negative control")) %>%
  select(clusterid,childid,phase,antigenf,logmfi) %>%
  pivot_wider(id_cols=c("clusterid","childid","phase"),names_from="antigenf",values_from = logmfi) %>%
  select(clusterid,childid,phase,levels(d$antigenf)[1:20])

# re-arrange the columns so they are still in order
# d_corr[c("clusterid","childid","phase",levels(d$antigenf)[1:20])]


d_corr_ests <- cor(d_corr[,-c(1:3)])

library(ggcorrplot)
antigen_cor_plot <- ggcorrplot(d_corr_ests, 
                               # method = "circle",
                               type = "lower",
                               lab = TRUE,
                               lab_size = 3,
                               insig = "blank",
                               colors = c(cbpal[2], "white", cbpal[3])
                               )

antigen_cor_plot

ggsave(filename = here("figures","mordor-ab-ab-correlation-matrix.pdf"),antigen_cor_plot, device = "pdf",width=9, height = 9)

```

# Session Info
```{r session info}
sessionInfo()
```
