==> R/08-mordor-ab-community-means.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Community level summaries"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Summarize antibody responses at the community level

# Preamble

```{r preamble}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
# drop GST negative control values
# drop trachoma and strongy because
# no transmission in this population
#-------------------------------
d1 <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public) %>%
  filter(!antigen %in% c("gst","pgp3","ct694","nie")) %>%
  mutate(antigenf = factor(antigenf))
  
#-------------------------------
# drop cholera CTB because so 
# highly cross-reactive with ETEC LTB
#-------------------------------
d2 <- d1 %>%
  filter( !antigen %in% c("ctb") ) %>%
  mutate(antigenf = factor(antigenf))

#-------------------------------
# drop baseline measurements
#-------------------------------
d2 <- d2 %>%
  filter(phase > 0)

#-------------------------------
# age restrictions based on 
# age-antibody curves
# to exclude ages with evidence
# for maternal IgG contributions
#
# Bacteria and protozoa: 6 mo - 59 mo
#
# ETEC LTB: 6 mo - 24 mo (no heterogeneity at older ages)
#
# Malaria antigens: 12 - 59 mo
#-------------------------------
d2 <- d2 %>%
  filter(agem >= 6) %>%
  mutate(agedrop = case_when(
    antigen %in% c("etec") & (agem > 24) ~ 1,
    antigen %in% c("hrp2","csp","glurp","lsa1","ama","pfmsp1","pvmsp1","pomsp1","pmmsp1") & (agem < 12) ~ 1,
    TRUE ~ 0
  )
         ) %>%
  filter(agedrop == 0)

```

# Create composite seropositivity indicators

For enteric pathogens with multiple antigens (Cryptosporidium, Giardia, Salmonella, Campylobacter), create a series of seropositivity indicators that indicate if a child was positive to either antigen.

Then bind those to the original data frame to create a seropositivity dataset for analyses.

```{r composite seropositivity indicators}
d_enteric_comps <- d2 %>%
  filter(pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  group_by(pathogen,childid,phase) %>%
  mutate(seropos = max(seropos)) %>%
  slice(1) %>%
  mutate(antigenf = case_when(
    antigen %in% c("cp17","cp23") ~ "Cryptosporidium Cp17 or Cp23",
    antigen %in% c("vsp3","vsp5") ~ "Giardia VSP-3 or VSP-5",
    antigen %in% c("salb","sald") ~ "Salmonella groups B or D LPS",
    antigen %in% c("p18","p39") ~ "Campylobacter p18 or p39"
  )) %>%
  select(clusterid,phase,arm,childid,agem,sex,dbsid,pathogen,antigenf,seropos)


```

```{r create seropositivity dataset, warning=FALSE}
d_serop <- d2 %>%
  # drop pathogens w/ composite indicators
  filter(!pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  # add composite indicators
  bind_rows(d_enteric_comps) %>%
  # create new factor levels that includes composite indicators
  mutate(antigenf = factor(antigenf,levels = c(
    "Campylobacter p18 or p39",
    "ETEC LTB",
    "Salmonella groups B or D LPS",
    "Cryptosporidium Cp17 or Cp23",
    "Giardia VSP-3 or VSP-5",
   levels(d2$antigenf)[10:19]) )
         ) %>%
  # drop extra variables (MFI doesn't correspond to composite seropos so drop)
  select(-mfi,-logmfi,-agedrop)

```

# IgG levels (MFI)

## Estimate MFI by community
```{r mfi by community}

#-----------------------------
# count N and estimate 
# mean log10 MFI by community
# averaged over all post-tr
# phases
#-----------------------------
mu_mfi <- d2 %>%
  group_by(pathogen,antigenf,clusterid,arm) %>%
  mutate(nomiss = ifelse(!is.na(logmfi),1,0)) %>%
  summarize(nobs = sum(nomiss), 
            mulogmfi = mean(logmfi),
            sdlogmfi = sd(logmfi),
            selogmfi = sdlogmfi/sqrt(nobs),
            mulogmfi_lb = mulogmfi - 1.96*selogmfi,
            mulogmfi_ub = mulogmfi + 1.96*selogmfi,
            .groups = "keep") %>%
  # identify malaria antigens vs others
    mutate(group = ifelse(substr(pathogen,1,2)=="P.","malaria","enterics\nand\nstrep"))

```


## Heatmaps of MFI by community

```{r mfi heatmap enterics}

# sort clusters by mean seroprev across all antigens
cl_rank <- mu_mfi %>%
  filter(group != "malaria") %>%
  group_by(clusterid) %>%
  summarize(mean_mfi = mean(mulogmfi), .groups = "keep") %>%
  slice(1) %>%
  arrange(mean_mfi) %>%
  ungroup() %>%
  mutate(clrank = row_number(),
         clrankf = factor(clrank))

plot_mu_mfi <- mu_mfi %>%
  filter(group != "malaria") %>%
  left_join(cl_rank,by="clusterid") %>%
  mutate(antigenf = factor(antigenf,levels=rev(levels(antigenf))),
         arm = factor(arm, levels=c("placebo","azithro"), labels = c("placebo","azithromycin"))
         ) 


#----------------------------------
# make a heatmap of all antigens
# inspiration: http://www.roymfrancis.com/a-guide-to-elegant-tiled-heatmaps-in-r/
#----------------------------------
#define a colour for fonts
textcol <- "grey20"
p <- ggplot(plot_mu_mfi,aes(x=clrankf,y=antigenf,fill=mulogmfi)) +
  #facet over village
  facet_grid(.~arm,scales='free',space='free',switch="y")+
  
  #add border white colour of line thickness 0.25
  geom_tile(colour="white",size=0.25)+
  #remove y axis labels, 
  labs(x="Columns: communities stratified by arm then sorted by mean seroprevalence",y="",title="")+
  #remove extra space
  scale_y_discrete(expand=c(0,0),position="right")+
  scale_x_discrete(expand=c(0,0))+
  # scale_x_discrete(expand=c(0,0),
  #                  breaks=1:9,labels=1:9)+
  #change the scale_fill_manual
  scale_fill_distiller(palette="GnBu",na.value="grey90",
                       direction=0,
                       guide=guide_colorbar(title="log10 MFI-bg",face='bold'))+
  # scale_fill_manual(values=rev(brewer.pal(7,"YlGnBu")),na.value="grey90",guide=guide_colorbar(title="log10\nMFI-bg",face='bold'))+
  #one unit on x-axis is equal to one unit on y-axis.
  #equal aspect ratio x and y axis
  # coord_equal()+
  #set base size for all font elements
  theme_grey(base_size=10)+
  #theme options
  theme(
    
    legend.title=element_text(color=textcol,size=8),
    legend.position = "left",
    #change legend text properties
    legend.text=element_text(colour=textcol,size=7,face="plain"),
    #change legend key height
    legend.key.height=grid::unit(0.8,"cm"),
    #set a slim legend
    legend.key.width=grid::unit(0.4,"cm"),
    
    #set remove x axis ticks
    axis.title.x=element_text(hjust=0),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    #set y axis text colour and adjust vertical justification
    axis.text.y=element_text(size=10,vjust = 0.2,colour=textcol),
    #change axis ticks thickness
    axis.ticks=element_line(size=0.4),
    
    #change title font, size, colour and justification
    plot.title=element_text(colour=textcol,hjust=0,size=12,face="bold"),
    #adjust facet labels
    strip.text.x = element_text(size=10),
    strip.text.y = element_text(size=8),
    #remove plot background
    plot.background=element_blank(),
    #remove plot border
    panel.border=element_blank(),
    #remove plot margins
    # plot.margin=margin(grid::unit(1,"cm"))
  )
p

```


```{r mfi heatmap malaria}

# sort clusters by mean seroprev across all antigens
cl_rank <- mu_mfi %>%
  filter(group == "malaria") %>%
  group_by(clusterid) %>%
  summarize(mean_mfi = mean(mulogmfi), .groups = "keep") %>%
  slice(1) %>%
  arrange(mean_mfi) %>%
  ungroup() %>%
  mutate(clrank = row_number(),
         clrankf = factor(clrank))

plot_mu_mfi <- mu_mfi %>%
  filter(group == "malaria") %>%
  left_join(cl_rank,by="clusterid") %>%
  mutate(antigenf = factor(antigenf,levels=rev(levels(antigenf))),
         arm = factor(arm, levels=c("placebo","azithro"), labels = c("placebo","azithromycin"))
         ) 


#----------------------------------
# make a heatmap of all antigens
# inspiration: http://www.roymfrancis.com/a-guide-to-elegant-tiled-heatmaps-in-r/
#----------------------------------
#define a colour for fonts
textcol <- "grey20"
p <- ggplot(plot_mu_mfi,aes(x=clrankf,y=antigenf,fill=mulogmfi)) +
  #facet over village
  facet_grid(.~arm,scales='free',space='free',switch="y")+
  
  #add border white colour of line thickness 0.25
  geom_tile(colour="white",size=0.25)+
  #remove y axis labels, 
  labs(x="Columns: communities stratified by arm then sorted by mean seroprevalence",y="",title="")+
  #remove extra space
  scale_y_discrete(expand=c(0,0),position="right")+
  scale_x_discrete(expand=c(0,0))+
  # scale_x_discrete(expand=c(0,0),
  #                  breaks=1:9,labels=1:9)+
  #change the scale_fill_manual
  scale_fill_distiller(palette="GnBu",na.value="grey90",
                       direction=0,
                       guide=guide_colorbar(title="log10 MFI-bg",face='bold'))+
  # scale_fill_manual(values=rev(brewer.pal(7,"YlGnBu")),na.value="grey90",guide=guide_colorbar(title="log10\nMFI-bg",face='bold'))+
  #one unit on x-axis is equal to one unit on y-axis.
  #equal aspect ratio x and y axis
  # coord_equal()+
  #set base size for all font elements
  theme_grey(base_size=10)+
  #theme options
  theme(
    
    legend.title=element_text(color=textcol,size=8),
    legend.position = "left",
    #change legend text properties
    legend.text=element_text(colour=textcol,size=7,face="plain"),
    #change legend key height
    legend.key.height=grid::unit(0.8,"cm"),
    #set a slim legend
    legend.key.width=grid::unit(0.4,"cm"),
    
    #set remove x axis ticks
    axis.title.x=element_text(hjust=1,size=8),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    #set y axis text colour and adjust vertical justification
    axis.text.y=element_text(size=10,vjust = 0.2,colour=textcol),
    #change axis ticks thickness
    axis.ticks=element_line(size=0.4),
    
    #change title font, size, colour and justification
    plot.title=element_text(colour=textcol,hjust=0,size=12,face="bold"),
    #adjust facet labels
    strip.text.x = element_text(size=10),
    strip.text.y = element_text(size=8),
    #remove plot background
    plot.background=element_blank(),
    #remove plot border
    panel.border=element_blank(),
    #remove plot margins
    # plot.margin=margin(grid::unit(1,"cm"))
  )
p

```


# Seroprevalence

## Estimate seroprevalence by community
```{r seroprev by community}

#-----------------------------
# count N and estimate 
# seroprevalence by community
# averaged over all post-tr
# phases
#-----------------------------
mu_serop <- d_serop %>%
  group_by(pathogen,antigenf,clusterid,arm) %>%
  mutate(nomiss = ifelse(!is.na(seropos),1,0)) %>%
  summarize(nobs = sum(nomiss), 
            npos = sum(seropos),
            seroprev = mean(seropos),
            .groups = "keep") 
#-----------------------------
# estimate exact binomial
# 95% CIs for community means
#-----------------------------
# estimate exact binomial CIs for prevalence
mu_serop <- mu_serop %>% 
  rowwise() %>% 
  mutate(seroprev_ci = list(enframe(binom.test(x=npos, n=nobs, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(seroprev_ci) %>% 
    spread(name, value) %>% 
    rename("seroprev_lb" = "1", "seroprev_ub" = "2") %>%
  # identify malaria antigens vs others
    mutate(group = ifelse(substr(pathogen,1,2)=="P.","malaria","bacteria\nand\nprotozoa"))

```




## Heatmaps of seroprevalence by community

Create a separate heatmap for enterics and malaria since the seroprevalence levels are so different.

```{r seroprev heatmap enterics}

# sort clusters by mean seroprev across all antigens
cl_rank <- mu_serop %>%
  filter(group != "malaria") %>%
  group_by(clusterid) %>%
  summarize(mean_serop = mean(seroprev), .groups = "keep") %>%
  slice(1) %>%
  arrange(mean_serop) %>%
  ungroup() %>%
  mutate(clrank = row_number(),
         clrankf = factor(clrank))

plot_mu_serop <- mu_serop %>%
  filter(group != "malaria") %>%
  left_join(cl_rank,by="clusterid") %>%
  mutate(antigenf = factor(antigenf,levels=rev(levels(antigenf))),
         arm = factor(arm, levels=c("placebo","azithro"), labels = c("placebo","azithromycin"))
         ) 


#----------------------------------
# make a heatmap of all antigens
# inspiration: http://www.roymfrancis.com/a-guide-to-elegant-tiled-heatmaps-in-r/
#----------------------------------
#define a colour for fonts
textcol <- "grey20"
p <- ggplot(plot_mu_serop,aes(x=clrankf,y=antigenf,fill=seroprev*100)) +
  #facet over village
  facet_grid(.~arm,scales='free',space='free',switch="y")+
  
  #add border white colour of line thickness 0.25
  geom_tile(colour="white",size=0.25)+
  #remove y axis labels, 
  labs(x="Columns: communities stratified by arm then sorted by mean seroprevalence",y="",title="", hjust=1)+
  #remove extra space
  scale_y_discrete(expand=c(0,0),position="right")+
  scale_x_discrete(expand=c(0,0))+
  # scale_x_discrete(expand=c(0,0),
  #                  breaks=1:9,labels=1:9)+
  #change the scale_fill_manual
  scale_fill_distiller(palette="GnBu",na.value="grey90",
                       direction=0,
                       guide=guide_colorbar(title="sero-\nprev (%)",face='bold'))+
  # scale_fill_manual(values=rev(brewer.pal(7,"YlGnBu")),na.value="grey90",guide=guide_colorbar(title="log10\nMFI-bg",face='bold'))+
  #one unit on x-axis is equal to one unit on y-axis.
  #equal aspect ratio x and y axis
  # coord_equal()+
  #set base size for all font elements
  theme_grey(base_size=10)+
  #theme options
  theme(
    
    legend.title=element_text(color=textcol,size=8),
    legend.position = "left",
    #change legend text properties
    legend.text=element_text(colour=textcol,size=7,face="plain"),
    #change legend key height
    legend.key.height=grid::unit(0.4,"cm"),
    #set a slim legend
    legend.key.width=grid::unit(0.4,"cm"),
    
    #set remove x axis ticks
    axis.title.x=element_text(hjust=0, size=8),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    #set y axis text colour and adjust vertical justification
    axis.text.y=element_text(size=10,vjust = 0.2,colour=textcol),
    #change axis ticks thickness
    axis.ticks=element_line(size=0.4),
    
    #change title font, size, colour and justification
    plot.title=element_text(colour=textcol,hjust=0,size=12,face="bold"),
    #adjust facet labels
    strip.text.x = element_text(size=10),
    strip.text.y = element_text(size=8),
    #remove plot background
    plot.background=element_blank(),
    #remove plot border
    panel.border=element_blank(),
    #remove plot margins
    # plot.margin=margin(grid::unit(1,"cm"))
  )
p

ggsave(filename = here("figures","mordor-ab-community-seroprev-heatmap-bacteria-protozoa.pdf"),p, device = "pdf", width=7,height=1.9)


```

```{r seroprev heatmap malaria}

# sort clusters by mean seroprev across all antigens
cl_rank <- mu_serop %>%
  filter(group == "malaria") %>%
  group_by(clusterid) %>%
  summarize(mean_serop = mean(seroprev), .groups = "keep") %>%
  slice(1) %>%
  arrange(mean_serop) %>%
  ungroup() %>%
  mutate(clrank = row_number(),
         clrankf = factor(clrank))

plot_mu_serop <- mu_serop %>%
  filter(group == "malaria") %>%
  left_join(cl_rank,by="clusterid") %>%
  mutate(antigenf = factor(antigenf,levels=rev(levels(antigenf))),
         arm = factor(arm, levels=c("placebo","azithro"), labels = c("placebo","azithromycin"))
         ) 


#----------------------------------
# make a heatmap of all antigens
# inspiration: http://www.roymfrancis.com/a-guide-to-elegant-tiled-heatmaps-in-r/
#----------------------------------
#define a colour for fonts
textcol <- "grey20"
p <- ggplot(plot_mu_serop,aes(x=clrankf,y=antigenf,fill=seroprev*100)) +
  #facet over village
  facet_grid(.~arm,scales='free',space='free',switch="y")+
  
  #add border white colour of line thickness 0.25
  geom_tile(colour="white",size=0.25)+
  #remove y axis labels, 
  labs(x="Columns: communities stratified by arm then sorted by mean seroprevalence",y="",title="")+
  #remove extra space
  scale_y_discrete(expand=c(0,0),position="right")+
  scale_x_discrete(expand=c(0,0))+
  # scale_x_discrete(expand=c(0,0),
  #                  breaks=1:9,labels=1:9)+
  #change the scale_fill_manual
  scale_fill_distiller(palette="GnBu",na.value="grey90",
                       direction=0,
                       guide=guide_colorbar(title="sero-\nprev (%)",face='bold'))+
  # scale_fill_manual(values=rev(brewer.pal(7,"YlGnBu")),na.value="grey90",guide=guide_colorbar(title="log10\nMFI-bg",face='bold'))+
  #set base size for all font elements
  theme_grey(base_size=10)+
  #theme options
  theme(
    
    legend.title=element_text(color=textcol,size=8),
    legend.position = "left",
    #change legend text properties
    legend.text=element_text(colour=textcol,size=7,face="plain"),
    #change legend key height
    legend.key.height=grid::unit(0.8,"cm"),
    #set a slim legend
    legend.key.width=grid::unit(0.4,"cm"),
    
    #set remove x axis ticks
    axis.title.x=element_text(hjust=0,size=8),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    #set y axis text colour and adjust vertical justification
    axis.text.y=element_text(size=10,vjust = 0.2,colour=textcol),
    #change axis ticks thickness
    axis.ticks=element_line(size=0.4),
    
    #change title font, size, colour and justification
    plot.title=element_text(colour=textcol,hjust=0,size=12,face="bold"),
    #adjust facet labels
    strip.text.x = element_text(size=10),
    strip.text.y = element_text(size=8),
    #remove plot background
    plot.background=element_blank(),
    #remove plot border
    panel.border=element_blank(),
    #remove plot margins
    # plot.margin=margin(grid::unit(1,"cm"))
  )
p

ggsave(filename = here("figures","mordor-ab-community-seroprev-heatmap-malaria.pdf"),p, device = "pdf", width=6.5,height=2.5)


```




## Cluster level seroprevalence

Bacteria and protozoa

```{r plot seroprev by community enterics, fig.width = 6, fig.height = 4}
set.seed(189724)
pcols <- cbpal[c(2,6)]
p_comm <- ggplot(data = mu_serop %>% filter(group != "malaria" ), aes(x = arm, y = seroprev, color = arm)) +
  facet_wrap(~antigenf, ncol = 3) +
  geom_jitter(width = 0.1, alpha = 0.5) +
  geom_boxplot(color = "gray40",fill = NA, outlier.shape = NA, width = 0.5) +
  scale_y_continuous(breaks=seq(0,1,by=0.2),labels=sprintf("%1.0f",100*seq(0,1,by=0.2)))+
  scale_color_manual(values = pcols) +
  coord_cartesian(ylim=c(0,1)) +
  labs(x = NULL, y = "community-level seroprevalence (%)") +
  theme(
    legend.position = "none"
  )

p_comm

ggsave(filename = here("figures","mordor-ab-community-seroprev-bacteria-protozoa.pdf"),p_comm, device = "pdf", width=6,height=6)

```

Malaria

```{r plot seroprev by community malaria, fig.width = 5, fig.height = 4}
set.seed(3243)
pcols <- cbpal[c(2,6)]
p_comm <- ggplot(data = mu_serop %>% filter(group == "malaria"), aes(x = arm, y = seroprev, color = arm)) +
  facet_wrap(~antigenf, ncol = 3) +
  geom_jitter(width = 0.1, alpha = 0.5) +
  geom_boxplot(color = "gray40",fill = NA, outlier.shape = NA, width = 0.5) +
  scale_y_continuous(breaks=seq(0,1,by=0.2),labels=sprintf("%1.0f",100*seq(0,1,by=0.2)))+
  scale_color_manual(values = pcols) +
  coord_cartesian(ylim=c(0,1)) +
  labs(x = NULL, y = "community-level seroprevalence (%)") +
  theme(
    legend.position = "none"
  )

p_comm


ggsave(filename = here("figures","mordor-ab-community-seroprev-malaria.pdf"),p_comm, device = "pdf", width=5.5,height=6)

```
# Community mean and ICC summary

## Estimate community level ICCs

Estimate the intra-cluster correlation (ICC) for seroprevalence measures.

```{r estimate iccs, warning=FALSE, message=FALSE}

d_serop <- d_serop %>% 
  mutate(clusteridf = factor(clusterid))

set.seed(1231)
icc_ests <- foreach(abi = levels(d_serop$antigenf), .combine = rbind) %do% {
  di <- d_serop %>% filter(antigenf==abi)
  fiti <- rptBinary(seropos ~ 1 + (1|clusteridf), grname = "clusteridf", data=di, nboot = 1000, parallel = TRUE)
  icci <- fiti$R[1,1]
  icci_lb <- fiti$CI_emp[[1]][1]
  icci_ub <- fiti$CI_emp[[1]][2]
  res <- data.frame(antigenf=abi, icc=icci, icc_lb = icci_lb, icc_ub = icci_ub)
  return(res)
}
  



```

## Estimate means and SDs

Estimate the between-community SD in seroprevalence. Summarize mean seroprevalence and the SD of the community-level means. 

```{r community mean sd}
c_serop <- mu_serop %>%
  ungroup() %>%
  group_by(antigenf) %>%
  summarize(nobs=sum(nobs),
            npos=sum(npos),
            musd = sd(seroprev)) %>%
  mutate(mu = npos/nobs)

```


## Community mean, SD, ICC Table

Create a summary table of mean (SD) for community-level seroprevalence, along with ICC and the 95% CI for the ICC.

```{r icc table}

c_serop2 <- c_serop %>%
  left_join(icc_ests, by="antigenf") %>%
  rename(icc_lb = "X2.5.", icc_ub = "X97.5.") %>%
  mutate(muprint = paste0(sprintf("%1.2f",mu)," (",sprintf("%1.3f",musd),")"),
         iccprint = paste0(sprintf("%1.2f",icc)," (",sprintf("%1.2f",icc_lb),", ",sprintf("%1.2f",icc_ub),")")
  )

# stack to put malaria antigens on top
c_serop3 <- bind_rows(c_serop2[7:15,],c_serop2[1:6,]) %>%
  mutate(m=30) %>%
  select(antigenf,m,nobs,muprint,iccprint)


knitr::kable(c_serop3, 
             row.names = FALSE,
             col.names = c("Antibody","m","n","Seroprevalence (SD)", "ICC (95% CI)*"),
             digits = c(0,0,0,0),
             format.args = list(big.mark = ','),
             align = c("l","c","c","c","c"),
             caption = "Community mean and standard deviation (SD) seroprevalence for n children measured in m communities, along with community-level ICC") %>%
  kable_styling(bootstrap_options = "striped") %>%
  pack_rows(group_label = "Malaria †",start_row = 1, end_row = 9) %>%
  pack_rows(group_label = "Bacteria and protozoa ‡", start_row = 10, end_row = 15) %>%
  footnote(symbol = c("Intra-cluster correlation coefficient (ICC). Its 95% confidence interval (CI) was estimated with a parametric bootstrap and 1000 iterations.",
                      "Malaria antibody seroprevalence estimated among children ages 12-59 months, per the age group used in the primary analysis.",
                      "Bacteria and protozoan antibody seroprevalence estimated among children ages 6-24 months (ETEC LTB) or among children ages 6-59 months (all others), per the age groups used in the primary analysis"))

```


# Session Info
```{r session info}
sessionInfo()
```
