==> R/09-mordor-ab-age-curves.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Compare age-dependent means and seroprevalence by arm"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Compare age-dependent mean antibody levels and seroprevalence between the placebo and biannual azithromycin arms. 

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
#-------------------------------
d <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public)

# drop GST negative control values
# drop trachoma and strongy because
# no transmission in this population
d2 <- d %>%
  filter(!antigen %in% c("gst","pgp3","ct694","nie")) 

# drop cholera CTB because so highly cross-reactive with ETEC LTB
d2 <- d2 %>%
  filter( !antigen %in% c("ctb") ) %>%
  mutate(antigenf = factor(antigenf))
```

# Create composite seropositivity indicators

For enteric pathogens with multiple antigens (Cryptosporidium, Giardia, Salmonella, Campylobacter), create a series of seropositivity indicators that indicate if a child was positive to either antigen.

```{r composite seropositivity indicators}
d_enteric_comps <- d2 %>%
  filter(pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  group_by(pathogen,childid,phase) %>%
  mutate(seropos = max(seropos)) %>%
  slice(1) %>%
  mutate(antigenf = case_when(
    antigen %in% c("cp17","cp23") ~ "Cryptosporidium Cp17 or Cp23",
    antigen %in% c("vsp3","vsp5") ~ "Giardia VSP-3 or VSP-5",
    antigen %in% c("salb","sald") ~ "Salmonella groups B or D LPS",
    antigen %in% c("p18","p39") ~ "Campylobacter p18 or p39"
  ),
  # create new factor levels that includes composite indicators
  antigenf = factor(antigenf,levels = c(
    "Campylobacter p18 or p39",
    "ETEC LTB",
    "Salmonella groups B or D LPS",
    "Cryptosporidium Cp17 or Cp23",
    "Giardia VSP-3 or VSP-5",
   levels(d2$antigenf)[10:19]) )
         ) %>%
  select(clusterid,phase,arm,childid,agem,sex,dbsid,pathogen,antigenf,seropos)


```

# Estimate age-dependent means by arm

```{r age-dependent means by arm}
#-----------------------------
# create an indicator for 
# children living in intervention
# and control communities
#
# create a dummy indicator equal
# to 1 for all communities to zero
# out the marginal predictions
# over all community random effects
#
# create a factor variable
# for clusterid (for model)
#-----------------------------
d3 <- d2 %>%
  mutate(tr0 = ifelse(arm == "placebo",1,0),
         tr1 = ifelse(arm == "azithro",1,0),
         dummy = 1,
         clusteridf = factor(clusterid)
         )

#----------------------------------------
# fit a GAM model fitting a separate
# spline by age for each group
# estimate simulatenous 95% confidence 
# intervals
#----------------------------------------

fit_age <- foreach(abi = levels(d3$antigenf), .combine = rbind) %do% {
  cat("\n Estimating splines for",abi)
  di <- d3 %>% filter(antigenf == abi)
  fit_age <- mgcv::gam(logmfi ~ s(agem,bs="cr") +
                         arm +
                         s(clusteridf,bs="re",by=dummy),
                       data=di)
  newd0 <- di %>%  mutate(dummy=0, arm = "placebo")
  newd1 <- di %>%  mutate(dummy=0, arm = "azithro")
  fit_tr0_ci <- gamCI(m=fit_age,newdata=newd0,nreps=1000)
  fit_tr1_ci <- gamCI(m=fit_age,newdata=newd1,nreps=1000)
  fit_ci <- bind_rows(fit_tr0_ci,fit_tr1_ci)
  
}

fit_age2 <- fit_age %>%
  mutate(arm = factor(arm,levels = c("placebo","azithro")))


```

# Estimate age-dependent seroprevalence by arm

```{r age-dependent seroprevalence by arm}
#----------------------------------------
# fit a GAM model fitting a separate
# spline by age for each group
# estimate simulatenous 95% confidence 
# intervals
#----------------------------------------

fitp_age <- foreach(abi = levels(d3$antigenf), .combine = rbind) %do% {
  cat("\n Estimating splines for",abi)
  di <- d3 %>% filter(antigenf == abi)
  fit_age <- mgcv::gam(seropos ~ s(agem,bs="cr") + 
                         arm +
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link="cloglog"),
                       data=di)
  newd0 <- di %>%  mutate(dummy=0, arm = "placebo")
  newd1 <- di %>%  mutate(dummy=0, arm = "azithro")
  fit_tr0_ci <- gamCI(m=fit_age,newdata=newd0,nreps=1000)
  fit_tr1_ci <- gamCI(m=fit_age,newdata=newd1,nreps=1000)
  fit_ci <- bind_rows(fit_tr0_ci,fit_tr1_ci)
  
}

# convert linear predictor to prevalance
fitp_age2 <- fitp_age %>%
  mutate(arm = factor(arm,levels = c("placebo","azithro"))) %>%
  mutate(fit = cloglogfn(fit),
         uprP = cloglogfn(uprP),
         lwrP = cloglogfn(lwrP),
         uprS = cloglogfn(uprS),
         lwrS = cloglogfn(lwrS),
         ) 

```

Repeat for composite seropositivity measures for enterics

```{r age-dependent seroprevalence by arm composites}
#----------------------------------------
# fit a GAM model fitting a separate
# spline by age for each group
# estimate simulatenous 95% confidence 
# intervals
#----------------------------------------
d_enteric_comps2 <- d_enteric_comps %>%
  ungroup() %>%
  mutate(tr0 = ifelse(arm == "placebo",1,0),
         tr1 = ifelse(arm == "azithro",1,0),
         dummy = 1,
         clusteridf = factor(clusterid),
         antigenf = droplevels(antigenf)
         )

fitp_comp_age <- foreach(abi = levels(d_enteric_comps2$antigenf), .combine = rbind) %do% {
  cat("\n Estimating splines for",abi)
  di <- d_enteric_comps2 %>% filter(antigenf == abi)
  fit_age <- mgcv::gam(seropos ~ s(agem,bs="cr") + 
                         arm +
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link = "cloglog"),
                       data=di)
  newd0 <- di %>%  mutate(dummy=0, arm = "placebo")
  newd1 <- di %>%  mutate(dummy=0, arm = "azithro")
  fit_tr0_ci <- gamCI(m=fit_age,newdata=newd0,nreps=1000)
  fit_tr1_ci <- gamCI(m=fit_age,newdata=newd1,nreps=1000)
  fit_ci <- bind_rows(fit_tr0_ci,fit_tr1_ci)
  
}

# convert linear predictor to prevalance
fitp_comp_age2 <- fitp_comp_age %>%
  mutate(arm = factor(arm,levels = c("placebo","azithro"))) %>%
  mutate(fit = cloglogfn(fit),
         uprP = cloglogfn(uprP),
         lwrP = cloglogfn(lwrP),
         uprS = cloglogfn(uprS),
         lwrS = cloglogfn(lwrS),
         ) 



```

# Estimate age-dependent force of infection

Repeat the age-seroprevalence fits from above, but use the derivative of the linear predictor to estimate age-dependent force of infection:

\begin{equation}
\lambda(a) = \eta'(a)\exp(\eta(a))
\end{equation}

where $\eta'(a)$ is the first derivative of the linear predictor $\eta(a)$ from a binomial generalized additive model with a complementary log-log link.

Estimate $\lambda(a)$ only over ages included in the primary analysis, which is 6-24 months for enterics (except _Salmonella_), and 12-59 months for malaria. 

```{r age-specific FOI, warning=FALSE}
#-------------------------------
# age restrictions based on 
# age-antibody curves
# to exclude ages without any
# variation in antibody levels
# (esp. enterics, which become saturated)
#
# Enterics: 6 mo - 24
# (except for salmonella, 6-59)
#
# Strep A: 6-59
#
# Malaria antigens: 12 - 59 mo
#-------------------------------
d_foi <- d3 %>%
  filter(agem >= 6) %>%
  mutate(agedrop = case_when(
    antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","etec","ctb") & (agem > 24) ~ 1,
    antigen %in% c("hrp2","csp","glurp","lsa1","ama","pfmsp1","pvmsp1","pomsp1","pmmsp1") & (agem < 12) ~ 1,
    TRUE ~ 0
  )) %>%
  filter(agedrop == 0)

set.seed(1234)
fit_foi_age <- foreach(abi = levels(d_foi$antigenf), .combine = rbind) %do% {
  cat("\n Estimating FOI for",abi)
  di <- d_foi %>% filter(antigenf == abi)
  fit_age <- mgcv::gam(seropos ~ s(agem,bs="cr") + arm +
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link="cloglog"),
                       data=di)
  
  # estimate average seroconversion rate by arm
  newd0 <- data.frame(antigenf = abi, agem=seq(min(di$agem),max(di$agem),by=0.01), dummy=0,clusteridf="C1",arm = "placebo")
  newd1 <- data.frame(antigenf = abi, agem=seq(min(di$agem),max(di$agem),by=0.01), dummy=0,clusteridf="C1",arm = "azithro")
  foi_tr0_ci <- foiCI(m=fit_age,newdata=newd0,fdvar="agem",nreps=1000)
  foi_tr1_ci <- foiCI(m=fit_age,newdata=newd1,fdvar="agem",nreps=1000)
  foi_ci <- bind_rows(foi_tr0_ci,foi_tr1_ci)
  
}
```

```{r age-specific FOI enterics composites, warning=FALSE}
#-------------------------------
# repeat FOI estimation based
# on the composite seropositivity
# indicators for some enterics
# 
# fit age-seroprev models on 
# all ages, but only estimate
# FOI in specific age regions
# determined in the primary analysis
#-------------------------------
d_foi_comp <- d_enteric_comps2 %>%
  filter(agem >= 6) %>%
  mutate( estfoi = ifelse(agem >= 6,1,0),
          estfoi = ifelse(antigenf != "Salmonella LPS B or D" & (agem > 24), 0,estfoi)
          ) 

set.seed(1234)
fit_foi_comp_age <- foreach(abi = levels(d_foi_comp$antigenf), .combine = rbind) %do% {
  cat("\n Estimating FOI for",abi)
  di <- d_foi_comp %>% filter(antigenf == abi)
  fit_age <- mgcv::gam(seropos ~ s(agem,bs="cr",by=estfoi) + arm +
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link="cloglog"),
                       data=di)
  
  # estimate average seroconversion rate by arm
  newd0 <- data.frame(antigenf = abi, agem=seq(min(di$agem[di$estfoi==1]),max(di$agem[di$estfoi==1]),by=0.01), dummy=0,estfoi=1,clusteridf="C1",arm = "placebo")
  newd1 <- data.frame(antigenf = abi, agem=seq(min(di$agem[di$estfoi==1]),max(di$agem[di$estfoi==1]),by=0.01), dummy=0,estfoi=1,clusteridf="C1",arm = "azithro")
  foi_tr0_ci <- foiCI(m=fit_age,newdata=newd0,fdvar="agem",nreps=1000)
  foi_tr1_ci <- foiCI(m=fit_age,newdata=newd1,fdvar="agem",nreps=1000)
  foi_ci <- bind_rows(foi_tr0_ci,foi_tr1_ci)
  
}


```


# Enterics and bacteria figure

Age-dependent MFI figure
```{r age-dependent means by arm figure enterics, fig.width = 6, fig.height = 12}
#-----------------------------
# plot age-dependent means
# by arm
#-----------------------------

# create a variable for faceting panels by pathogen
fit_age_enterics <- fit_age2 %>%
  filter(antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","salb","sald","etec","ctb","speb")) %>%
  mutate(antigenn=ifelse(antigenf %in% c("Giardia VSP-5","Cryptosporidium Cp23","Campylobacter p39","Salmonella group D LPS"),2,1)
         )

#-----------------------------
# labels for the panels in the plot 
#
# create a bounding box
# to shade ages used in
# the primary analysis
# and in age-dependent FOI
# estimation
#-----------------------------
pplabs <- fit_age_enterics %>%
  group_by(antigenf) %>%
  select(pathogen,antigen, antigenf, arm) %>%
  slice(1) %>%
  mutate(antigenn=ifelse(antigenf %in% c("Giardia VSP-5","Cryptosporidium Cp23","Campylobacter p39","Salmonella group D LPS"),2,1)
         ) %>%
  mutate(minX = 6, 
         maxX = ifelse(antigenf %in% c("Salmonella group B LPS", "Salmonella group D LPS","Streptococcus group A SPEB"),60,24),
         arm = "placebo"
         )

# custom log labels
log10labs <- c( 
  expression(10^0),
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)
pcols <- cbpal[c(2,6)]

plot_age_mfi_enterics <- ggplot(data = fit_age_enterics, aes(x = agem, y = fit, group = arm, color = arm, fill = arm)) +
  # facet by pathogen and antigen
  facet_grid(pathogen~antigenn, drop = TRUE) +
   # add shade box
  geom_rect(data=pplabs, inherit.aes = FALSE, aes(xmin = minX, xmax = maxX, ymin = 0, ymax = 4.5), alpha=0.1) +
  # 95% CI bands for spline fit
  geom_ribbon(aes(ymin = lwrS, ymax = uprS), color = NA, alpha = 0.4) +
  # spline fit
  geom_line() +
  # labels
  geom_text(data=pplabs, aes(x=0, y=4.6, label=antigenf, hjust=0, vjust=0), color="gray20", size=2.5)+
  # plot aesthetics
  scale_y_continuous(breaks = 0:4, labels = log10labs) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,4.65)) +
  labs(x = "Age, months", y = "IgG Luminex response (MFI-bg)", tag = "a") +
  theme_minimal() +
  theme(
    strip.text.x = element_blank(),
    strip.text.y = element_blank(),
    legend.position = "top",
    plot.tag = element_text(face = "bold")
  )


```

Seroprevalence figure. Use composite estimates for pathogens with multiple antigens

```{r age-dependent seroprevalence by arm figure, warning = FALSE, fig.width = 12, fig.height = 12}

#-----------------------------
# plot age-dependent seroprev
# by arm for enterics
#-----------------------------
fitp_enterics <- fitp_age2 %>%
  filter(pathogen %in% c("ETEC","Streptococcus")) %>%
  bind_rows(fitp_comp_age2) %>%
  mutate(antigenf = factor(antigenf,levels = c(levels(fitp_comp_age2$antigenf)[1],"ETEC LTB",levels(fitp_comp_age2$antigenf)[2:4],"Streptococcus group A SPEB")))

#-----------------------------
# create a bounding box
# to shade ages used in
# the primary analysis
# and in age-dependent FOI
# estimation
#-----------------------------
d_shade_box <- data.frame(antigenf = levels(fitp_enterics$antigenf)) %>%
  mutate(minX = 6, 
         maxX = ifelse(antigenf %in% c("Salmonella groups B or D LPS","Streptococcus group A SPEB"),60,24),
         arm = "placebo",
         antigenf = factor(antigenf,levels(fitp_enterics$antigenf))
         )

# labels for the panels in the plot 
pplabs2 <- fitp_enterics %>%
  group_by(antigenf) %>%
  select(pathogen, antigenf, arm) %>%
  slice(1)

pcols <- cbpal[c(2,6)]

plot_age_seroprev_enterics <- ggplot(data = fitp_enterics , aes(x = agem, y = fit, group = arm, color = arm, fill = arm)) +
  # facet by pathogen
  facet_grid(antigenf~.) +
  # add shade box
  geom_rect(data=d_shade_box, inherit.aes = FALSE, aes(xmin = minX, xmax = maxX, ymin = 0, ymax = 1), alpha=0.1) +
  # approximate pointwise 95% CIs
  geom_ribbon(aes(ymin = lwrS, ymax = uprS), color = NA, alpha = 0.3) +
  # spline fits
  geom_line() +
  # labels
  geom_text(data=pplabs2, aes(x=0, y=1.03, label=antigenf, hjust=0, vjust=0), color="gray20", size=2.5)+
  # plot aesthetics
  scale_y_continuous(breaks = seq(0,1,by=0.2), labels = sprintf("%1.0f",seq(0,1,by=0.2)*100)) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,1.04)) +
  labs(x = "Age, months", y = "IgG seroprevalence (%)", tag = "b") +
  theme_minimal() +
  theme(
    strip.text.x = element_blank(),
    strip.text.y = element_blank(),
    legend.position = "top",
    plot.tag = element_text(face = "bold")
  )


```

Composite of MFI and seroprevalence

```{r enterics age plot composite, warning=FALSE, fig.width = 12, fig.height = 18}
pcurve_composite_enterics <- grid.arrange(plot_age_mfi_enterics,
                                          plot_age_seroprev_enterics,
                                          ncol=2,nrow=1,widths=c(2,1.1))

ggsave(here("figures","mordor-ab-age-curves-bacteria-protozoa.pdf"),pcurve_composite_enterics, device = "pdf", width = 8, height = 11)

```


Seroconversion rate figure. Use composite estimates for pathogens with multiple antigens.

```{r age-dependent FOI by arm figure, warning = FALSE, fig.width = 5, fig.height = 15}

#-----------------------------
# plot age-dependent FOI
# by arm for enterics
# multiply seroconversion rate
# per month of age by 12
# to convert units to per year
#-----------------------------
foip_enterics <- fit_foi_age %>%
  filter(antigenf %in% c("ETEC LTB","Streptococcus group A SPEB")) %>%
  bind_rows(fit_foi_comp_age) %>%
  mutate(antigenf = factor(antigenf,levels = c(levels(d_foi_comp$antigenf)[1],"ETEC LTB", levels(d_foi_comp$antigenf)[2:4],"Streptococcus group A SPEB")),
         arm = factor(arm, levels = c("placebo","azithro"))
         )

pcols <- cbpal[c(2,6)]
plot_age_foi_enterics <- ggplot(data = foip_enterics , aes(x = agem, y = 12*foi, group = arm, color = arm, fill = arm)) +
  # facet by pathogen
  facet_grid(antigenf~., scales = "free_y") +
  # approximate pointwise 95% CIs
  geom_ribbon(aes(ymin = 12*foi_lwrS, ymax = 12*foi_uprS), color = NA, alpha = 0.3) +
  # spline fits
  geom_line() +
  # labels
  # geom_text(data=pplabs2, aes(x=0, y=8.2, label=antigenf, hjust=0, vjust=0), color="gray20", size=2.5)+
  # plot aesthetics
  # scale_y_continuous(breaks = seq(0,8,by=2)) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  # coord_cartesian(ylim = c(0,8.4)) +
  labs(x = "Age, months", y = expression(paste("Seroconversion rate per year (force of infection), ",lambda,"(a)"))) +
  theme_minimal() +
  theme(
    strip.text.x = element_blank(),
    # strip.text.y = element_blank(),
    legend.position = "top"
  )

plot_age_foi_enterics
```


## Campylobacter seroprevalence figure

Create a separate Campylobacter seroprevalence figure for an article panel

```{r campy seroprev figure}
dcamp <- d_enteric_comps2 %>% filter(antigenf == "Campylobacter p18 or p39")
fit_camp <- mgcv::gam(seropos ~ s(agem,bs="cr", by = tr0) + 
                         s(agem,bs="cr", by = tr1) + 
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link = "cloglog"),
                       data=dcamp)
  newd0 <- dcamp %>%  mutate(dummy=0, tr0 = 1, tr1 = 0)
  newd1 <- dcamp %>%  mutate(dummy=0, tr1 = 1, tr0 = 0)
  fit_tr0_ci <- gamCI(m=fit_camp,newdata=newd0,nreps=1000)
  fit_tr1_ci <- gamCI(m=fit_camp,newdata=newd1,nreps=1000)
  fit_ci <- bind_rows(fit_tr0_ci,fit_tr1_ci)
  

# convert linear predictor to prevalance
fit_camp2<- fit_ci %>%
  mutate(arm = ifelse(tr1==1,"azithro","placebo"),
         arm = factor(arm,levels = c("placebo","azithro"))) %>%
  mutate(fit = cloglogfn(fit),
         uprP = cloglogfn(uprP),
         lwrP = cloglogfn(lwrP),
         uprS = cloglogfn(uprS),
         lwrS = cloglogfn(lwrS),
         ) 


#-----------------------------
# create a bounding box
# to shade ages used in
# the primary analysis
# and in age-dependent FOI
# estimation
#-----------------------------
d_shade_box <- data.frame(minX = 6, maxX = 24,arm = "placebo")


# labels for the panels in the plot 
pplabs2 <- fitp_enterics %>%
  group_by(antigenf) %>%
  select(pathogen, antigenf, arm) %>%
  slice(1)

pcols <- cbpal[c(2,6)]

plot_age_seroprev_campy <- ggplot(data = fit_camp2 , aes(x = agem, y = fit, group = arm, color = arm, fill = arm)) +
  # add shade box
  geom_rect(data=d_shade_box, inherit.aes = FALSE, aes(xmin = minX, xmax = maxX, ymin = 0, ymax = 1), alpha=0.1) +
  # approximate pointwise 95% CIs
  # geom_ribbon(aes(ymin = lwrS, ymax = uprS), color = NA, alpha = 0.3) +
  # spline fits
  geom_line() +
  # labels
  annotate("text",x=54,y=0.97,label="placebo",color=pcols[1],hjust=1) +
  annotate("text",x=48,y=0.85,label="azithromycin",color=pcols[2],hjust=1) +
  scale_y_continuous(breaks = seq(0,1,by=0.2), labels = sprintf("%1.0f",seq(0,1,by=0.2)*100)) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,1.04)) +
  labs(x = "Age, months", y = expression(paste(italic("Campylobacter")," IgG seroprevalence (%)"))) +
  theme_minimal(base_size = 14) +
  theme(
    strip.text.x = element_blank(),
    strip.text.y = element_blank(),
    legend.position = "none"
  )


plot_age_seroprev_campy

ggsave(here("figures","mordor-ab-age-curves-campy.png"),plot_age_seroprev_campy, device = "png", width = 4, height = 4)
ggsave(here("figures","mordor-ab-age-curves-campy.pdf"),plot_age_seroprev_campy, device = "pdf", width = 4, height = 4)

```

# Malaria figure

Age-dependent MFI figure
```{r age-dependent means by arm figure malaria, fig.width = 6, fig.height = 12}
#-----------------------------
# plot age-dependent means
# by arm
#-----------------------------
fit_age_malaria <- fit_age2 %>%
  filter(!antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","salb","sald","etec","ctb","speb")) %>%
  mutate(antigenf = droplevels(antigenf))

#-----------------------------
# create a bounding box
# to shade ages used in
# the primary analysis
# and in age-dependent FOI
# estimation
#-----------------------------
d_shade_box <- data.frame(antigenf = levels(fit_age_malaria$antigenf)) %>%
  mutate(minX = 12, 
         maxX = 60,
         arm = "placebo",
         antigenf = factor(antigenf,levels(fit_age_malaria$antigenf))
         )

# custom log labels
log10labs <- c( 
  expression(10^0),
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)
pcols <- cbpal[c(2,6)]

plot_age_mfi_malaria <- ggplot(data = fit_age_malaria, aes(x = agem, y = fit, group = arm, color = arm, fill = arm)) +
  # facet by pathogen and antigen
  facet_grid(antigenf~., drop = TRUE) +
  # add shade box
  geom_rect(data=d_shade_box, inherit.aes = FALSE, aes(xmin = minX, xmax = maxX, ymin = 0, ymax = 4.5), alpha=0.1) +
  # 95% CI bands for spline fit
  geom_ribbon(aes(ymin = lwrS, ymax = uprS), color = NA, alpha = 0.4) +
  # spline fit
  geom_line() +
  # plot aesthetics
  scale_y_continuous(breaks = 0:4, labels = log10labs) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,4.65)) +
  labs(x = "Age, months", y = "IgG Luminex response (MFI-bg)", tag = "a") +
  theme_minimal() +
  theme(
    strip.text.x = element_blank(),
    strip.text.y = element_text(angle = 0),
    legend.position = "top",
    plot.tag = element_text(face = "bold")
  )

```

Seroprevalence figure. 

```{r age-dependent seroprevalence by arm figure malaria, warning = FALSE, fig.width = 12, fig.height = 12}

#-----------------------------
# plot age-dependent seroprev
# by arm for malaria
#-----------------------------
fitp_malaria<- fitp_age2 %>%
  filter(!antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","salb","sald","etec","ctb","speb")) %>%
  mutate(antigenf = droplevels(antigenf))

#-----------------------------
# create a bounding box
# to shade ages used in
# the primary analysis
# and in age-dependent FOI
# estimation
#-----------------------------
d_shade_box <- data.frame(antigenf = levels(fitp_malaria$antigenf)) %>%
  mutate(minX = 12, 
         maxX = 60,
         arm = "placebo",
         antigenf = factor(antigenf,levels(fitp_malaria$antigenf))
         )

pcols <- cbpal[c(2,6)]

plot_age_seroprev_malaria <- ggplot(data = fitp_malaria , aes(x = agem, y = fit, group = arm, color = arm, fill = arm)) +
  # facet by pathogen
  facet_grid(antigenf~.) +
  # add shade box
  geom_rect(data=d_shade_box, inherit.aes = FALSE, aes(xmin = minX, xmax = maxX, ymin = 0, ymax = 1), alpha=0.1) +
  # approximate pointwise 95% CIs
  geom_ribbon(aes(ymin = lwrS, ymax = uprS), color = NA, alpha = 0.3) +
  # spline fits
  geom_line() +
  # plot aesthetics
  scale_y_continuous(breaks = seq(0,1,by=0.2), labels = sprintf("%1.0f",seq(0,1,by=0.2)*100)) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,1.04), xlim = c(0,60)) +
  labs(x = "Age, months", y = "IgG seroprevalence (%)", tag = "b") +
  theme_minimal() +
  theme(
    strip.text.x = element_blank(),
    strip.text.y = element_blank(),
    legend.position = "top",
    plot.tag = element_text(face = "bold")
  )


```

Composite of MFI and seroprevalence
```{r malaria age plot composite, warning=FALSE, fig.width = 9, fig.height = 18}
pcurve_composite_malaria <- grid.arrange(plot_age_mfi_malaria,
                                          plot_age_seroprev_malaria,
                                          ncol=2,nrow=1,widths=c(1.5,1))

ggsave(here("figures","mordor-ab-age-curves-malaria.pdf"),pcurve_composite_malaria, device = "pdf", width = 8, height = 19)
```



Seroconversion rate figure.

```{r age-dependent FOI by arm figure malaria, warning = FALSE, fig.width = 5, fig.height = 15}

#-----------------------------
# plot age-dependent FOI
# by arm for enterics
# multiply seroconversion rate
# per month of age by 12
# to convert units to per year
#-----------------------------
foip_malaria <- fit_foi_age %>%
  mutate(antigenf = factor(antigenf,levels = levels(d_foi$antigenf)[12:20])) %>%
  filter(!is.na(antigenf)) %>%
  mutate(arm = factor(arm, levels = c("placebo","azithro")))
  

pcols <- cbpal[c(2,6)]

plot_age_foi_malaria <- ggplot(data = foip_malaria , aes(x = agem, y = 12*foi, group = arm, color = arm, fill = arm)) +
  # facet by pathogen
  facet_grid(antigenf~., scales = "free_y") +
  # approximate pointwise 95% CIs
  geom_ribbon(aes(ymin = 12*foi_lwrS, ymax = 12*foi_uprS), color = NA, alpha = 0.3) +
  # spline fits
  geom_line() +
  # plot aesthetics
  scale_y_continuous(breaks = seq(0,1,by=0.5)) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,2), xlim = c(0,60)) +
  labs(x = "Age, months", y = expression(paste("Seroconversion rate per year (force of infection), ",lambda,"(a)"))) +
  theme_minimal() +
  theme(
    strip.text.x = element_blank(),
    legend.position = "top"
  )

plot_age_foi_malaria
```

# Session Info
```{r session info}
sessionInfo()
```
