==> R/10-mordor-ab-mean-diffs.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Compare arms using MFI, seroprevalence, and SCR"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Compare mean antibody levels, seroprevalence, and seroconversion rates between the placebo and biannual azithromycin arms. 

# Preamble

```{r preamble}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
#-------------------------------
d <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public)

#-------------------------------
# drop GST negative control values
# drop trachoma and strongy because
# no transmission in this population
#-------------------------------
d1 <- d %>%
  filter(!antigen %in% c("gst","pgp3","ct694","nie")) %>%
  mutate(antigenf = factor(antigenf))

#-------------------------------
# drop cholera CTB because so 
# highly cross-reactive with ETEC LTB
#-------------------------------
d1 <- d1 %>%
  filter( !antigen %in% c("ctb") ) %>%
  mutate(antigenf = factor(antigenf))

#-------------------------------
# drop baseline measurements
#-------------------------------
d2 <- d1 %>%
  filter(phase > 0)

#-------------------------------
# age restrictions based on 
# age-antibody curves
# to exclude ages with evidence
# for maternal IgG contributions
#
# Bacteria and protozoa: 6 mo - 59 mo
#
# ETEC LTB: 6 mo - 24 mo (no heterogeneity at older ages)
#
# Malaria antigens: 12 - 59 mo
#-------------------------------
d2 <- d2 %>%
  filter(agem >= 6) %>%
  mutate(agedrop = case_when(
    antigen %in% c("etec") & (agem > 24) ~ 1,
    antigen %in% c("hrp2","csp","glurp","lsa1","ama","pfmsp1","pvmsp1","pomsp1","pmmsp1") & (agem < 12) ~ 1,
    TRUE ~ 0
  )
         ) %>%
  filter(agedrop == 0)

```

# Create composite seropositivity indicators

For enteric pathogens with multiple antigens (Cryptosporidium, Giardia, Salmonella, Campylobacter), create a series of seropositivity indicators that indicate if a child was positive to either antigen.

Then bind those to the original data frame to create a seropositivity dataset for analyses.

```{r composite seropositivity indicators}
d_enteric_comps <- d2 %>%
  filter(pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  group_by(pathogen,childid,phase) %>%
  mutate(seropos = max(seropos)) %>%
  slice(1) %>%
  mutate(antigenf = case_when(
    antigen %in% c("cp17","cp23") ~ "Cryptosporidium Cp17 or Cp23",
    antigen %in% c("vsp3","vsp5") ~ "Giardia VSP-3 or VSP-5",
    antigen %in% c("salb","sald") ~ "Salmonella groups B or D LPS",
    antigen %in% c("p18","p39") ~ "Campylobacter p18 or p39"
  )) %>%
  select(clusterid,phase,arm,childid,agem,sex,dbsid,pathogen,antigenf,seropos)


```

```{r create seropositivity dataset, warning=FALSE}
d_serop <- d2 %>%
  # drop pathogens w/ composite indicators
  filter(!pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  # add composite indicators
  bind_rows(d_enteric_comps) %>%
  # create new factor levels that includes composite indicators
  mutate(antigenf = factor(antigenf,levels = c(
    "Campylobacter p18 or p39",
    "ETEC LTB",
    "Salmonella groups B or D LPS",
    "Cryptosporidium Cp17 or Cp23",
    "Giardia VSP-3 or VSP-5",
   levels(d2$antigenf)[10:19]) )
         ) %>%
  # drop extra variables (MFI doesn't correspond to composite seropos so drop)
  dplyr::select(-mfi,-logmfi,-agedrop)

```

# Summarize the number of measurements by phase and arm

Cross-tabulate measurements by phase and arm (and overall)

```{r cross tab of n by arm}
#-----------------------------
# summarize total samples
# analyzed for IgG responses
# by phase and arm
#-----------------------------
d_n <- d1 %>%
  filter(antigen=="vsp3" & !is.na(logmfi))

dim(d_n)
table(d_n$phase)
table(d_n$arm)
table(d_n$phase,d_n$arm)


#-----------------------------
# summarize samples among 
# children ages 6-24 months
# (some force of infection analyses)
#-----------------------------
d_n2 <- d1 %>%
  filter(antigen == "vsp3" & !is.na(logmfi)) %>%
  filter(agem >=6 & agem <=24)

dim(d_n2)
table(d_n2$phase)
table(d_n2$arm)
table(d_n2$phase,d_n2$arm)

dim(d_n2 %>% filter(phase>0))
table(d_n2$arm[d_n2$phase>0])

```

# Estimate mean mfi by arm

## Means by arm
```{r means by arm}
#-----------------------------
# tally number of observations
# by antigen
#-----------------------------
n_mfi <- d2 %>%
  group_by(pathogen,antigenf,arm) %>%
  mutate(nomiss = ifelse(!is.na(logmfi),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)


#-----------------------------
# estimate means by arm
# and difference in means
#-----------------------------
mu_mfi <- d2 %>%
  group_by(pathogen,antigenf,arm) %>%
  summarize(mean_mfi = mean(logmfi), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf), names_from = "arm", values_from = mean_mfi) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_mfi <- left_join(n_mfi, mu_mfi, by=c("pathogen","antigenf"))
```

## Bootstrap 95% CIs
```{r bootstrap mean MFI diff}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in means
#-----------------------------
nreps <- 1000
ncl <- length(unique(d2$clusterid))
bsamp <- matrix(sample(unique(d2$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_mfi_boot <- d2 %>% ungroup() %>% select(clusterid,antigenf,arm,logmfi)
mu_mfi_boot <- foreach(bi = 1:nreps, .combine = rbind) %dopar% {
  set.seed(bi)
  
  # if(bi %% 50 == 0) cat(bi,"\n") else cat(".")

  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_mfi_boot, by = "clusterid")
  di %>%
  group_by(antigenf,arm) %>%
  summarize(mean_mfi = mean(logmfi), .groups = "keep") %>%
  pivot_wider(id_cols = "antigenf", names_from = "arm", values_from = mean_mfi) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)
  
}
```

```{r mean mfi bootstrap cis}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_mfi_ci <- mu_mfi_boot %>%
  group_by(antigenf) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025),
         mean_p_ub = quantile(placebo, probs = 0.975),
         mean_a_lb = quantile(azithro, probs = 0.025),
         mean_a_ub = quantile(azithro, probs = 0.975),
         diff_bmean = mean(diff),
         diff_bse = sd(diff),
         diff_lb = quantile(diff, probs = 0.025),
         diff_ub = quantile(diff, probs = 0.975)) %>%
  select(antigenf,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_mfi2 <- mu_mfi %>%
  left_join(mu_mfi_ci, by = "antigenf")
  
```

## Exact permutation P-values
```{r permutation p-values}
#-----------------------------
# estimate permutation p-values
# for the difference in means
#
# pooling over phase 6+
#
# conduct the permutation
# test on unweighted cluster
# level means
#-----------------------------
clmeans_mfi <- d2 %>% 
    group_by(clusterid,arm,pathogen,antigenf) %>%
    summarize(mean_mfi = mean(logmfi), .groups = "keep")

set.seed(87592)
permute_mfi <- foreach(abi = levels(d2$antigenf), .combine = rbind) %do% {
  di <- clmeans_mfi %>% filter(antigenf == abi)
  pti <- oneway_test(mean_mfi ~ arm, data= clmeans_mfi,subset = clmeans_mfi$antigenf==abi, distribution = "exact", alternative = "two.sided")
  data.frame(antigenf = abi, permute_p = pvalue(pti))
  }

```

## Benjamini-Hochberg FDR correction
```{r bh correction MFI}
#-----------------------------
# Add Benjamini-Hochberg FDR
# corrected p-values, nested
# within bacteria/protozoa and malaria
#-----------------------------
bh_p_enterics <- p.adjust(permute_mfi$permute_p[1:10],method = "BH")
bh_p_malaria <- p.adjust(permute_mfi$permute_p[11:19],method = "BH")

```

## Bacteria & Protozoa MFI summary table
```{r MFI means table bacteria}
mu_mfi_table <- mu_mfi2 %>%
  ungroup() %>%
  left_join(permute_mfi, by = "antigenf") %>%
  mutate(diff_ci = paste0(sprintf("%1.3f",diff)," (",sprintf("%1.3f",diff_lb),", ", sprintf("%1.3f",diff_ub),")")) %>%
  mutate(permute_p = as.numeric(permute_p)) %>%
  select(antigenf, placebo_n, placebo, azithro_n, azithro, diff_ci, permute_p) %>%
  mutate(bh_p = c(bh_p_enterics,bh_p_malaria))

knitr::kable(mu_mfi_table %>% slice(1:10),
             col.names = c("Pathogen, antigen", "N","Mean","N","Mean","Diff. (95% CI)","P-value*","P-value†"), 
             digits = c(0,0,3,0,3,0,2,2),
             align = "lccccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1, "Placebo" = 2, "Azithromycin" = 2, "Difference in log10 IgG" = 1,  " " = 2)) %>%
  footnote(
    symbol=c("Exact permutation P-value.","Permutation P-value with false discovery rate correction using the Benjamini-Hochberg method")
  )

```


## Bacteria and Protozoa MFI summary figure

```{r MFI means figure bacteria, warning=FALSE}
#------------------------------
# plot mean log10 MFI by arm
#------------------------------

# make some labels
d_labels <- data.frame(group = rep("bacteria\nand\nprotozoa",2),
                       antigenf=rep("Campylobacter p18",2),
                       arm = c("azithromycin","placebo"),
                       mean = rep(3.5,2))

# format the output for plotting
mu_mfi_placebo <- mu_mfi2 %>%
  ungroup() %>%
  select(pathogen,antigenf,mean = placebo, mean_lb = mean_p_lb, mean_ub = mean_p_ub) %>%
  mutate(arm = "placebo")
mu_mfi_azithro <- mu_mfi2 %>%
  ungroup() %>%
  select(pathogen,antigenf,mean = azithro, mean_lb = mean_a_lb, mean_ub = mean_a_ub) %>%
  mutate(arm = "azithromycin")

mu_mfi_plot <- bind_rows(mu_mfi_placebo,mu_mfi_azithro) %>%
  mutate(antigenf = factor(antigenf, levels = rev(levels(mu_mfi2$antigenf)))) %>%
  arrange(antigenf)
mu_mfi_plot$group = "bacteria\nand\nprotozoa"
mu_mfi_plot$group[grep("P. ",mu_mfi_plot$pathogen)] <- "malaria"
  
# make figure
pcols <- cbpal[c(6,2)]
plot_mfi_means_enterics <- ggplot(data = mu_mfi_plot %>% filter(group != "malaria"), aes(x = antigenf, color = arm)) +
  # plot points
  geom_errorbar(aes(ymin = mean_lb, ymax = mean_ub),width = 0.4, alpha=1,
                position = position_nudge(x = rep(c(-0.2,0.2),20))) +
  geom_point(aes(y = mean), alpha=1,
             position = position_nudge(x = rep(c(-0.2,0.2),20)) ) +
  
  scale_color_manual(values = pcols) +
  scale_x_discrete(position = "top") +
  # add annotation and labels
  geom_text(data=d_labels, aes(x = antigenf, y = mean, label=arm), position = position_nudge(x = c(0.3,-0.3) ), size = 3, hjust = 1 ) +
  labs(x = NULL, y = "Mean log10 IgG Luminex response\n(MFI-bg)") +
  
  # figure layout
  coord_flip() +
  # theme_minimal() +
  theme(
    legend.position = "none",
    strip.text.y = element_text(angle = 0)
  )

```

```{r MFI diff figure bacteria}
#------------------------------
# plot the difference between
# arms (companion to the means)
#------------------------------

# format the output for plotting
mu_mfi_plot2 <- mu_mfi2 %>%
  ungroup() %>%
  mutate(antigenf = factor(antigenf, levels = rev(levels(mu_mfi2$antigenf))))
mu_mfi_plot2$group = "bacteria\nand\nprotozoa"
mu_mfi_plot2$group[grep("P. ",mu_mfi_plot2$pathogen)] <- "malaria"

# make the figure
plot_mfi_diff_enterics <- ggplot(data = mu_mfi_plot2 %>% filter(group != "malaria"), aes(x = antigenf, y = diff)) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.4,alpha=1,color = vircols[4]) +
  geom_point(alpha=1, color = vircols[4]) +
  scale_color_manual(values = pcols) +
  scale_y_continuous(breaks = seq(-0.6,0.2,by=0.2), labels = sprintf("%1.1f",seq(-0.6,0.2,by=0.2))) +
  labs(x = NULL, y = "difference in log10 Luminex response (MFI-bg)\nazithromycin - placebo") +
  coord_flip(ylim=c(-0.7,0.3)) +
  # theme_minimal() +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank()  
    )


  
```

```{r mfi composite figure bacteria, fig.width = 10, fig.height = 5, warning = FALSE}
plot_mfi_comp_enterics <- grid.arrange(plot_mfi_means_enterics,plot_mfi_diff_enterics,ncol=2,nrow=1, widths = c(4,3))

ggsave(here("figures","mordor-ab-diff-mfi-bacteria-protozoa.pdf"),plot_mfi_comp_enterics,device = "pdf",width = 8, height = 3)
```



## Malaria MFI summary table
```{r MFI means table malaria}
knitr::kable(mu_mfi_table %>% slice(11:19),
             col.names = c("Pathogen, antigen", "N","Mean","N","Mean","Diff. (95% CI)","P-value*","P-value†"), 
             digits = c(0,0,3,0,3,0,2,2),
             align = "lccccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1, "Placebo" = 2, "Azithromycin" = 2, "Difference in log10 IgG" = 1,  " " = 2)) %>%
  footnote(symbol=c("Exact permutation P-value.","Permutation P-value with false discovery rate correction using the Benjamini-Hochberg method"))

```

## Malaria MFI summary figure

```{r MFI means figure malaria}
#------------------------------
# plot mean log10 MFI by arm
#------------------------------

# make some labels
d_labels <- data.frame(group = rep("malaria",2),
                       antigenf=rep("P. falciparum AMA1",2),
                       arm = c("azithromycin","placebo"),
                       mean = rep(2.5,2))

pcols <- cbpal[c(6,2)]
plot_mfi_means_malaria <- ggplot(data = mu_mfi_plot %>% filter(group == "malaria"), aes(x = antigenf, color = arm)) +
  # plot points
  geom_errorbar(aes(ymin = mean_lb, ymax = mean_ub),width = 0.4, alpha=1,
                position = position_nudge(x = rep(c(-0.2,0.2),20))) +
  geom_point(aes(y = mean), alpha=1,
             position = position_nudge(x = rep(c(-0.2,0.2),20)) ) +
  
  scale_color_manual(values = pcols) +
  scale_x_discrete(position = "top") +
  # add annotation and labels
  geom_text(data=d_labels, aes(x = antigenf, y = mean, label=arm), position = position_nudge(x = c(0.3,-0.3) ), size = 3, hjust = 1 ) +
  labs(x = NULL, y = "Mean log10 IgG Luminex response\n(MFI-bg)") +
  
  # figure layout
  coord_flip() +
  # theme_minimal() +
  theme(
    legend.position = "none"  
    )

```

```{r MFI diff figure malaria}
#------------------------------
# plot the difference between
# arms (companion to the means)
#------------------------------
plot_mfi_diff_malaria <- ggplot(data = mu_mfi_plot2 %>% filter(group == "malaria"), aes(x = antigenf, y = diff)) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.4,alpha=1, color = vircols[4]) +
  geom_point(alpha=1, color = vircols[4]) +
  scale_color_manual(values = pcols) +
  scale_y_continuous(breaks = seq(-0.6,0.2,by=0.2), labels = sprintf("%1.1f",seq(-0.6,0.2,by=0.2))) +
  labs(x = NULL, y = "difference in log10 Luminex response (MFI-bg)\nazithromycin - placebo") +
  coord_flip(ylim = c(-0.6,0.2)) +
  # theme_minimal() +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank() 
  )


  
```

```{r mfi composite figure malaria, fig.width = 10, fig.height = 5, warning = FALSE}
plot_mfi_comp_malaria <- grid.arrange(plot_mfi_means_malaria,plot_mfi_diff_malaria,ncol=2,nrow=1, widths = c(4,3))

ggsave(here("figures","mordor-ab-diff-mfi-malaria.pdf"),plot_mfi_comp_malaria,device = "pdf",width = 8, height = 3)
```


# Seroprevalence


## Estimate means by arm
```{r seroprev by arm}

#-----------------------------
# tally number of observations
# by antigen (including composites)
#-----------------------------
n_serop <- d_serop %>%
  group_by(pathogen,antigenf,arm) %>%
  mutate(nomiss = ifelse(!is.na(seropos),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)

#-----------------------------
# estimate seroprevalence by arm
# and difference in seroprevalence
#-----------------------------
mu_serop <- d_serop %>%
  group_by(pathogen,antigenf,arm) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_serop <- left_join(n_serop, mu_serop, by=c("pathogen","antigenf"))
```

## Bootstrap 95% CIs
```{r bootstrap mean seropos diff}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in seroprev
#-----------------------------
nreps <- 1000
ncl <- length(unique(d_serop$clusterid))
bsamp <- matrix(sample(unique(d_serop$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_serop_boot <- d_serop %>% ungroup %>% select(clusterid,antigenf,arm,seropos)
mu_serop_boot <- foreach(bi = 1:nreps, .combine = rbind) %dopar% {
  set.seed(bi)
  # if(bi %% 50 == 0) cat(bi,"\n") else cat(".")
  
  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_serop_boot, by = "clusterid")
  di %>%
  group_by(antigenf,arm) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = "antigenf", names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)

}
```

```{r mean seroprev bootstrap cis}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_serop_ci <- mu_serop_boot %>%
  group_by(antigenf) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025),
         mean_p_ub = quantile(placebo, probs = 0.975),
         mean_a_lb = quantile(azithro, probs = 0.025),
         mean_a_ub = quantile(azithro, probs = 0.975),
         diff_bmean = mean(diff),
         diff_bse = sd(diff),
         diff_lb = quantile(diff, probs = 0.025),
         diff_ub = quantile(diff, probs = 0.975)) %>%
  select(antigenf,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_serop2 <- mu_serop %>%
  left_join(mu_serop_ci, by = "antigenf")
  
```

## Exact permutation p-values
```{r permutation p-values seroprev}
#-----------------------------
# estimate permutation p-values
# for the difference in seroprev
#
# pooling over phase 6+
#
# conduct the permutation
# test on unweighted cluster
# level means
#-----------------------------
clmeans_serop <- d_serop %>% 
    group_by(clusterid,arm,pathogen,antigenf) %>%
    summarize(mean_serop = mean(seropos), .groups = "keep")

set.seed(34527)
permute_serop <- foreach(abi = levels(d_serop$antigenf), .combine = rbind) %do% {
  di <- clmeans_serop %>% filter(antigenf == abi)
  pti <- oneway_test(mean_serop ~ arm, data= clmeans_serop,subset = clmeans_serop$antigenf==abi, distribution = "exact", alternative = "two.sided")
  data.frame(antigenf = abi, permute_p = pvalue(pti))
  }

```

## Benjamini-Hochberg FDR correction
```{r bh correction seroprev}
#-----------------------------
# Add Benjamini-Hochberg FDR
# corrected p-values, nested
# within enterics and malaria
#-----------------------------
bh_p_enterics <- p.adjust(permute_serop$permute_p[1:6],method = "BH")
bh_p_malaria <- p.adjust(permute_serop$permute_p[7:15],method = "BH")

```

## Bacteria & Protozoa seroprevalence table
```{r seroprev means table bacteria}
mu_serop_table <- mu_serop2 %>%
  ungroup() %>%
  left_join(permute_serop, by = "antigenf") %>%
  mutate(diff_ci = paste0(sprintf("%1.3f",diff)," (",sprintf("%1.3f",diff_lb),", ", sprintf("%1.3f",diff_ub),")")) %>%
  mutate(permute_p = as.numeric(permute_p)) %>%
  select(antigenf,placebo_n, placebo, azithro_n, azithro, diff_ci, permute_p) %>%
  mutate(bh_p = c(bh_p_enterics,bh_p_malaria))

knitr::kable(mu_serop_table %>% slice(1:6),
             col.names = c("Pathogen, antigen", "N", "Proportion seropositive","N", "Proportion seropositive","Difference (95% CI)","P-value*","P-value†"), 
             digits = 2,
             align = "lcccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1, "Placebo" = 2,"Azithromycin" = 2, " " = 3)) %>%
  footnote(
    symbol=c("Exact permutation P-value.","Permutation P-value with false discovery rate correction using the Benjamini-Hochberg method")
  )

```

## Bacteria & Protozoa seroprevalence figure

```{r seroprev figure bacteria}
#------------------------------
# plot seroprevalence by arm
#------------------------------

# make some labels
d_labels <- data.frame(group = rep("bacteria\nand\nprotozoa",2),
                       antigenf=rep("Campylobacter p18 or p39",2),
                       arm = c("azithromycin","placebo"),
                       mean = rep(0.85,2))

# format the output for plotting
mu_serop_placebo <- mu_serop2 %>%
  ungroup() %>%
  select(pathogen,antigenf,mean = placebo, mean_lb = mean_p_lb, mean_ub = mean_p_ub) %>%
  mutate(arm = "placebo")
mu_serop_azithro <- mu_serop2 %>%
  ungroup() %>%
  select(pathogen,antigenf,mean = azithro, mean_lb = mean_a_lb, mean_ub = mean_a_ub) %>%
  mutate(arm = "azithromycin")

mu_serop_plot <- bind_rows(mu_serop_placebo,mu_serop_azithro) %>%
  mutate(antigenf = factor(antigenf, levels = rev(levels(mu_serop2$antigenf)))) %>%
  arrange(antigenf)
mu_serop_plot$group = "bacteria\nand\nprotozoa"
mu_serop_plot$group[grep("P. ",mu_serop_plot$pathogen)] <- "malaria"
  
# make figure
pcols <- cbpal[c(6,2)]
plot_serop_means_enterics <- ggplot(data = mu_serop_plot %>% filter(group != "malaria"), aes(x = antigenf, color = arm)) +
  # plot points
  geom_errorbar(aes(ymin = mean_lb, ymax = mean_ub),width = 0.4, alpha=1,
                position = position_nudge(x = rep(c(-0.2,0.2),20))) +
  geom_point(aes(y = mean), alpha=1,
             position = position_nudge(x = rep(c(-0.2,0.2),20)) ) +
  
  scale_color_manual(values = pcols) +
  scale_x_discrete(position = "top") +
  scale_y_continuous(breaks = seq(0,1,by=0.2), labels=sprintf("%1.0f",seq(0,1,by=0.2)*100)) +
  # add annotation and labels
  geom_text(data=d_labels, aes(x = antigenf, y = mean, label=arm), position = position_nudge(x = c(0.2,-0.2) ), size = 3, hjust = 1 ) +
  labs(x = NULL, y = "IgG seroprevalence (%)\n ") +
  
  # figure layout
  coord_flip(ylim=c(0,1)) +
  # theme_minimal() +
  theme(
    legend.position = "none"
  )

```

```{r seroprev diff figure enterics}
#------------------------------
# plot the difference between
# arms (companion to the means)
#------------------------------

# format the output for plotting
mu_serop_plot2 <- mu_serop2 %>%
  ungroup() %>%
  mutate(antigenf = factor(antigenf, levels = rev(levels(mu_serop2$antigenf))))
mu_serop_plot2$group = "bacteria\nand\nprotozoa"
mu_serop_plot2$group[grep("P. ",mu_serop_plot2$pathogen)] <- "malaria"

# make the figure
plot_serop_diff_enterics <- ggplot(data = mu_serop_plot2 %>% filter(group != "malaria"), aes(x = antigenf, y = diff)) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.4,alpha=1, color = vircols[4]) +
  geom_point(alpha=1, color = vircols[4]) +
  scale_color_manual(values = pcols) +
  scale_y_continuous(breaks = seq(-0.2,0.1,by=0.05), labels=sprintf("%1.0f",seq(-0.2,0.1,by=0.05)*100)) +
  labs(x = NULL, y = "Difference in seroprevalence (%)\nazithromycin - placebo") +
  coord_flip(ylim = c(-0.22,0.11)) +
  # theme_minimal() +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank() 
  )


  
```

```{r serop composite figure enterics, fig.width = 10, fig.height = 5, warning = FALSE}
plot_serop_comp_enterics <- grid.arrange(plot_serop_means_enterics,plot_serop_diff_enterics,ncol=2,nrow=1, widths = c(4,2))

ggsave(here("figures","mordor-ab-diff-seroprev-bacteria-protozoa.pdf"),plot_serop_comp_enterics,device = "pdf",width = 8, height = 2.5)
```


## Malaria seroprevalence table
```{r seroprev means table malaria}
knitr::kable(mu_serop_table %>% slice(7:15),
             col.names = c("Pathogen, antigen", "N", "Proportion seropositive","N", "Proportion seropositive","Difference (95% CI)","P-value*","P-value†"), 
             digits = 2,
             align = "lcccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1, "Placebo" = 2,"Azithromycin" = 2, " " = 3)) %>%
  footnote(
    symbol=c("Exact permutation P-value.","Permutation P-value with false discovery rate correction using the Benjamini-Hochberg method")
  )

```


## Malaria seroprevalence figure

```{r seroprev figure malaria}
#------------------------------
# plot seroprevalence by arm
#------------------------------

# make some labels
d_labels <- data.frame(group = rep("bacteria\nand\nprotozoa",2),
                       antigenf=rep("P. falciparum MSP-1",2),
                       arm = c("azithromycin","placebo"),
                       mean = rep(0.65,2))


# make figure
pcols <- cbpal[c(6,2)]
plot_serop_means_malaria <- ggplot(data = mu_serop_plot %>% filter(group == "malaria"), aes(x = antigenf, color = arm)) +
  # plot points
  geom_errorbar(aes(ymin = mean_lb, ymax = mean_ub),width = 0.4, alpha=1,
                position = position_nudge(x = rep(c(-0.2,0.2),20))) +
  geom_point(aes(y = mean), alpha=1,
             position = position_nudge(x = rep(c(-0.2,0.2),20)) ) +
  
  scale_color_manual(values = pcols) +
  scale_x_discrete(position = "top") +
  scale_y_continuous(breaks = seq(0,1,by=0.2), labels=sprintf("%1.0f",seq(0,1,by=0.2)*100)) +
  # add annotation and labels
  geom_text(data=d_labels, aes(x = antigenf, y = mean, label=arm), position = position_nudge(x = c(0.25,-0.25) ), size = 3, hjust = 1 ) +
  labs(x = NULL, y = "IgG seroprevalence (%)\n ") +
  
  # figure layout
  coord_flip(ylim=c(0,0.8)) +
  # theme_minimal() +
  theme(
    legend.position = "none"
  )

```

```{r seroprev diff figure malaria}
#------------------------------
# plot the difference between
# arms (companion to the means)
#------------------------------
# make the figure
plot_serop_diff_malaria <- ggplot(data = mu_serop_plot2 %>% filter(group == "malaria"), aes(x = antigenf, y = diff)) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.4,alpha=1,color = vircols[4]) +
  geom_point(alpha=1, color=vircols[4]) +
  scale_color_manual(values = pcols) +
  scale_y_continuous(breaks = seq(-0.15,0.1,by=0.05), labels=sprintf("%1.0f",seq(-0.15,0.1,by=0.05)*100)) +
  labs(x = NULL, y = "Difference in seroprevalence (%)\nazithromycin - placebo") +
  coord_flip() +
  # theme_minimal() +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank()
  )


  
```

```{r serop composite figure malaria, fig.width = 10, fig.height = 5, warning = FALSE}
plot_serop_comp_malaria <- grid.arrange(plot_serop_means_malaria,plot_serop_diff_malaria,ncol=2,nrow=1, widths = c(4,2))

ggsave(here("figures","mordor-ab-diff-seroprev-malaria.pdf"),plot_serop_comp_malaria,device = "pdf",width = 8, height = 2.5)
```

# Force of infection by arm

Estimate serological force of infection through the seroconverion rate. Estimate the seroconversion rate from age-structured seroprevalence.

We use a generalized linear model with a complementary log-log link fit with current status, age-prevalence data, equivalent to a proportional hazards model. We model the baseline hazard in age using a semi-parametric, cubic spline, $g(\cdot)$, and include a binary indicator for treatment group. Let $A$ be a child's age and let $T$ be a binary indicator for azithromycin treatment. The model is:

\begin{equation}
  \log - \log(1-P(Y|A,T)) = g(A) + \beta X
\end{equation}

The hazard or force of infection at age $a$ is: $\lambda(a) = F'(a)/(1-F(a))$. From the definition of the hazard, we have: 

$1- F(a) = \exp [ - \int_0^a \lambda(a) da ]$ and $\int_0^a \lambda(a) da = - \log [1-F(a)]$.

This is the cumulative hazard. The average hazard per year (units of $a$) also divided $-\log[1-F(a)]$ by $a$.  So, all of the information is embedded in the cumulative distribution function, $F(a)$, which here is the age-dependent seroprevalence curve. 

The average hazard over the age period $a_1$ to $a_2$ is:

\begin{equation}
  \int_{a_1}^{a_2} \lambda(a) da = \frac{\log[1-F(a_1)]-\log[1-F(a_2)]}{a_2 - a_1}
\end{equation}

The approach assumes no seroreversion over the age range $(a_1, a_2)$, which manifests as monotonically increasing seroprevalence with age. Limit the analysis of enteric pathogens to ages 6-24 months where this assumption is more tenable (_Salmonella_ serogroups B and D is an exception, where seroprevalence continues to increase through 59 months) 

```{r FOI  estimates}
#---------------------------------
# for enterics, drop children >24 mo
# (excluding Salmonella)
#---------------------------------
d_foi <- d_serop %>%
  mutate(dropobs = ifelse(antigenf %in% c("Campylobacter p18 or p39","ETEC LTB","Cryptosporidium Cp17 or Cp23","Giardia VSP-3 or VSP-5") & (agem > 24),1,0)) %>%
  filter(dropobs==0) %>%
  select(-dropobs) %>%
  mutate(antigenf = droplevels(antigenf),
         clusteridf = factor(clusterid),
         dummy = 1)

# print the number of observations
# included in each FOI analysis for legend reporting
(table(d_foi$antigenf))

set.seed(123)
foi_ests <- foreach(ab=levels(d_foi$antigenf),.combine=rbind) %do% {
  
    pd <- filter(d_foi, antigenf==ab)
    gfit <- gam(seropos ~ s(agem, bs="cr") + arm +
                s(clusteridf, bs = "re",by=dummy), 
                data=pd,
                family=binomial(link="cloglog"))
    gsum <- summary(gfit)
    # estimate average seroconversion rate by arm
    # use a parametric bootstrap of the posterior VCOV matrix to get inference
    newd0 <- data.frame(agem=c(min(pd$agem),max(pd$agem)), arm = "placebo", dummy=0, clusteridf = "community-1")
    newd1 <- data.frame(agem=c(min(pd$agem),max(pd$agem)), arm = "azithro", dummy=0, clusteridf = "community-1")
    lambda0 <- avgFOI(gfit, newdata=newd0, a1=min(pd$agem),a2=max(pd$agem),nreps=10000)
    lambda1 <- avgFOI(gfit, newdata=newd1, a1=min(pd$agem),a2=max(pd$agem),nreps=10000)
    # estimate HR and 95% CI from the model from the model 
    hr <- as.numeric(exp(gfit$coefficients[2]))
    loghr_se <-as.numeric(gsum$se[2])
    hr_p <- as.numeric(gsum$p.pv[2])
    # return the results
    res <- data.frame(antigenf = ab,
                      lambda0 = lambda0$mufoi,
                      lambda0_se = lambda0$mufoi_se,
                      lambda0_min95 = lambda0$mufoi_lb,
                      lambda0_max95 = lambda0$mufoi_ub,
                      lambda1 = lambda1$mufoi,
                      lambda1_se = lambda1$mufoi_se,
                      lambda1_min95 = lambda1$mufoi_lb,
                      lambda1_max95 = lambda1$mufoi_ub,
                      hr = hr, 
                      loghr_se = loghr_se,
                      hr_min95 = exp(log(hr) - 1.96*loghr_se),
                      hr_max95 = exp(log(hr) + 1.96*loghr_se),
                      hr_p = hr_p
                      )
}
```

## Benjamini-Hochberg FDR correction
```{r bh correction FOI}
#-----------------------------
# Add Benjamini-Hochberg FDR
# corrected p-values, nested
# within bacteria/protozoa and malaria
#-----------------------------
bh_p_enterics <- p.adjust(foi_ests$hr_p[1:6],method = "BH")
bh_p_malaria <- p.adjust(foi_ests$hr_p[7:15],method = "BH")

```

## Bacteria & Protozoa FOI table
```{r foi table of results enterics}

foi_ests_tab <- foi_ests %>%
  mutate(lambda0c = paste0(sprintf("%1.2f",lambda0*12)," (",
                            sprintf("%1.2f",lambda0_min95*12),", ", 
                            sprintf("%1.2f",lambda0_max95*12),")"),
         lambda1c = paste0(sprintf("%1.2f",lambda1*12)," (", 
                            sprintf("%1.2f",lambda1_min95*12),", ", 
                            sprintf("%1.2f",lambda1_max95*12),")"),
         hrc = paste0(sprintf("%1.2f",hr)," (",
                       sprintf("%1.2f",hr_min95),", ",
                       sprintf("%1.2f",hr_max95),")")
         ) %>%
  mutate(bh_p = c(bh_p_enterics, bh_p_malaria))

knitr::kable(foi_ests_tab %>% select(antigenf,lambda0c,lambda1c,hrc,hr_p,bh_p) %>% slice(1:6),digits=3,
             caption="Average seroconversion rate (SCR) per year estimated from age-structured seroprevalence",
             align = "lcccc",
             col.names = c("Antigens","Placebo SCR (95% CI)","Azithromycin SCR (95% CI)", "HR (95% CI)","P-value*","P-value†"),
             row.names = FALSE)  %>%
  kable_styling(bootstrap_options = "striped",full_width = TRUE) %>%
  footnote(symbol = c("P-value from proportional hazards model","P-value with false discovery rate correction using the Benjamini-Hochberg method"))

```

## Bacteria & Protozoa FOI figure

```{r foi ests figure bacteria}
#------------------------------
# plot SCR by arm
# estimates are multiplied by
# 12 to convert from per-month
# to per-year
#------------------------------

# make some labels
d_labels <- data.frame(group = rep("bacteria\nand\nprotozoa",2),
                       antigenf=rep("Campylobacter p18 or p39",2),
                       arm = c("azithromycin","placebo"),
                       mean = rep(2.75,2))

# format the output for plotting
mu_scr_placebo <- foi_ests %>%
  ungroup() %>%
  select(antigenf,mean = lambda0, mean_lb = lambda0_min95, mean_ub = lambda0_max95) %>%
  mutate(arm = "placebo")
mu_scr_azithro <- foi_ests %>%
  ungroup() %>%
  select(antigenf,mean = lambda1, mean_lb = lambda1_min95, mean_ub = lambda1_max95) %>%
  mutate(arm = "azithromycin")

mu_scr_plot <- bind_rows(mu_scr_placebo,mu_scr_azithro) %>%
  mutate(antigenf = factor(antigenf, levels = rev(levels(d_foi$antigenf)))) %>%
  arrange(antigenf)
mu_scr_plot$group = "bacteria\nand\nprotozoa"
mu_scr_plot$group[substr(as.character(mu_scr_plot$antigenf),1,3)=="P. "] <- "malaria"
  
# make figure (enterics)
pcols <- cbpal[c(6,2)]
plot_scr_means_enterics <- ggplot(data = mu_scr_plot %>% filter(group != "malaria"), aes(x = antigenf, color = arm)) +
  # plot points
  geom_errorbar(aes(ymin = 12*mean_lb, ymax = 12*mean_ub),width = 0.4, alpha=1,
                position = position_nudge(x = rep(c(-0.2,0.2),20))) +
  geom_point(aes(y = 12*mean), alpha=1,
             position = position_nudge(x = rep(c(-0.2,0.2),20)) ) +
  
  scale_color_manual(values = pcols) +
  scale_x_discrete(position = "top") +
  scale_y_continuous(breaks = seq(0,4,by=1)) +
  # add annotation and labels
  geom_text(data=d_labels, aes(x = antigenf, y = mean, label=arm), position = position_nudge(x = c(0.2,-0.2) ), size = 3, hjust = 0 ) +
  labs(x = NULL, y = "Force of infection\n(seroconversion rate per child-year)") +
  
  # figure layout
  coord_flip(ylim=c(0,4)) +
  # theme_minimal() +
  theme(
    legend.position = "none"
  )

```

```{r scr hazard ratio figure bacteria}
#------------------------------
# plot the hazard ratio between
# arms (companion to the means)
#------------------------------

# format the output for plotting
mu_scr_plot2 <- foi_ests %>%
  ungroup() %>%
  mutate(antigenf = factor(antigenf, levels = rev(levels(d_foi$antigenf))))
mu_scr_plot2$group = "bacteria\nand\nprotozoa"
mu_scr_plot2$group[substr(as.character(mu_scr_plot2$antigenf),1,3)=="P. "] <- "malaria"

# make the figure (enterics)
plot_scr_hr_enterics <- ggplot(data = mu_scr_plot2 %>% filter(group != "malaria"), aes(x = antigenf, y = hr)) +
  geom_hline(yintercept = 1, color = "gray60") +
  geom_errorbar(aes(ymin = hr_min95, ymax = hr_max95),width = 0.4,alpha=1, color = vircols[4]) +
  geom_point(alpha=1, color=vircols[4]) +
  scale_color_manual(values = pcols) +
  scale_y_continuous(breaks = c(0.5,0.8,1,1.2,1.5,2),trans = "log") +
  labs(x = NULL, y = "Hazard ratio \nazithromycin / placebo") +
  coord_flip(ylim=c(0.4,2)) +
  # theme_minimal() +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank(),
    panel.grid.minor = element_blank()
  )


  
```


```{r scr composite figure bacteria, fig.width = 10, fig.height = 5, warning = FALSE}
plot_scr_comp <- grid.arrange(plot_scr_means_enterics,plot_scr_hr_enterics,ncol=2,nrow=1, widths = c(4,2))

ggsave(here("figures","mordor-ab-scr-hr-bacteria-protozoa.pdf"),plot_scr_comp,device = "pdf",width = 8, height = 2.5)
```


## Malaria FOI table
```{r foi table of results malaria}
knitr::kable(foi_ests_tab %>% select(antigenf,lambda0c,lambda1c,hrc,hr_p,bh_p) %>% slice(7:15),digits=3,
             caption="Average seroconversion rate (SCR) per year estimated from age-structured seroprevalence",
             align = "lcccc",
             col.names = c("Antigens","Placebo SCR (95% CI)","Azithromycin SCR (95% CI)", "HR (95% CI)","P-value*","P-value†"),
             row.names = FALSE)  %>%
  kable_styling(bootstrap_options = "striped",full_width = TRUE) %>%
  footnote(symbol = c("P-value from proportional hazards model","P-value with false discovery rate correction using the Benjamini-Hochberg method"))

```

## Malaria FOI figure

```{r foi ests figure malaria}
#------------------------------
# plot SCR by arm
# estimates are multiplied by
# 12 to convert from per-month
# to per-year
#------------------------------

# make some labels
d_labels <- data.frame(group = rep("malaria",2),
                       antigenf=rep("P. falciparum MSP-1",2),
                       arm = c("azithromycin","placebo"),
                       mean = rep(0.3,2))

# make figure (malaria)
pcols <- cbpal[c(6,2)]
plot_scr_means_malaria <- ggplot(data = mu_scr_plot %>% filter(group == "malaria"), aes(x = antigenf, color = arm)) +
  # plot points
  geom_errorbar(aes(ymin = 12*mean_lb, ymax = 12*mean_ub),width = 0.4, alpha=1,
                position = position_nudge(x = rep(c(-0.2,0.2),20))) +
  geom_point(aes(y = 12*mean), alpha=1,
             position = position_nudge(x = rep(c(-0.2,0.2),20)) ) +
  
  scale_color_manual(values = pcols) +
  scale_x_discrete(position = "top") +
  scale_y_continuous(breaks = seq(0,0.8,by=0.2)) +
  # add annotation and labels
  geom_text(data=d_labels, aes(x = antigenf, y = mean, label=arm), position = position_nudge(x = c(0.25,-0.25) ), size = 3, hjust = 1 ) +
  labs(x = NULL, y = "Force of infection\n(seroconversion rate per child-year)") +
  
  # figure layout
  coord_flip(ylim=c(0,0.8)) +
  # theme_minimal() +
  theme(
    legend.position = "none"
  )

```

```{r scr hazard ratio figure malaria}
#------------------------------
# plot the hazard ratio between
# arms (companion to the means)
#------------------------------

# make the figure (malaria)
plot_scr_hr_malaria <- ggplot(data = mu_scr_plot2 %>% filter(group == "malaria"), aes(x = antigenf, y = hr)) +
  geom_hline(yintercept = 1, color = "gray60") +
  geom_errorbar(aes(ymin = hr_min95, ymax = hr_max95),width = 0.4,alpha=1,color = vircols[4]) +
  geom_point(alpha=1, color = vircols[4]) +
  scale_y_continuous(breaks = c(0.5,0.8,1,1.2,1.5,2),trans = "log") +
  labs(x = NULL, y = "Hazard ratio \nazithromycin / placebo") +
  coord_flip(ylim=c(0.4,2)) +
  # theme_minimal() +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank(),
    panel.grid.minor = element_blank()
  )


  
```


```{r scr composite figure malaria, fig.width = 10, fig.height = 5, warning = FALSE}
plot_scr_comp_malaria <- grid.arrange(plot_scr_means_malaria,plot_scr_hr_malaria,ncol=2,nrow=1, widths = c(4,2))

ggsave(here("figures","mordor-ab-scr-hr-malaria.pdf"),plot_scr_comp_malaria,device = "pdf",width = 8, height = 2.5)
```

# Save estimates for comparison with longitudinal ests

```{r save foi estimates}
write_csv(foi_ests_tab,here("data","mordor-ab-cross-sectional-foi-ests.csv"))
```

# Session Info
```{r session info}
sessionInfo()
```
