==> R/12-mordor-ab-longitudinal.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Longitudinal estimates of seroconversion"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Estimate prospective seroconversion rates among children measured longitudinally. Assume that events ocurred at the midpoint of the interval, consistent with the mortality outcome measurement and with past longitudinal analyses of enteric seronconversion.

Keenan, J. D., Bailey, R. L., West, S. K., Arzika, A. M., Hart, J., Weaver, J., Kalua, K., Mrango, Z., Ray, K. J., Cook, C., Lebas, E., O’Brien, K. S., Emerson, P. M., Porco, T. C., Lietman, T. M. & MORDOR Study Group. Azithromycin to Reduce Childhood Mortality in Sub-Saharan Africa. _N. Engl. J. Med._ 378, 1583–1592 (2018). https://pubmed.ncbi.nlm.nih.gov/29694816/

Arnold, B. F., Martin, D. L., Juma, J., Mkocha, H., Ochieng, J. B., Cooley, G. M., Omore, R., Goodhew, E. B., Morris, J. F., Costantini, V., Vinjé, J., Lammie, P. J. & Priest, J. W. Enteropathogen antibody dynamics and force of infection among children in low-resource settings. _eLife_ 8, (2019). https://pubmed.ncbi.nlm.nih.gov/31424386/

# Preamble

```{r preamble}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
#-------------------------------
d <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public)

#-------------------------------
# drop GST negative control values
# drop trachoma and strongy because
# no transmission in this population
#-------------------------------
d1 <- d %>%
  filter(!antigen %in% c("gst","pgp3","ct694","nie")) %>%
  mutate(antigenf = factor(antigenf))

#-------------------------------
# drop cholera CTB because so 
# highly cross-reactive with ETEC LTB
#-------------------------------
d2 <- d1 %>%
  filter( !antigen %in% c("ctb") ) %>%
  mutate(antigenf = factor(antigenf))

#-------------------------------
# age restrictions based on 
# age-antibody curves
# to exclude ages without any
# variation in antibody levels
# 
# For enterics, limit from birth
# to 24 months.
#
# (except for salmonella, strep A, 6 - 59 mo)
#
# Malaria antigens: 12 - 59 mo
#-------------------------------
d2 <- d2 %>%
  mutate(agedrop = case_when(
    antigen %in% c("cp17","cp23","vsp3","vsp5","p18","p39","etec","ctb") & (agem > 24) ~ 1,
    antigen %in% c("salb","sald","speb") & (agem <6) ~ 1,
    antigen %in% c("hrp2","csp","glurp","lsa1","ama","pfmsp1","pvmsp1","pomsp1","pmmsp1") & (agem < 12) ~ 1,
    TRUE ~ 0
  )
         ) %>%
  filter(agedrop == 0)

```

# Identify children measured longitudinally

Identify children without longitudinal measurements. Tally them, then exclude them.

```{r identify longitudinal children}

#-------------------------------
# identify the number of observations
# for each child
#-------------------------------
d_nolong <- d2 %>%
  group_by(antigenf,childid) %>%
  arrange(phase) %>%
  mutate(nobs = n())

#-------------------------------
# print number of children with 
# different numbers of measurements
# for enterics (6-24 mos) 
# salmonella/strep (6-59 mos)
# and malaria (12-59 mos)
#-------------------------------
table(d_nolong$nobs[d_nolong$antigen=="etec"])/1:4
table(d_nolong$nobs[d_nolong$antigen=="salb"])/1:5
table(d_nolong$nobs[d_nolong$antigen=="ama"])/1:5

#-------------------------------
# print number of observations 
# excluded by outcome
#-------------------------------
table(d_nolong$antigenf,d_nolong$nobs>1)

#-------------------------------
# restrict to children with >1 obs
#-------------------------------
dl <- d_nolong %>%
  filter(nobs>1) %>%
  arrange(antigenf,childid,phase) %>%
  mutate(phase_diff = phase - lag(phase))

#-----------------------------
# inspect children who have
# negative age differences
# between visits
# there appear to be 115
# children whose age difference
# was <=0 between phases
# this appears to be a data
# error. Drop them from the
# longitudinal analysis
#-----------------------------
dl <- dl %>%
  group_by(antigenf,childid) %>%
  arrange(antigenf, childid,phase) %>%
  mutate(age_diff = agem - lag(agem),
         min_agediff = min(age_diff, na.rm = TRUE))
length(unique(dl$childid[dl$min_agediff<=0]))
dl <- dl %>%
  filter(min_agediff >0) %>%
  select(-age_diff,-min_agediff)

# observations by antigen and phase
table(dl$antigenf,dl$phase)

# observations by antigen and months between phases
# most repeated measures are 6-12 months apart
table(dl$antigenf,dl$phase_diff)


#-------------------------------
# print final number of children with 
# different numbers of measurements
# for enterics (6-24 mos) 
# salmonella/strep (6-59 mos)
# and malaria (12-59 mos)
#-------------------------------

table(dl$antigenf)

# Enterics
table(dl$nobs[dl$antigen=="etec"])/2:4
sum(table(dl$nobs[dl$antigen=="etec"])/2:4)

# Salmonella and Streptococcus
table(dl$nobs[dl$antigen=="salb"])/2:5
sum(table(dl$nobs[dl$antigen=="salb"])/2:5)

# malaria
table(dl$nobs[dl$antigen=="ama"])/2:5
sum(table(dl$nobs[dl$antigen=="ama"])/2:5)

#-------------------------------
# summarize the median and
# range of measurements per child
# in the three longitudinal analysis
# sets
#-------------------------------
dl_enterics <- dl %>%
  filter(antigen=="etec") %>%
  group_by(childid) %>%
  slice(1) 
summary(dl_enterics$nobs)
dl_sal <- dl %>%
  filter(antigen=="salb") %>%
  group_by(childid) %>%
  slice(1) 
summary(dl_sal$nobs)
dl_mal <- dl %>%
  filter(antigen=="ama") %>%
  group_by(childid) %>%
  slice(1) 
summary(dl_mal$nobs)

  

```

# Create composite seropositivity indicators

For enteric pathogens with multiple antigens (Cryptosporidium, Giardia, Salmonella, Campylobacter), create a series of seropositivity indicators that indicate if a child was positive to either antigen.

Then bind those to the original data frame to create a seropositivity dataset for analyses.

```{r composite seropositivity indicators, eval = FALSE}
d_enteric_comps <- dl %>%
  filter(pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  group_by(pathogen,childid,phase) %>%
  mutate(seropos = max(seropos)) %>%
  slice(1) %>%
  mutate(antigenf = case_when(
    antigen %in% c("cp17","cp23") ~ "Cryptosporidium Cp17 or Cp23",
    antigen %in% c("vsp3","vsp5") ~ "Giardia VSP-3 or VSP-5",
    antigen %in% c("salb","sald") ~ "Salmonella groups B or D LPS",
    antigen %in% c("p18","p39") ~ "Campylobacter p18 or p39"
  )) %>%
  select(clusterid,phase,arm,childid,agem,sex,dbsid,pathogen,antigenf,seropos)


```

```{r create seropositivity dataset, warning=FALSE, eval = FALSE}
d_serop <- dl %>%
  # drop pathogens w/ composite indicators
  filter(!pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  # add composite indicators
  bind_rows(d_enteric_comps) %>%
  # create new factor levels that includes composite indicators
  mutate(antigenf = factor(antigenf,levels = c(
    "Campylobacter p18 or p39",
    "ETEC LTB",
    "Salmonella groups B or D LPS",
    "Cryptosporidium Cp17 or Cp23",
    "Giardia VSP-3 or VSP-5",
   levels(d2$antigenf)[10:19]) )
         ) %>%
  # drop extra variables (MFI doesn't correspond to composite seropos so drop)
  dplyr::select(-mfi,-logmfi,-agedrop)

```



# Plot longitudinal IgG trajectories

Identify different trajectories, based on whether children experienced a 4-fold increase or decrease in IgG to above or below the seropositivity cutoff.  This is consistent with the analysis from Arnold et al. 2019, which used a definition of "incident boosting" as a 4-fold increase in IgG to above the seropositivity cutoff, and "incident waning" as a 4-fold decrease in IgG from above the seropositivity cutoff.
https://pubmed.ncbi.nlm.nih.gov/31424386/ 
https://osf.io/2j5fe/

```{r identify igg changes}

#-----------------------------
# identify incident 
# seroconversions and reversions
# in the dataset with 
# measurements separated by
# antigen (i.e., not combined responses for enterics)
# this is done to visualize
# the 4-fold changes in IgG
# which only make sense for
# individual antigens (not combined)
#-----------------------------

# group the data by child and
# use lags to identify
# time between measurements,
# sero-conversions + sero-reversions 
# between measurements
# set the first measurement to 
# missing for the incidence indicators
# (using the is.na(age_diff) to identify them)
dl2 <- dl %>%
  group_by(antigenf,childid) %>% 
  arrange(antigenf,childid,phase) %>%
  mutate(age_min  = min(agem),
         age_diff = agem - lag(agem),
         
         logmfi_lag  = lag(logmfi),
         logmfi_lead = lead(logmfi),
         logmfi_dlag  = logmfi - logmfi_lag,
         logmfi_dlead = logmfi_lead - logmfi,
         
         # incident seroconversions and reversions
         # including cumulative numbers
         seropos_lag  = lag(seropos),
         seroi = ifelse(seropos==1 & seropos_lag==0,1,0),
         seroi = ifelse(is.na(age_diff),NA,seroi),
         seroin = cumsum(ifelse(is.na(seroi),0,seroi)),
         seroin = ifelse(seroi==1,seroin,0),
         seror = ifelse(seropos==0 & seropos_lag==1,1,0),
         seror = ifelse(is.na(age_diff),NA,seror),
         serorn = cumsum(ifelse(is.na(seror),0,seror)),
         serorn = ifelse(seror==1,serorn,0)
         ) %>%
  select(
         clusterid,childid,arm,phase,
         starts_with("age"),
         antigen,antigenf,
         mfi,logmfi,logmfi_dlag,logmfi_dlead,
         serocut,seropos,
         seroi,seroin,seror,serorn,
         -logmfi_lag,-logmfi_lead,-seropos_lag)

#-----------------------------
# alternate definition
# incident boosting based on a 4-fold increase in MFI
# with a second measure above the seropositivity cutoff
# incident waning based on a 4-fold decrease in MFI
# with the first measure above the seropositivity cutoff
#-----------------------------
dl2 <- dl2 %>%
  mutate(seroi4fold = ifelse(logmfi_dlag>log10(4) & logmfi>log10(serocut),1,0),
         seroin4fold = cumsum(ifelse(is.na(seroi4fold),0,seroi4fold)),
         seroin4fold = ifelse(seroi4fold==1,seroin4fold,0),
         
         seror4fold=ifelse(logmfi_dlag< -log10(4) & lag(logmfi)>log10(serocut),1,0),
         serorn4fold = cumsum(ifelse(is.na(seror4fold),0,seror4fold)),
         serorn4fold = ifelse(seror4fold==1,serorn4fold,0)
         )

#-----------------------------
# create a label for different
# antibody patterns
#-----------------------------
dl2pp <- dl2 %>%
  arrange(antigenf,childid,phase) %>%
  group_by(antigenf,childid) %>%
  mutate(measnum=row_number()) %>%
  mutate(
    logmfi_chng = ifelse(logmfi_dlag>0,"Increasing","Decreasing"),
    logmfi_chng = ifelse(is.na(logmfi_chng) & logmfi_dlead>0,"Increasing",logmfi_chng),
    logmfi_chng = ifelse(is.na(logmfi_chng) & logmfi_dlead<0,"Decreasing",logmfi_chng),
    logmfi_chng = factor(logmfi_chng,levels=c("Increasing","Decreasing")),
    
    sero_chng = ifelse(lead(seroi4fold)==1,">4-fold increase to above cutoff","<4-fold change"),
    sero_chng = ifelse(lead(seror4fold)==1,">4-fold decrease from above cutoff",sero_chng),
    sero_chng = ifelse(is.na(sero_chng),lead(sero_chng),sero_chng),
    sero_chng = ifelse(is.na(sero_chng),lag(sero_chng),sero_chng),
    sero_chng = factor(sero_chng,levels=c(">4-fold increase to above cutoff",">4-fold decrease from above cutoff","<4-fold change")),
    
    sero_comp = ifelse(lead(seroi)==1 & lead(seroi4fold==1),">4-fold increase, across cutoff","<4-fold change"),
    sero_comp = ifelse(lead(seroi)==0 & lead(seroi4fold==1),">4-fold increase, above cutoff",sero_comp),
    sero_comp = ifelse(lead(seror4fold)==1,">4-fold decrease",sero_comp),
    sero_comp = factor(sero_comp,levels=c(">4-fold increase, across cutoff",">4-fold increase, above cutoff",">4-fold decrease","<4-fold change")),
    
    everseroi = max(seroi,na.rm=TRUE)
    
  ) 


```

Plot IgG levels by phase for individual children

```{r plot mfi by phase, fig.width = 8, fig.height = 12}

# custom log labels
log10labs <- c( 
  expression(10^0),
  expression(10^1),
  expression(10^2),
  expression(10^3),
  expression(10^4)
)


pmfi_long <- ggplot(data = dl2pp, aes(x = phase, y = logmfi, group = childid, color = sero_comp)) +
  facet_wrap(~antigenf,ncol=4) +
  # first draw all the lines in light grey
  geom_line(alpha=0.2,color = "gray90") +
  # then draw the lines with color 
  # (setting <4-fold change color to missing to better display the >4-fold change lines)
  geom_line(alpha=0.25) +
  geom_hline(aes(yintercept = log10(serocut)),linetype = "dashed") +
  scale_color_manual(values=c(cbpal[c(2,8,3)],NA),
                     guide=guide_legend(title="MFI change:",
                                        override.aes = list(alpha=1), nrow=2,ncol=3))+
  scale_y_continuous(breaks = 0:4, labels = log10labs) +
  scale_x_continuous(breaks=c(0,6,12,24,36)) + 
  labs(x = "study phase (months since baseline)", y = "Luminex response (MFI-bg)") +
  theme_minimal() +
  theme(
      legend.position="top",
      panel.grid.minor=element_blank(),
      panel.grid.major.y=element_blank(),
      axis.ticks.y=element_line(color="gray40")
    )

pmfi_long

# save a pdf file
ggsave(filename = here("figures","mordor-ab-long-trajectories.pdf"),device = "pdf",width=8,height=12)
```

# Estimate prospective rates

Estimate seroconversion rates and seroreversion rates for each pathogen. For enterics, use combined responses and seropositivity to determine seroconversion and seroreversion. For malaria, examine individual antibody responses.

## Identify incident seroconversion and seroreversion

Defined as an increase in IgG to above seropositivity cutoff (seroconversion) or a decrease in IgG from above to below seropositivity cutoff. 

```{r identify incident seroconversion and seroreversions}

#-----------------------------
# create composite
# seroprevalence indicators
# that use information from
# multiple antigens if available
# for the enterics
#-----------------------------
dl3 <- dl2 %>%
  ungroup() %>%
  mutate(antigenf2 = antigenf,
         antigenf2 = case_when(antigen %in% c("p18","p39") ~ "Campylobacter p18 or p39",
                               antigen %in% c("salb","sald") ~ "Salmonella groups B or D LPS",
                               antigen %in% c("cp17","cp23") ~ "Cryptosporidium Cp17 or Cp23",
                               antigen %in% c("vsp3","vsp5") ~ "Giardia VSP-3 or VSP-5",
                               TRUE ~ as.character(antigenf)
                              ),
         antigenf2 = factor(antigenf2, levels = c(
           "Campylobacter p18 or p39",
           "ETEC LTB",
           "Salmonella groups B or D LPS",
           "Cryptosporidium Cp17 or Cp23",
           "Giardia VSP-3 or VSP-5",
           levels(d2$antigenf)[10:19]) )
         ) %>%
  select(antigenf2,antigenf,clusterid,childid,arm,phase,agem,seropos) %>%
  group_by(antigenf2,childid,phase) %>%
  mutate(seropos=max(seropos)) %>%
  slice(1) %>%
  select(antigenf=antigenf2,clusterid,childid,arm,phase,agem,seropos) 



  # mutate(nobs=n(),obsnum=row_number(),
  #        agediff=ifelse(obsnum==1,lead(agem)-agem,agem-lag(agem)),
  #        
#-----------------------------
# group the data by child and
# use lags to identify
# time in years between measurements,
# sero-conversions + sero-reversions 
# between measurements
# set the first measurement to 
# missing for the incidence indicators
# (using the is.na(age_diff) to identify them)
#-----------------------------
dl4 <- dl3 %>%
  group_by(antigenf,childid) %>% 
  arrange(antigenf,childid,phase) %>%
  mutate(obsnum = row_number()) %>%
  mutate(age_min  = min(agem),
         age_diff = agem - lag(agem),
         
         # incident seroconversions and reversions
         seropos_lag  = lag(seropos),
         seroi = ifelse(seropos==1 & seropos_lag==0,1,0),
         seroi = ifelse(is.na(age_diff),NA,seroi),
         seror = ifelse(seropos==0 & seropos_lag==1,1,0),
         seror = ifelse(is.na(age_diff),NA,seror)
         )

#-----------------------------
# estimate time at risk
# for seroconversion and reversion
# assumed to be 1/2 of time
# between measurements
# if indivs are seropositive
# at measurement 1 they are
# not at risk for seroconversion
# (and vice-versa for seroreversion)
#
# convert differences in age
# in months to years for
# eventual rate estimates per-year
#-----------------------------
dl5 <- dl4 %>%
  mutate(ptc = ifelse(seropos==0 & seror==0,age_diff,0),
         ptc = ifelse(seropos==1 & seroi==1,age_diff/2,ptc),
         ptc = ptc/12,
         ptr = ifelse(seropos==1 & seroi==0,age_diff,0),
         ptr = ifelse(seropos==0 & seror==1,age_diff/2,ptr),
         ptr = ptr/12)

```

## Estimate prospective rates

```{r estimate prospective rates}
#-----------------------------
# estimate sero-incidence rates
# for antibody boosting and waning 
# estimate SEs with a bootstrap
#-----------------------------

#-----------------------------
# estimate incidence rates
# for each antibody
# units are episodes per child-year
#-----------------------------
rate_ests <- dl5 %>%
  group_by(antigenf) %>%
  summarize(ni=sum(seroi,na.rm=T),
            nit=sum(ptc,na.rm=T),
            nr=sum(seror,na.rm=T),
            nrt=sum(ptr,na.rm=T),
            .groups = "keep") %>%
  mutate(seroi=ni/nit,seror=nr/nrt)

#-----------------------------
# get bootstrap CIs
# resampling clusters with 
# replacement due to repeated
# measures
#-----------------------------
dboot <- dl5 %>%
  group_by(antigenf) %>%
  select(clusterid,antigenf,seroi,seror,ptc,ptr)

nreps <- 1000
ids <- unique(dboot$clusterid)
bsamp <- matrix(sample(ids,size=length(ids)*nreps,replace=TRUE),
                nrow=length(ids),ncol=nreps)
bootests <- foreach(ab=levels(dboot$antigenf),.combine=rbind) %:%
  foreach(brep=1:nreps,.combine=rbind) %dopar% {
    set.seed(brep)
    di <- dboot %>% filter(antigenf==ab)
    di <- left_join(data.frame(clusterid=bsamp[,brep]),di,by=c("clusterid")) %>%
      select(-clusterid) %>% group_by(antigenf) %>%
      summarize_all(function(x) sum(x,na.rm=TRUE))
    }

rate_cis <- bootests %>% 
  group_by(antigenf) %>%
  mutate(sero_c=seroi/ptc,sero_r=seror/ptr) %>%
  # restrict to finite values (some ETEC bootstrap replicates have a denominator of 0)
  summarize(seroi_lb = quantile(sero_c[is.finite(sero_c)],probs=c(0.025),na.rm=T),
            seroi_ub = quantile(sero_c[is.finite(sero_c)],probs=c(0.975),na.rm=T),
            seror_lb = quantile(sero_r[is.finite(sero_r)],probs=c(0.025),na.rm=T),
            seror_ub = quantile(sero_r[is.finite(sero_r)],probs=c(0.975),na.rm=T), 
            .groups = "keep"
            )

rate_ests <- left_join(rate_ests,rate_cis,by="antigenf") %>%
  select(antigenf,ni,nit,starts_with("seroi"),nr,nrt,starts_with("seror"))

```

Confirm bootstrap distributions of the rates are smooth and reasonable
```{r check boot distributions, fig.width = 8, fig.height = 12}
bsum <- bootests %>% 
  group_by(antigenf) %>%
  mutate(sero_c=seroi/ptc,sero_r=seror/ptr)

ggplot(data=bsum,aes(x=sero_c)) + geom_density() + facet_wrap(~antigenf,ncol=2,scales="free") + labs(title = "Seroconversion") + theme_minimal()

ggplot(data=bsum,aes(x=sero_r)) + geom_density() + facet_wrap(~antigenf,ncol=2,scales="free") + labs(title = "Seroreversion") + theme_minimal()

```

## Table of results

```{r seroconversion rate table}
rate_ests_tab <- rate_ests[c(7:15,1:6),] %>%
  select(antigenf,nit,ni,seroi,seroi_lb,seroi_ub,nrt,nr,seror,seror_lb,seror_ub)


knitr::kable(rate_ests_tab,
             digits=2,
             caption="Longitudinal incidence rates of seroconversion and seroreversion per child year, MORDOR Niger, 2015-2018.",
             col.names = c("Pathogen, antigen(s)","Years at risk","n events","Incidence per year","min95","max95","Years at risk","n events","Incidence per year","min95","max95"),
             row.names = FALSE) %>%
  kableExtra::kable_styling(bootstrap_options = c("striped"),full_width = TRUE) %>%
  kableExtra::add_header_above(c(" " = 1, "Seroconversion *" = 5, "Seroreversion †" = 5)) %>% 
  kableExtra::group_rows("Malaria", 1,9) %>%
  kableExtra::group_rows("Bacteria and protozoans", 10,15) %>%
  kableExtra::footnote(symbol = c("Incident seroconversion defined by an increase in IgG levels (MFI-bg) to above the seropositivity cutoff.", "Incident seroreversion defined by a decrease in IgG levels to below the seropositivy cutoff.")
           )

```

## Save longitudinal rate estimates for other comparisons

Save longitudinal rates pooled over arms (total population)

```{r save long rates}
write_csv(rate_ests,here("data","mordor-ab-longitudinal-rate-ests.csv"))
```

# Estimate rates by arm

Estimate prospective incidence rates and incidence rate ratio by arm as a sensitivity analysis for the primary analysis that estimates the hazard ratio from a current status analysis of age-structured seroprevalence data.

```{r estimate prospective rates by arm}
#-----------------------------
# estimate sero-incidence rates
# for antibody boosting and waning 
# by arm
# estimate SEs with a bootstrap
#-----------------------------

#-----------------------------
# estimate incidence rates
# for each antibody
# units are episodes per child-year
#-----------------------------
rate_ests_by_arm <- dl5 %>%
  group_by(antigenf,arm) %>%
  summarize(ni=sum(seroi,na.rm=T),
            nit=sum(ptc,na.rm=T),
            nr=sum(seror,na.rm=T),
            nrt=sum(ptr,na.rm=T),
            .groups = "keep") %>%
  mutate(seroi=ni/nit,seror=nr/nrt)

#-----------------------------
# get bootstrap CIs
# resampling clusters with 
# replacement due to repeated
# measures
#-----------------------------
dboot <- dl5 %>%
  group_by(antigenf,arm) %>%
  select(clusterid,antigenf,arm,seroi,seror,ptc,ptr)

nreps <- 1000
ids <- unique(dboot$clusterid)
bsamp <- matrix(sample(ids,size=length(ids)*nreps,replace=TRUE),
                nrow=length(ids),ncol=nreps)
bootests <- foreach(ab=levels(dboot$antigenf),.combine=rbind) %:%
  foreach(brep=1:nreps,.combine=rbind) %dopar% {
    set.seed(brep)
    di <- dboot %>% filter(antigenf==ab)
    di <- left_join(data.frame(clusterid=bsamp[,brep]),di,by=c("clusterid")) %>%
      select(-clusterid) %>% group_by(antigenf,arm) %>%
      summarize_all(function(x) sum(x,na.rm=TRUE)) %>%
      mutate(brep = brep)
    }

# estimate bootstrap 95% CIs for rates
rate_cis_by_arm <- bootests %>% 
  group_by(antigenf,arm) %>%
  mutate(sero_c = seroi/ptc,
         sero_r = seror/ptr) %>%
  # restrict to finite values (some ETEC bootstrap replicates have a denominator of 0)
  summarize(seroi_lb = quantile(sero_c[is.finite(sero_c)],probs=c(0.025),na.rm=T),
            seroi_ub = quantile(sero_c[is.finite(sero_c)],probs=c(0.975),na.rm=T),
            seror_lb = quantile(sero_r[is.finite(sero_r)],probs=c(0.025),na.rm=T),
            seror_ub = quantile(sero_r[is.finite(sero_r)],probs=c(0.975),na.rm=T), 
            .groups = "keep"
            )

rate_ests_by_arm <- left_join(rate_ests_by_arm,rate_cis_by_arm,by=c("antigenf","arm")) %>%
  select(antigenf,arm,ni,nit,starts_with("seroi"),nr,nrt,starts_with("seror"))

# Estimate bootstrapped 95% CI for IRR
rate_boot_p <- bootests %>% 
  filter(arm=="placebo") %>% 
  select(antigenf, brep,
         seroi_p = seroi, ptc_p = ptc,
         seror_p = seror, ptr_p = ptr)

rate_boot_irr <- bootests %>% 
  filter(arm=="azithro") %>%
  select(antigenf, brep,
         seroi_a = seroi, ptc_a = ptc,
         seror_a = seror, ptr_a = ptr) %>%
  left_join(rate_boot_p, by = c("antigenf","brep")) %>%
  mutate(irate_azithro = seroi_a / ptc_a,
         irate_placebo = seroi_p / ptc_p,
         rrate_azithro = seror_a / ptr_a,
         rrate_placebo = seror_p / ptr_p,
         irri = irate_azithro / irate_placebo,
         isrr = rrate_azithro / rrate_placebo)

irr_cis <- rate_boot_irr %>%
  group_by(antigenf) %>%
  summarize(irr_lb = quantile(irri[is.finite(irri)],probs = c(0.025)),
            irr_ub = quantile(irri[is.finite(irri)],probs = c(0.975)), 
            srr_lb = quantile(isrr[is.finite(isrr)],probs = c(0.025)),
            srr_ub = quantile(isrr[is.finite(isrr)],probs = c(0.975)),
            .groups = "keep")

irr_ests_p <- rate_ests_by_arm %>%
  ungroup() %>%
  filter(arm=="placebo") %>%
  select(antigenf,ni_p = ni, nit_p = nit, seroi_p = seroi, seroi_lb_p = seroi_lb, seroi_ub_p = seroi_ub, nr_p = nr, nrt_p = nrt, seror_p = seror, seror_lb_p = seror_lb, seror_ub_p = seror_ub)
irr_ests <- rate_ests_by_arm %>%
  ungroup() %>%
  filter(arm == "azithro") %>%
  select(antigenf, ni_a = ni, nit_a = nit, seroi_a = seroi, seroi_lb_a = seroi_lb, seroi_ub_a = seroi_ub,  nr_a = nr, nrt_a = nrt, seror_a = seror, seror_lb_a = seror_lb, seror_ub_a = seror_ub) %>%
  left_join(irr_ests_p, by = "antigenf") %>%
  mutate(irr = seroi_a / seroi_p,
         srr = seror_a / seror_p) %>%
  left_join(irr_cis, by = "antigenf")

```

## Summary table - seroconversion rates
```{r summary irr table}

#-----------------------------
# format rates and IRRs for
# printing
# put malaria on top
#-----------------------------
irr_ests_tab <- irr_ests[c(7:15,1:6),] %>%
  mutate(seroi_a_print = paste0(sprintf("%1.2f",seroi_a)," (",sprintf("%1.2f",seroi_lb_a),", ",sprintf("%1.2f",seroi_ub_a),")"),
         seroi_p_print = paste0(sprintf("%1.2f",seroi_p)," (",sprintf("%1.2f",seroi_lb_p),", ",sprintf("%1.2f",seroi_ub_p),")"),
         irr_print = paste0(sprintf("%1.2f",irr)," (",sprintf("%1.2f",irr_lb),", ",sprintf("%1.2f",irr_ub),")"),
         
         seror_a_print = paste0(sprintf("%1.2f",seror_a)," (",sprintf("%1.2f",seror_lb_a),", ",sprintf("%1.2f",seror_ub_a),")"),
         seror_p_print = paste0(sprintf("%1.2f",seror_p)," (",sprintf("%1.2f",seror_lb_p),", ",sprintf("%1.2f",seror_ub_p),")"),
         srr_print = paste0(sprintf("%1.2f",srr)," (",sprintf("%1.2f",srr_lb),", ",sprintf("%1.2f",srr_ub),")"),
         ) %>%
  # set zero rates and infinite IRRs to zero / NA
  mutate(seroi_a_print = ifelse(ni_a == 0, "0", seroi_a_print),
         seroi_p_print = ifelse(ni_p == 0, "0", seroi_p_print),
         seror_a_print = ifelse(nr_a == 0, "0", seror_a_print),
         seror_p_print = ifelse(nr_p == 0, "0", seror_p_print),
         irr_print = ifelse(ni_a == 0 | ni_p == 0, "n/a", irr_print),
         srr_print = ifelse(nr_a == 0 | nr_p == 0, "n/a", srr_print)
         )



knitr::kable(irr_ests_tab %>% select(antigenf,ni_a,nit_a,seroi_a_print,ni_p,nit_p,seroi_p_print,irr_print),
             caption = "Seroconversion rates estimated longitudinally among children in MORDOR Niger, 2015 - 2018.",
             col.names = c("Pathogen, antigen", "N events","Person-Years","Rate per year (95% CI)","N events","Person-Years","Rate per year (95% CI)", "IRR (95% CI)"), 
             digits = 2,
             align = "lrrcrrcc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1, "Azithromycin" = 3,  "Placebo" = 3, " " = 1)) %>%
  kableExtra::column_spec(column = c(4,7,8), width = "150px") %>%
  kableExtra::group_rows(group_label = "Malaria", start_row = 1, end_row = 9) %>%
  kableExtra::group_rows(group_label = "Bacteria & protozoa", start_row = 10, end_row = 15) 

```
## Summary table - seroreversion rates
```{r summary srr table}

knitr::kable(irr_ests_tab %>% select(antigenf,nr_a,nrt_a,seror_a_print,nr_p,nrt_p,seror_p_print,srr_print),
             caption = "Seroreversion rates estimated longitudinally among children in MORDOR Niger, 2015 - 2018.",
             col.names = c("Pathogen, antigen", "N events","Person-Years","Rate per year (95% CI)","N events","Person-Years","Rate per year (95% CI)", "IRR (95% CI)"), 
             digits = 2,
             align = "lrrcrrcc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1, "Azithromycin" = 3,  "Placebo" = 3, " " = 1)) %>%
  kableExtra::column_spec(column = c(4,7,8), width = "150px") %>%
  kableExtra::group_rows(group_label = "Malaria", start_row = 1, end_row = 9) %>%
  kableExtra::group_rows(group_label = "Bacteria & protozoa", start_row = 10, end_row = 15) 

```

## Save longitudinal rate estimates by arm

Save longitudinal rates by arm for other comparisons

```{r save long irr ests}
write_csv(irr_ests,here("data","mordor-ab-longitudinal-irr-ests.csv"))
```

# Rate comparison

Compare longitudinal estimates with the cross-sectional (primary) analysis.

## Load saved estimates

Load saved cross-sectional estimates, and harmonize the data

# Load the saved estimates

```{r load saved estimates}
#-------------------------------
# load the cross-sectional estimates
# mordor-ab-mean-diffs.Rmd
#-------------------------------
dc <- read_csv(here("data","mordor-ab-cross-sectional-foi-ests.csv"))


#-------------------------------
# create a summary dataset
# of rates by arm
# the cross-sectional estimates
# are in units per-month, so 
# multiply them by 12 before
# combining
# 
# rename HR -> IRR
#-------------------------------
dc2 <- dc %>%
  mutate(seroi_a = lambda1*12,
         seroi_lb_a = lambda1_min95*12,
         seroi_ub_a = lambda1_max95*12,
         seroi_p = lambda0*12,
         seroi_lb_p = lambda0_min95*12,
         seroi_ub_p = lambda0_max95*12,
  ) %>%
  select(antigenf,starts_with("seroi_"),irr=hr,irr_lb=hr_min95,irr_ub=hr_max95) %>%
  mutate(analysis = "cross-sectional")

#-------------------------------
# stack the longitudinal and
# cross-sectional estimates
# convert labels to factor
#-------------------------------
d <- irr_ests %>%
  mutate(analysis = "longitudinal") %>%
  bind_rows(dc2) %>%
  mutate(analysis = factor(analysis, levels = c("cross-sectional","longitudinal"))) %>%
  mutate(antigenf = factor(antigenf,levels = 
                             c("Campylobacter p18 or p39",
                               "ETEC LTB",
                               "Salmonella groups B or D LPS",
                               "Cryptosporidium Cp17 or Cp23",
                               "Giardia VSP-3 or VSP-5",
                               "Streptococcus group A SPEB",
                               "P. falciparum MSP-1",
                               "P. falciparum AMA1",
                               "P. falciparum GLURP-Ro",
                               "P. falciparum LSA1",
                               "P. falciparum CSP",
                               "P. falciparum HRP2",
                               "P. malariae MSP-1",
                               "P. ovale MSP-1",
                               "P. vivax MSP-1"
                             )
                           )
         )
d$group = "enterics\nand\nstrep"
d$group[substr(as.character(d$antigenf),1,3)=="P. "] <- "malaria"

#-------------------------------
# tediously reshape intervention
# group to long format to 
# make it easier to plot estimates
#-------------------------------
da <- d %>%
  select(group,antigenf,analysis,ni=ni_a,nit=nit_a,seroi=seroi_a,seroi_lb = seroi_lb_a, seroi_ub = seroi_ub_a, nr = nr_a, nrt = nrt_a, seror = seror_a, seror_lb = seror_lb_a, seror_ub = seror_ub_a,irr,irr_lb,irr_ub,srr,srr_lb,srr_ub) %>%
  mutate(arm = "azithromycin")

dp <- d %>%
  select(group,antigenf,analysis,ni=ni_p,nit=nit_p,seroi=seroi_p,seroi_lb = seroi_lb_p, seroi_ub = seroi_ub_p, nr = nr_p, nrt = nrt_p, seror = seror_p, seror_lb = seror_lb_p, seror_ub = seror_ub_p) %>%
  mutate(arm = "placebo")

d2 <- da %>%
  bind_rows(dp) %>%
  mutate(arm = factor(arm, levels = c("placebo","azithromycin")))

```




## Malaria comparison figure

Seroconversion rate panel

```{r plot seroconversion rates malaria}

d_plot_rates <- d2 %>%
  filter(group=="malaria") %>%
  mutate(antigenf = factor(antigenf, levels=rev(levels(d2$antigenf)))) %>%
  arrange(antigenf,arm,desc(analysis))

plot_scr_malaria <- ggplot(data = d_plot_rates, aes(x = antigenf, y = seroi,color = analysis, shape= analysis )) +
  facet_grid(.~arm) +
  geom_errorbar(aes(ymin = seroi_lb, ymax = seroi_ub), width=0.2, position = position_nudge(c(-0.1,0.1))) + 
  geom_point(position = position_nudge(c(-0.1,0.1))) +
  scale_x_discrete(position = "top") +
  scale_y_continuous(breaks=seq(0,1,by=0.2)) + 
  scale_color_manual(values = c("black",cbpal[2])) +
  labs(x=NULL,y="Seroconversion rate per child-year\n") +
  coord_flip(ylim=c(0,1)) +
  theme_minimal() +
  theme(
    legend.position = "left"
  )


```

Hazard ratio panel

```{r plot HRs malaria}

d_plot_hr <- d2 %>%
  filter(group=="malaria" & arm == "azithromycin") %>%
  mutate(antigenf = factor(antigenf, levels=rev(levels(d2$antigenf)))) %>%
  arrange(antigenf,desc(analysis))


plot_hr_malaria <- ggplot(data = d_plot_hr, aes(x = antigenf, y = irr,color = analysis, shape= analysis )) +
  facet_grid(.~arm) +
  geom_hline(yintercept = 1, color = "gray70") +
  geom_errorbar(aes(ymin = irr_lb, ymax = irr_ub), width=0.2, position = position_nudge(c(-0.1,0.1))) + 
  geom_point(position = position_nudge(c(-0.1,0.1))) +
  scale_y_continuous(breaks=c(0.2,0.5,1,2,3), trans = "log") +
  scale_color_manual(values = c("black",cbpal[2])) +
  labs(x=NULL,y="HR (cross-sectional) or IRR (longitudinal) \nazithromycin / placebo") +
  coord_flip() +
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.y  = element_blank(),
    strip.text.x = element_text(color="white")
  )


```

Composite plot
```{r plot composite malaria, fig.height=3, fig.width = 12}
malaria_comp <- grid.arrange(plot_scr_malaria,plot_hr_malaria,ncol=2,nrow=1,widths=c(3,1))

ggsave(filename = here("figures","mordor-ab-long-v-cross-malaria.pdf"),malaria_comp,device = "pdf",height=3,width=10)

```

## Bacteria, protozoa comparison figure

Seroconversion rate panel

```{r plot seroconversion rates bacteria protozoa}

d_plot_rates <- d2 %>%
  filter(group!="malaria") %>%
  mutate(antigenf = factor(antigenf, levels=rev(levels(d2$antigenf)))) %>%
  arrange(antigenf,arm,desc(analysis))

plot_scr_enterics <- ggplot(data = d_plot_rates, aes(x = antigenf, y = seroi,color = analysis, shape= analysis )) +
  facet_grid(.~arm) +
  geom_errorbar(aes(ymin = seroi_lb, ymax = seroi_ub), width=0.2, position = position_nudge(c(-0.1,0.1))) + 
  geom_point(position = position_nudge(c(-0.1,0.1))) +
  scale_x_discrete(position = "top") +
  scale_color_manual(values = c("black",cbpal[2])) +
  labs(x=NULL,y="Seroconversion rate per child-year\n") +
  coord_flip() +
  theme_minimal() +
  theme(
    legend.position = "left"
  )


```

Hazard/rate ratio panel

```{r plot HRs bacteria}

d_plot_hr <- d2 %>%
  filter(group!="malaria" & arm == "azithromycin") %>%
  mutate(antigenf = factor(antigenf, levels=rev(levels(d2$antigenf)))) %>%
  arrange(antigenf,desc(analysis))


plot_hr_enterics <- ggplot(data = d_plot_hr, aes(x = antigenf, y = irr,color = analysis, shape= analysis )) +
  facet_grid(.~arm) +
  geom_hline(yintercept = 1, color = "gray70") +
  geom_errorbar(aes(ymin = irr_lb, ymax = irr_ub), width=0.2, position = position_nudge(c(-0.1,0.1))) + 
  geom_point(position = position_nudge(c(-0.1,0.1))) +
  scale_y_continuous(breaks=c(0.5,1,1.5,2), trans = "log") +
  scale_color_manual(values = c("black",cbpal[2])) +
  labs(x=NULL,y="HR (cross-sectional) or IRR (longitudinal) \nazithromycin / placebo") +
  coord_flip() +
  theme_minimal() +
  theme(
    legend.position = "non",
    axis.title.y = element_blank(),
    axis.text.y  = element_blank(),
    strip.text.x = element_text(color="white")
  )


```

Composite plot
```{r plot composite bacteria protozoa, fig.height=3, fig.width = 12}
enterics_comp <- grid.arrange(plot_scr_enterics,plot_hr_enterics,ncol=2,nrow=1,widths=c(3,1))

ggsave(filename = here("figures","mordor-ab-long-v-cross-bacteria-protozoa.pdf"),enterics_comp,device = "pdf",height=3,width=10)

```

# Session Info
```{r session info}
sessionInfo()
```
