==> R/13-mordor-ab-subgroup-analyses.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Subgroup analyses"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Compare mean antibody levels and seroprevalence by pre-specified subgroups: trial phase (6, 12, 24, 36) and by child age (age 6 months or younger at the start of the trial vs older).  The rationale for these subgroups is specified in the pre-analysis plan: https://osf.io/d9s4t/


# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
#-------------------------------
d <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public)

#-------------------------------
# drop GST negative control values
# drop trachoma and strongy because
# no transmission in this population
#-------------------------------
d2 <- d %>%
  filter(!antigen %in% c("gst","pgp3","ct694","nie")) %>%
  mutate(antigenf = factor(antigenf),
         phasef = factor(phase))

#-------------------------------
# drop cholera CTB because so 
# highly cross-reactive with ETEC LTB
#-------------------------------
d2 <- d2 %>%
  filter( !antigen %in% c("ctb") ) %>%
  mutate(antigenf = factor(antigenf))

#-------------------------------
# drop baseline measurements
#-------------------------------
d2 <- d2 %>%
  filter(phase > 0) %>%
  mutate(phasef = droplevels(phasef))

#-------------------------------
# age restrictions based on 
# age-antibody curves
# to exclude ages with evidence
# for maternal IgG contributions
#
# Bacteria and protozoa: 6 mo - 59 mo
#
# ETEC LTB: 6 mo - 24 mo (no heterogeneity at older ages)
#
# Malaria antigens: 12 - 59 mo
#-------------------------------
d2 <- d2 %>%
  filter(agem >= 6) %>%
  mutate(agedrop = case_when(
    antigen %in% c("etec") & (agem > 24) ~ 1,
    antigen %in% c("hrp2","csp","glurp","lsa1","ama","pfmsp1","pvmsp1","pomsp1","pmmsp1") & (agem < 12) ~ 1,
    TRUE ~ 0
  )
         ) %>%
  filter(agedrop == 0) %>%
  ungroup()

```

# Create composite seropositivity indicators

For enteric pathogens with multiple antigens (Cryptosporidium, Giardia, Salmonella, Campylobacter), create a series of seropositivity indicators that indicate if a child was positive to either antigen.

Then bind those to the original data frame to create a seropositivity dataset for analyses.

```{r composite seropositivity indicators}
d_enteric_comps <- d2 %>%
  filter(pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  group_by(pathogen,childid,phasef) %>%
  mutate(seropos = max(seropos)) %>%
  slice(1) %>%
  mutate(antigenf = case_when(
    antigen %in% c("cp17","cp23") ~ "Cryptosporidium Cp17 or Cp23",
    antigen %in% c("vsp3","vsp5") ~ "Giardia VSP-3 or VSP-5",
    antigen %in% c("salb","sald") ~ "Salmonella groups B or D LPS",
    antigen %in% c("p18","p39") ~ "Campylobacter p18 or p39"
  )) %>%
  select(clusterid,phasef,arm,childid,agem,agem_start_under6,sex,dbsid,pathogen,antigenf,seropos)


```

```{r create seropositivity dataset, warning=FALSE}
d_serop <- d2 %>%
  # drop pathogens w/ composite indicators
  filter(!pathogen %in% c("Cryptosporidium","Giardia","Salmonella","Campylobacter")) %>%
  # add composite indicators
  bind_rows(d_enteric_comps) %>%
  # create new factor levels that includes composite indicators
  mutate(antigenf = factor(antigenf,levels = c(
    "Campylobacter p18 or p39",
    "ETEC LTB",
    "Salmonella groups B or D LPS",
    "Cryptosporidium Cp17 or Cp23",
    "Giardia VSP-3 or VSP-5",
   levels(d2$antigenf)[10:19]) )
         ) %>%
  # drop extra variables (MFI doesn't correspond to composite seropos so drop)
  select(-mfi,-logmfi,-agedrop)

```

# Subgroup analysis by study phase

Assess effect modification by trial phase on MFI and seroprevalence.


## MFI by phase and arm

```{r means by arm and phase}
#-----------------------------
# tally number of observations
# by antigen
#-----------------------------
n_mfi <- d2 %>%
  group_by(pathogen,antigenf,arm,phasef) %>%
  mutate(nomiss = ifelse(!is.na(logmfi),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf,phasef), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)


#-----------------------------
# estimate means by arm, phase
# and difference in means
#-----------------------------
mu_mfi <- d2 %>%
  group_by(pathogen,antigenf,arm,phasef) %>%
  summarize(mean_mfi = mean(logmfi), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf,phasef), names_from = "arm", values_from = mean_mfi) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_mfi <- left_join(n_mfi, mu_mfi, by=c("pathogen","antigenf","phasef"))
```

```{r bootstrap mean MFI diff by phase}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in means
#-----------------------------
nreps <- 1000
ncl <- length(unique(d2$clusterid))
bsamp <- matrix(sample(unique(d2$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_mfi_boot <- d2 %>% select(clusterid,antigenf,phasef,arm,logmfi)
mu_mfi_boot <- foreach(bi = 1:nreps, .combine = rbind) %dopar% {
  set.seed(bi)
  # if(bi %% 50 == 0) cat(bi,"\n") else cat(".")

  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_mfi_boot, by = "clusterid")
  di %>%
  group_by(antigenf,phasef,arm) %>%
  summarize(mean_mfi = mean(logmfi), .groups = "keep") %>%
  pivot_wider(id_cols = c("antigenf","phasef"), names_from = "arm", values_from = mean_mfi) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)
  
}
```

```{r mean mfi bootstrap cis by phase}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_mfi_ci <- mu_mfi_boot %>%
  group_by(antigenf,phasef) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025),
         mean_p_ub = quantile(placebo, probs = 0.975),
         mean_a_lb = quantile(azithro, probs = 0.025),
         mean_a_ub = quantile(azithro, probs = 0.975),
         diff_bmean = mean(diff),
         diff_bse = sd(diff),
         diff_lb = quantile(diff, probs = 0.025),
         diff_ub = quantile(diff, probs = 0.975)) %>%
  select(antigenf,phasef,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_mfi2 <- mu_mfi %>%
  left_join(mu_mfi_ci, by = c("antigenf","phasef"))
  
```

### Estimate interaction p-values

Estimate interaction p-values on the additive scale using an ANOVA on cluster level means.

```{r mfi p-values for interaction phase}

#-----------------------------
# test for interaction on
# the additive scale using
# ANOVA F-test on cluster level means
# See Hayes and Moulten 2017, 
# section 10.8 (p241)
#-----------------------------
clmeans_mfi <- d2 %>% 
    group_by(antigenf,arm,phasef,clusterid) %>%
    summarize(mean_mfi = mean(logmfi), .groups = "keep")

mfi_phase_Ftest <- foreach(abi = levels(d2$antigenf), .combine = rbind) %do% {
    # restrict to a particular antigen
    # summarize cluster level means by phase and arm
    di <- clmeans_mfi %>% filter(antigenf == abi) 
  
    # ANOVA F-test on cluster level means
    # See Hayes and Moulten 2017, section 10.8 (p241)
    fiti <- lm(mean_mfi ~ arm*phasef, data=di)
    intFP <- as.numeric(anova(fiti)[3,4:5])
    res <- data.frame(antigenf=abi,phasef="36",intF=intFP[1],intP=intFP[2])
    
  }

#-----------------------------
# within pathogen group calc.
# Benjamini-Hochberg adjusted
# p-values, allowing for a 5% FDR.
#-----------------------------
mfi_phase_Ftest$intPbh <- c(p.adjust(mfi_phase_Ftest$intP[1:10]), 
                            p.adjust(mfi_phase_Ftest$intP[11:19])
                            )

```

### Table of results
```{r MFI means by phase table}
options(knitr.kable.NA = '') 
mu_mfi_table <- mu_mfi2 %>%
  ungroup() %>%
  left_join(mfi_phase_Ftest,by=c("antigenf","phasef")) %>%
  mutate(diff_ci = paste0(sprintf("%1.3f",diff)," (",sprintf("%1.3f",diff_lb),", ", sprintf("%1.3f",diff_ub),")")) %>%
  select(antigenf, phasef, placebo_n, placebo, azithro_n, azithro, diff_ci,intP,intPbh)

knitr::kable(mu_mfi_table,
             col.names = c("Pathogen, antigen", "Phase","N","Mean","N","Mean","log10 MFI-bg (95% CI)","P-value*","Adj P-value†"), 
             digits = 3,
             align = "lccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 2, "Placebo" = 2, "Azithromycin" = 2, "Difference in" = 1," " = 2)) %>%
  kableExtra::group_rows(group_label = "Bacteria and protozoa", start_row = 1, end_row = 40) %>%
  kableExtra::group_rows(group_label = "Malaria", start_row = 41, end_row = 76) %>%
  footnote(
    symbol=c("P-value for additive interaction by study phase.", "P-value for additive interaction, adjusted for a 5% False Discovery Rate using the Benjamini-Hochberg procedure.")
  )

```

### Figure of results

```{r MFI diff phase figure enterics}
#------------------------------
# plot the difference between
# arms by phase
#------------------------------

# format the output for plotting
mu_mfi_plot2 <- mu_mfi2 %>%
  ungroup() #%>%
  # mutate(antigenf = factor(antigenf, levels = rev(levels(mu_mfi2$antigenf))))
mu_mfi_plot2$group = "bacteria\nand\nprotozoa"
mu_mfi_plot2$group[grep("P. ",mu_mfi_plot2$pathogen)] <- "malaria"

# make the figure
pcols <- brewer.pal(9,"PuBuGn")[5:8]
plot_mfi_diff_phase_enterics <- ggplot(data = mu_mfi_plot2 %>% filter(group!="malaria"), aes(x = phasef, y = diff, color = phasef, fill = phasef)) +
  facet_wrap(antigenf~., ncol=4, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.8,0.8,by=0.4), labels = sprintf("%1.1f",seq(-0.8,0.8,by=0.4))) +
  scale_color_manual(values = pcols) +
  labs(x = "Study phase (months)", y = "Difference in log10 Luminex response (MFI-bg)\nazithromycin - placebo", tag = "a") +
  coord_cartesian(ylim = c(-0.8,0.8)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r MFI diff phase figure malaria}
#------------------------------
# plot the difference between
# arms by phase
#------------------------------
pcols <- brewer.pal(9,"PuBuGn")[5:8]
plot_mfi_diff_phase_malaria <- ggplot(data = mu_mfi_plot2 %>% filter(group=="malaria"), aes(x = phasef, y = diff, color = phasef, fill = phasef)) +
  facet_wrap(antigenf~., ncol=3, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.8,0.8,by=0.4), labels = sprintf("%1.1f",seq(-0.8,0.8,by=0.4))) +
  scale_color_manual(values = pcols) +
  labs(x = "Study phase (months)", y = "Difference in log10 Luminex response (MFI-bg)\nazithromycin - placebo", tag = "b") +
  coord_cartesian(ylim = c(-0.8,0.8)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r mfi diff phase composite figure, fig.width = 10, fig.height = 6}
plot_mfi_phase_comp <- grid.arrange(plot_mfi_diff_phase_enterics,plot_mfi_diff_phase_malaria,ncol=2,nrow=1, widths = c(1,0.75))

ggsave(here("figures","mordor-ab-diff-mfi-by-phase.pdf"),plot_mfi_phase_comp,device = "pdf",width = 12, height = 4)
```

## Seroprevalence by phase and arm

```{r seroprev by arm phase}

#-----------------------------
# tally number of observations
# by antigen (including composites)
#-----------------------------
n_serop <- d_serop %>%
  group_by(pathogen,antigenf,arm, phasef) %>%
  mutate(nomiss = ifelse(!is.na(seropos),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf, phasef), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)

#-----------------------------
# estimate seroprevalence by arm
# and difference in seroprevalence
#-----------------------------
mu_serop <- d_serop %>%
  group_by(pathogen,antigenf,arm, phasef) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf, phasef), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_serop <- left_join(n_serop, mu_serop, by=c("pathogen","antigenf", "phasef"))
```


```{r bootstrap mean seropos diff}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in seroprev
#-----------------------------
nreps <- 1000
ncl <- length(unique(d_serop$clusterid))
bsamp <- matrix(sample(unique(d_serop$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_serop_boot <- d_serop %>% select(clusterid,antigenf,arm,phasef,seropos)
mu_serop_boot <- foreach(bi = 1:nreps, .combine = rbind) %dopar% {
  set.seed(bi)
  # if(bi %% 50 == 0) cat(bi,"\n") else cat(".")
  
  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_serop_boot, by = "clusterid")
  di %>%
  group_by(antigenf,arm,phasef) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c("antigenf","phasef"), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)

}
```


```{r mean seroprev bootstrap cis}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_serop_ci <- mu_serop_boot %>%
  group_by(antigenf,phasef) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025),
         mean_p_ub = quantile(placebo, probs = 0.975),
         mean_a_lb = quantile(azithro, probs = 0.025),
         mean_a_ub = quantile(azithro, probs = 0.975),
         diff_bmean = mean(diff),
         diff_bse = sd(diff),
         diff_lb = quantile(diff, probs = 0.025),
         diff_ub = quantile(diff, probs = 0.975)) %>%
  select(antigenf,phasef,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_serop2 <- mu_serop %>%
  left_join(mu_serop_ci, by = c("antigenf","phasef"))
  
```

### Estimate interaction p-values

Estimate interaction p-values on the additive scale using an ANOVA on cluster level means.

```{r seroprev p-values for interaction phase}

#-----------------------------
# test for interaction on
# the additive scale using
# ANOVA F-test on cluster level means
# See Hayes and Moulten 2017, 
# section 10.8 (p241)
#-----------------------------
clmeans_serop <- d_serop %>% 
    group_by(antigenf,arm,phasef,clusterid) %>%
    summarize(seroprev = mean(seropos), .groups = "keep")

serop_phase_Ftest <- foreach(abi = levels(d_serop$antigenf), .combine = rbind) %do% {
    # restrict to a particular antigen
    # summarize cluster level means by phase and arm
    di <- clmeans_serop %>% filter(antigenf == abi) 
  
    # ANOVA F-test on cluster level means
    # See Hayes and Moulten 2017, section 10.8 (p241)
    fiti <- lm(seroprev ~ arm*phasef, data=di)
    intFP <- as.numeric(anova(fiti)[3,4:5])
    res <- data.frame(antigenf=abi,phasef="36",intF=intFP[1],intP=intFP[2])
    
  }

#-----------------------------
# within pathogen group calc.
# Benjamini-Hochberg adjusted
# p-values, allowing for a 5% FDR.
#-----------------------------
serop_phase_Ftest$intPbh <- c(p.adjust(serop_phase_Ftest$intP[1:6]), 
                            p.adjust(serop_phase_Ftest$intP[7:15])
                            )
```


### Table of results
```{r seroprev means by phase table}
mu_serop_table <- mu_serop2 %>%
  ungroup() %>%
  left_join(serop_phase_Ftest, by = c("antigenf","phasef")) %>%
  mutate(diff_ci = paste0(sprintf("%1.2f",diff)," (",sprintf("%1.2f",diff_lb),", ", sprintf("%1.2f",diff_ub),")")) %>%
  select(antigenf,phasef, placebo_n, placebo, azithro_n, azithro, diff_ci, intP, intPbh)

knitr::kable(mu_serop_table,
             col.names = c("Pathogen, antigen","Phase", "N","Proportion\nSeropositive","N","Proportion\nSeropositive","Difference (95% CI)","P-value*","Adj P-value†"), 
             digits = c(0,0,0,2,0,2,0,3,3),
             align = "lccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 2, "Placebo" = 2, "Azithromycin" = 2, " " = 3)) %>%
  kableExtra::group_rows(group_label = "Bacteria and protozoa", start_row = 1, end_row = 24) %>%
  kableExtra::group_rows(group_label = "Malaria", start_row = 25, end_row = 60) %>%
  footnote(
    symbol=c("P-value for additive interaction by study phase.", "P-value for additive interaction, adjusted for a 5% False Discovery Rate using the Benjamini-Hochberg procedure.")
  )

```


### Figure of results

```{r serop diff phase figure bacteria protozoa}
#------------------------------
# plot the difference between
# arms by phase
#------------------------------

# format the output for plotting
mu_serop_plot2 <- mu_serop2 %>%
  ungroup() #%>%
  # mutate(antigenf = factor(antigenf, levels = rev(levels(mu_mfi2$antigenf))))
mu_serop_plot2$group = "bacteria\nand\nprotozoa"
mu_serop_plot2$group[grep("P. ",mu_serop_plot2$pathogen)] <- "malaria"

# make the figure
pcols <- brewer.pal(9,"PuBuGn")[5:8]
plot_serop_diff_phase_enterics <- ggplot(data = mu_serop_plot2 %>% filter(group!="malaria"), aes(x = phasef, y = diff, color = phasef, fill = phasef)) +
  facet_wrap(antigenf~., ncol=3, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.3,0.3,by=0.1), labels = sprintf("%1.0f",100*seq(-0.3,0.3,by=0.1))) +
  scale_color_manual(values = pcols) +
  labs(x = "Study phase (months)", y = "Difference in seroprevalence (%)\nazithromycin - placebo", tag = "a") +
  coord_cartesian(ylim = c(-0.3,0.3)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r seroprev diff phase figure malaria}
#------------------------------
# plot the difference between
# arms by phase
#------------------------------
pcols <- brewer.pal(9,"PuBuGn")[5:8]
plot_serop_diff_phase_malaria <- ggplot(data = mu_serop_plot2 %>% filter(group=="malaria"), aes(x = phasef, y = diff, color = phasef, fill = phasef)) +
  facet_wrap(antigenf~., ncol=3, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.3,0.3,by=0.1), labels = sprintf("%1.0f",100*seq(-0.3,0.3,by=0.1))) +
  scale_color_manual(values = pcols) +
  labs(x = "Study phase (months)", y = "Difference in seroprevalence (%)\nazithromycin - placebo", tag = "b") +
  coord_cartesian(ylim = c(-0.3,0.3)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )


```

```{r serop composite figure, fig.width = 10, fig.height = 6}
plot_serop_comp_phase <- grid.arrange(plot_serop_diff_phase_enterics,plot_serop_diff_phase_malaria,ncol=2,nrow=1, widths = c(1,1))

ggsave(here("figures","mordor-ab-diff-seroprev-by-phase.pdf"),plot_serop_comp_phase,device = "pdf",width = 13, height = 5)
```

# Subgroup analysis by age at first treatment

Since IgG response increases cumulatively with age, we hypothesized that children who were older than 6 months at the trial's first treatment date might already have IgG responses that were high from previous infections. This pre-specified analysis examines the effect of intervention on MFI and seroprevalence, stratified by whether the child was $\leq$ 6 months at the first treatment (phase 0).

Note, the code below largely repeats the subgroup analysis by phase, so many object names are re-used.


## MFI by age at trial start and arm

```{r means by arm and age}
#-----------------------------
# tally number of observations
# by antigen
#-----------------------------
n_mfi <- d2 %>%
  group_by(pathogen,antigenf,arm,agem_start_under6) %>%
  mutate(nomiss = ifelse(!is.na(logmfi),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf,agem_start_under6), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)


#-----------------------------
# estimate means by arm, age group
# and difference in means
#-----------------------------
mu_mfi <- d2 %>%
  group_by(pathogen,antigenf,arm,agem_start_under6) %>%
  summarize(mean_mfi = mean(logmfi), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf,agem_start_under6), names_from = "arm", values_from = mean_mfi) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_mfi <- left_join(n_mfi, mu_mfi, by=c("pathogen","antigenf","agem_start_under6"))
```


```{r bootstrap mean MFI diff by age}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in means
#-----------------------------
nreps <- 1000
ncl <- length(unique(d2$clusterid))
bsamp <- matrix(sample(unique(d2$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_mfi_boot <- d2 %>% select(clusterid,antigenf,agem_start_under6,arm,logmfi)
mu_mfi_boot <- foreach(bi = 1:nreps, .combine = rbind) %dopar% {
  set.seed(bi)
  # if(bi %% 50 == 0) cat(bi,"\n") else cat(".")

  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_mfi_boot, by = "clusterid")
  di %>%
  group_by(antigenf,agem_start_under6,arm) %>%
  summarize(mean_mfi = mean(logmfi), .groups = "keep") %>%
  pivot_wider(id_cols = c("antigenf","agem_start_under6"), names_from = "arm", values_from = mean_mfi) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)
  
}
```

```{r mean mfi bootstrap cis by age}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_mfi_ci <- mu_mfi_boot %>%
  group_by(antigenf,agem_start_under6) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025),
         mean_p_ub = quantile(placebo, probs = 0.975),
         mean_a_lb = quantile(azithro, probs = 0.025),
         mean_a_ub = quantile(azithro, probs = 0.975),
         diff_bmean = mean(diff),
         diff_bse = sd(diff),
         diff_lb = quantile(diff, probs = 0.025),
         diff_ub = quantile(diff, probs = 0.975)) %>%
  select(antigenf,agem_start_under6,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_mfi2 <- mu_mfi %>%
  left_join(mu_mfi_ci, by = c("antigenf","agem_start_under6"))
  
```



### Estimate interaction p-values

Estimate interaction p-values on the additive scale using an ANOVA on cluster level means.

```{r mfi p-values for interaction by age group}

#-----------------------------
# test for interaction on
# the additive scale using
# ANOVA F-test on cluster level means
# See Hayes and Moulten 2017, 
# section 10.8 (p241)
#-----------------------------
clmeans_mfi <- d2 %>% 
    group_by(antigenf,arm,agem_start_under6,clusterid) %>%
    summarize(mean_mfi = mean(logmfi), .groups = "keep")

mfi_age_Ftest <- foreach(abi = levels(d2$antigenf), .combine = rbind) %do% {
    # restrict to a particular antigen
    # summarize cluster level means by phase and arm
    di <- clmeans_mfi %>% filter(antigenf == abi) 
  
    # ANOVA F-test on cluster level means
    # See Hayes and Moulten 2017, section 10.8 (p241)
    fiti <- lm(mean_mfi ~ arm*agem_start_under6, data=di)
    intFP <- as.numeric(anova(fiti)[3,4:5])
    res <- data.frame(antigenf=abi,agem_start_under6="Yes",intF=intFP[1],intP=intFP[2])
    
  }

#-----------------------------
# within pathogen group calc.
# Benjamini-Hochberg adjusted
# p-values, allowing for a 5% FDR.
#-----------------------------
mfi_age_Ftest$intPbh <- c(p.adjust(mfi_age_Ftest$intP[1:10]), 
                            p.adjust(mfi_age_Ftest$intP[11:19])
                            )
```

### Table of results
```{r MFI means by age group table}
options(knitr.kable.NA = '') 
mu_mfi_table <- mu_mfi2 %>%
  ungroup() %>%
  left_join(mfi_age_Ftest,by=c("antigenf","agem_start_under6")) %>%
  mutate(diff_ci = paste0(sprintf("%1.3f",diff)," (",sprintf("%1.3f",diff_lb),", ", sprintf("%1.3f",diff_ub),")")) %>%
  select(antigenf, agem_start_under6, placebo_n, placebo, azithro_n, azithro, diff_ci,intP, intPbh)

knitr::kable(mu_mfi_table,
             col.names = c("Pathogen, antigen", "≤6 months at trial start","N","Mean","N","Mean","log10 MFI-bg (95% CI)","P-value*","Adj P-value†"), 
             digits = 3,
             align = "lccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 2, "Placebo" = 2, "Azithromycin" = 2,  "Difference in" = 1, " " = 2)) %>%
  kableExtra::group_rows(group_label = "Bacteria and protozoa", start_row = 1, end_row = 20) %>%
  kableExtra::group_rows(group_label = "Malaria", start_row = 21, end_row = 38) %>%
  footnote(
    symbol=c("P-value for additive interaction by age at trial start.", "P-value for additive interaction, adjusted for a 5% False Discovery Rate using the Benjamini-Hochberg procedure.")
  )
```


### Figure of results

```{r MFI diff age figure bacteria protozoa}
#------------------------------
# plot the difference between
# arms by age group
#------------------------------

# format the output for plotting
mu_mfi_plot2 <- mu_mfi2 %>%
  ungroup() #%>%
  # mutate(antigenf = factor(antigenf, levels = rev(levels(mu_mfi2$antigenf))))
mu_mfi_plot2$group = "bacteria\nand\nprotozoa"
mu_mfi_plot2$group[grep("P. ",mu_mfi_plot2$pathogen)] <- "malaria"

# make the figure
pcols <- brewer.pal(9,"PuBuGn")[7:8]
plot_mfi_diff_age_enterics <- ggplot(data = mu_mfi_plot2 %>% filter(group!="malaria"), aes(x = agem_start_under6, y = diff, color = agem_start_under6, fill = agem_start_under6)) +
  facet_wrap(antigenf~., ncol=4, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=0.6) +
  geom_point(alpha=0.6) +
  scale_y_continuous(breaks = seq(-0.8,0.8,by=0.4), labels = sprintf("%1.1f",seq(-0.8,0.8,by=0.4))) +
  scale_color_manual(values = pcols) +
  labs(x = "Under 6 months at trial start", y = "difference in log10 Luminex response (MFI-bg)\nazithromycin - placebo", tag = "a") +
  coord_cartesian(ylim = c(-0.8,0.8)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r MFI diff age figure malaria}
#------------------------------
# plot the difference between
# arms by age group
#------------------------------
pcols <- brewer.pal(9,"PuBuGn")[7:8]
plot_mfi_diff_age_malaria <- ggplot(data = mu_mfi_plot2 %>% filter(group=="malaria"), aes(x = agem_start_under6, y = diff, color = agem_start_under6, fill = agem_start_under6)) +
  facet_wrap(antigenf~., ncol=3, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.6,0.6,by=0.3), labels = sprintf("%1.1f",seq(-0.6,0.6,by=0.3))) +
  scale_color_manual(values = pcols) +
  labs(x = "Under 6 months at trial start", y = "difference in log10 Luminex response (MFI-bg)\nazithromycin - placebo", tag = "b") +
  coord_cartesian(ylim = c(-0.6,0.6)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r mfi diff age composite figure, fig.width = 10, fig.height = 6}
plot_mfi_age_comp <- grid.arrange(plot_mfi_diff_age_enterics,plot_mfi_diff_age_malaria,ncol=2,nrow=1, widths = c(1,0.75))

ggsave(here("figures","mordor-ab-diff-mfi-by-age-group.pdf"),plot_mfi_age_comp,device = "pdf",width = 12, height = 4)
```


## Seroprevalence by age and arm

```{r seroprev by arm age}

#-----------------------------
# tally number of observations
# by antigen (including composites)
#-----------------------------
n_serop <- d_serop %>%
  group_by(pathogen,antigenf,arm, agem_start_under6) %>%
  mutate(nomiss = ifelse(!is.na(seropos),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf, agem_start_under6), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)

#-----------------------------
# estimate seroprevalence by arm
# and difference in seroprevalence
#-----------------------------
mu_serop <- d_serop %>%
  group_by(pathogen,antigenf,arm, agem_start_under6) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c(pathogen, antigenf, agem_start_under6), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_serop <- left_join(n_serop, mu_serop, by=c("pathogen","antigenf", "agem_start_under6"))
```


```{r bootstrap mean seropos diff by age}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in seroprev
#-----------------------------
nreps <- 1000
ncl <- length(unique(d_serop$clusterid))
bsamp <- matrix(sample(unique(d_serop$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_serop_boot <- d_serop %>% select(clusterid,antigenf,arm,agem_start_under6,seropos)
mu_serop_boot <- foreach(bi = 1:nreps, .combine = rbind) %dopar% {
  set.seed(bi)
  # if(bi %% 50 == 0) cat(bi,"\n") else cat(".")
  
  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_serop_boot, by = "clusterid")
  di %>%
  group_by(antigenf,arm,agem_start_under6) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c("antigenf","agem_start_under6"), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)

}
```


```{r mean seroprev bootstrap cis by age}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_serop_ci <- mu_serop_boot %>%
  group_by(antigenf,agem_start_under6) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025),
         mean_p_ub = quantile(placebo, probs = 0.975),
         mean_a_lb = quantile(azithro, probs = 0.025),
         mean_a_ub = quantile(azithro, probs = 0.975),
         diff_bmean = mean(diff),
         diff_bse = sd(diff),
         diff_lb = quantile(diff, probs = 0.025),
         diff_ub = quantile(diff, probs = 0.975)) %>%
  select(antigenf,agem_start_under6,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_serop2 <- mu_serop %>%
  left_join(mu_serop_ci, by = c("antigenf","agem_start_under6"))
  
```

### Estimate interaction p-values

Estimate interaction p-values on the additive scale using an ANOVA on cluster level means.

```{r seroprev p-values for interaction age}

#-----------------------------
# test for interaction on
# the additive scale using
# ANOVA F-test on cluster level means
# See Hayes and Moulten 2017, 
# section 10.8 (p241)
#-----------------------------
clmeans_serop <- d_serop %>% 
    group_by(antigenf,arm,agem_start_under6,clusterid) %>%
    summarize(seroprev = mean(seropos), .groups = "keep")

serop_age_Ftest <- foreach(abi = levels(d_serop$antigenf), .combine = rbind) %do% {
    # restrict to a particular antigen
    # summarize cluster level means by phase and arm
    di <- clmeans_serop %>% filter(antigenf == abi) 
  
    # ANOVA F-test on cluster level means
    # See Hayes and Moulten 2017, section 10.8 (p241)
    fiti <- lm(seroprev ~ arm*agem_start_under6, data=di)
    intFP <- as.numeric(anova(fiti)[3,4:5])
    res <- data.frame(antigenf=abi,agem_start_under6="Yes",intF=intFP[1],intP=intFP[2])
    
  }

#-----------------------------
# within pathogen group calc.
# Benjamini-Hochberg adjusted
# p-values, allowing for a 5% FDR.
#-----------------------------
serop_age_Ftest$intPbh <- c(p.adjust(serop_age_Ftest$intP[1:6]), 
                            p.adjust(serop_age_Ftest$intP[7:15])
                            )

```


### Table of results
```{r seroprev means by age table}
mu_serop_table <- mu_serop2 %>%
  ungroup() %>%
  left_join(serop_age_Ftest, by = c("antigenf","agem_start_under6")) %>%
  mutate(diff_ci = paste0(sprintf("%1.2f",diff)," (",sprintf("%1.2f",diff_lb),", ", sprintf("%1.2f",diff_ub),")")) %>%
  select(antigenf,agem_start_under6, placebo_n, placebo, azithro_n, azithro, diff_ci, intP,intPbh)

knitr::kable(mu_serop_table,
             col.names = c("Pathogen, antigen","≤6 months at trial start", "N","Proportion seropositive","N","Proportion seropositive","Difference (95% CI)","P-value*","Adj P-value†"), 
             digits = c(0,0,0,2,0,2,0,3,3),
             align = "lccccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 2, "Placebo" = 2, "Azithromycin" = 2,  " " = 3)) %>%
  kableExtra::group_rows(group_label = "Bacteria and protozoa", start_row = 1, end_row = 12) %>%
  kableExtra::group_rows(group_label = "Malaria", start_row = 13, end_row = 30) %>%
  footnote(
    symbol=c("P-value for additive interaction by study phase.", "P-value for additive interaction, adjusted for a 5% False Discovery Rate using the Benjamini-Hochberg procedure.")
  )

```



### Figure of results

```{r serop diff age figure bacteria protozoa}
#------------------------------
# plot the difference between
# arms by phase
#------------------------------

# format the output for plotting
mu_serop_plot2 <- mu_serop2 %>%
  ungroup() #%>%
  # mutate(antigenf = factor(antigenf, levels = rev(levels(mu_mfi2$antigenf))))
mu_serop_plot2$group = "bacteria\nand\nprotozoa"
mu_serop_plot2$group[grep("P. ",mu_serop_plot2$pathogen)] <- "malaria"

# make the figure
pcols <- brewer.pal(9,"PuBuGn")[7:8]
plot_serop_diff_age_enterics <- ggplot(data = mu_serop_plot2 %>% filter(group!="malaria"), aes(x = agem_start_under6, y = diff, color = agem_start_under6, fill = agem_start_under6)) +
  facet_wrap(antigenf~., ncol=3, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.3,0.3,by=0.1), labels = sprintf("%1.0f",100*seq(-0.3,0.3,by=0.1))) +
  scale_color_manual(values = pcols) +
  labs(x = "Under age 6 months at trial start", y = "Difference in seroprevalence (%)\nazithromycin - placebo", tag = "a") +
  coord_cartesian(ylim = c(-0.3,0.3)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )

```

```{r seroprev diff age figure malaria}
#------------------------------
# plot the difference between
# arms by phase
#------------------------------
pcols <- brewer.pal(9,"PuBuGn")[7:8]
plot_serop_diff_age_malaria <- ggplot(data = mu_serop_plot2 %>% filter(group=="malaria"), aes(x = agem_start_under6, y = diff, color = agem_start_under6, fill = agem_start_under6)) +
  facet_wrap(antigenf~., ncol=3, nrow = 3) +
  geom_hline(yintercept = 0, color = "gray60") +
  geom_errorbar(aes(ymin = diff_lb, ymax = diff_ub),width = 0.2,alpha=1) +
  geom_point(alpha=1) +
  scale_y_continuous(breaks = seq(-0.3,0.3,by=0.1), labels = sprintf("%1.0f",100*seq(-0.3,0.3,by=0.1))) +
  scale_color_manual(values = pcols) +
  labs(x = "Under age 6 months at trial start", y = "Difference in seroprevalence (%)\nazithromycin - placebo", tag = "b") +
  coord_cartesian(ylim = c(-0.2,0.2)) +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold")
  )


```

```{r age composite figure, fig.width = 10, fig.height = 6}
plot_serop_comp_age <- grid.arrange(plot_serop_diff_age_enterics,plot_serop_diff_age_malaria,ncol=2,nrow=1, widths = c(1,1))

ggsave(here("figures","mordor-ab-diff-seroprev-by-age-group.pdf"),plot_serop_comp_age,device = "pdf",width = 13, height = 5)
```


# Session Info
```{r session info}
sessionInfo()
```
