==> R/16-mordor-ab-pf-infection-comparison.Rmd <==
---
title: "Effect of biannual azithromycin distribution on antibody responses to malaria, bacterial, and protozoan pathogens in Niger"
subtitle: "Compare community malaria parasitemia and P. falciparum seroprevalence, and stratified analysis by baseline parasitemia"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Examine the association between community level seroprevalence and malaria parasitemia (by thick smear microscopy), a pathogen with independent measures. Estimate separate associations for antigens that elicit a longer- and shorter-duration IgG response. There is modest correlation between measures with some evidence for saturation at high levels of seroprevalence with increases in malaria parasitemia.

Finally, estimate the effects of intervention stratified by communities with baseline malaria parasitemia (by thick smear) ≤5% versus communities with higher prevalence.  This analysis was not pre-specified. It was suggested during peer review as a way to interrogate the supposition that IgG might be a more appropriate outcome measure in lower transmission settings compared to settings with extremely high transmission, where it becomes saturated.  There is modest support for this hypothesis based on this exploratory analysis, below.

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","mordor-ab-Config.R"))

#----------------------------
# source the functions file
# includes reused functions
#----------------------------
source(here("R","mordor-ab-Functions.R"))
```



# Load the antibody data

```{r load antibody data}
#-------------------------------
# load the formatted analysis
# data created by
# mordor-ab-format-data.Rmd
#-------------------------------
d <- read_rds(here("data","mordor-ab-analysis-public.rds"))

#-------------------------------
# drop "_public" suffix from IDs
#-------------------------------
d1 <- d %>%
  rename(clusterid = clusterid_public, childid = childid_public)

#-------------------------------
# restrict to Pf malaria antigens
#-------------------------------
d2 <- d1 %>%
  filter( antigen %in% c("hrp2","csp","glurp","lsa1","ama","pfmsp1") ) %>%
  mutate(antigenf = factor(antigenf))

```

# Create composite seropositivity indicators

Create an indicator if a child is positive to any one of the P. falciparum antigens.

```{r composite seropositivity indicator}
#-------------------------------
# create a composite seropositivity
# indicator across all P. falciparum
# antibody responses
#-------------------------------
d3 <- d2 %>%
  group_by(childid,phase) %>%
  mutate(seropos = max(seropos),
         ) %>%
  arrange(clusterid,childid,phase) %>%
  slice(1) %>%
  mutate(antigenf = "P. falciparum (MSP-1, AMA1, GLRUP-Ro, LSA1, CSP, HRP2)") %>%
  select(clusterid,phase,arm,childid,agem,sex,malaria_thicksmear,dbsid,pathogen,antigenf,seropos)

#-------------------------------
# create a composite seropositivity
# indicator across all long-lived
# P. falciparum antibody responses
#-------------------------------
d3long <- d2 %>%
  filter(antigen %in% c("pfmsp1","ama")) %>%
  group_by(childid,phase) %>%
  mutate(seropos = max(seropos)) %>%
  arrange(clusterid,childid,phase) %>%
  slice(1) %>%
  mutate(antigenf = "P. falciparum (MSP-1, AMA1)") %>%
  select(clusterid,phase,arm,childid,agem,sex,malaria_thicksmear,dbsid,pathogen,antigenf,seropos)

#-------------------------------
# create a composite seropositivity
# indicator across all short-lived
# P. falciparum antibody responses
#-------------------------------
d3short <- d2 %>%
  filter(antigen %in% c("glurp","lsa1","csp","hrp2")) %>%
  group_by(childid,phase) %>%
  mutate(seropos = max(seropos)) %>%
  arrange(clusterid,childid,phase) %>%
  slice(1) %>%
  mutate(antigenf = "P. falciparum (GLRUP-Ro, LSA1, CSP, HRP2)") %>%
  select(clusterid,phase,arm,childid,agem,sex,malaria_thicksmear,dbsid,pathogen,antigenf,seropos)


```

# Pf Seroprevalence vs malaria parasitemia

## All antigens
```{r infection v ab all antigens}

#-----------------------------
# count N and estimate 
# seroprevalence by community
# averaged over all phases
#
# exclude children <12 months
# due to maternal IgG contributions
#-----------------------------
mu_p <- d3 %>%
  group_by(pathogen,antigenf,clusterid,arm) %>%
  filter(agem>=12) %>%
  mutate(nomiss_ab = ifelse(!is.na(seropos),1,0),
         nomiss_ts = ifelse(!is.na(malaria_thicksmear),1,0)) %>%
  summarize(nobs_ab = sum(nomiss_ab), 
            npos_ab = sum(seropos, na.rm=TRUE),
            nobs_ts = sum(nomiss_ts), 
            npos_ts = sum(malaria_thicksmear, na.rm=TRUE),
            seroprev = mean(seropos),
            pfprev = mean(malaria_thicksmear, na.rm=TRUE),
            .groups = "keep") 
#-----------------------------
# estimate exact binomial
# 95% CIs for community means
#-----------------------------
# estimate exact binomial CIs for seroprevalence
mu_p <- mu_p %>% 
  rowwise() %>% 
  mutate(seroprev_ci = list(enframe(binom.test(x=npos_ab, n=nobs_ab, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(seroprev_ci) %>% 
    spread(name, value) %>% 
    rename("seroprev_lb" = "1", "seroprev_ub" = "2") %>%
  # identify malaria antigens vs others
    mutate(group = ifelse(substr(pathogen,1,2)=="P.","malaria","bacteria\nand\nprotozoa"))

# estimate exact binomial CIs for prevalence
mu_p <- mu_p %>% 
  rowwise() %>% 
  mutate(pfprev_ci = list(enframe(binom.test(x=npos_ts, n=nobs_ts, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(pfprev_ci) %>% 
    spread(name, value) %>% 
    rename("pfprev_lb" = "1", "pfprev_ub" = "2")

#-----------------------------
# print the total Ns included
# in the analysis for each
# measure of malaria
#-----------------------------
(sum(mu_p$nobs_ab))
(sum(mu_p$nobs_ts))

```


Plot the relationship between community level malaria parasitemia and _P. falciparum_ IgG seroprevalence. The correlations are estimated using Spearman rank correlation and the line is a locally weighted regression smoother, trimmed to avoid edge effects. 

The first plot includes all antigens, and the later plots are stratified by shorter- and longer-duration IgG responses.

```{r pf infection vs seroprev all antigens scatter, message=FALSE, warning=FALSE}

# Loess fit, trimmed to avoid edge effects
lfit <- loess(seroprev ~pfprev, data=mu_p)
mu_p$plfit <- predict(lfit)
mu_p$plfit[mu_p$pfprev<0.023 | mu_p$pfprev>0.2 ] <- NA

# spearman correlation coefficient
( spcor <- cor.test(mu_p$pfprev,mu_p$seroprev,method="spearman", exact = TRUE) )
spcor_print <- sprintf("%1.2f",spcor$estimate)


plot_serop_vs_mp <- ggplot(data=mu_p, aes(x = pfprev, color = arm)) +
  geom_point(aes(y= seroprev),alpha=0.8) +
  # geom_smooth(data=mu_p %>% filter(pfprev>0.01 & pfprev<0.1165),aes(y=seroprev),method = loess, se = FALSE, color="gray40",lwd=0.5) +
  geom_line(aes(y=plfit), color="gray40",lwd=0.5) +
  annotate("text",x=0.01,y=0.95,label=paste("rho ==",deparse(spcor_print)),parse=TRUE, hjust=0)+
  annotate("text",x=0.04,y=0.45,label="azithromycin",hjust=0,color=cbpal[6])+
  annotate("text",x=0.17,y=0.65,label="placebo",hjust=0,color=cbpal[2])+
  scale_y_continuous(breaks=seq(0,1,by=0.1),labels = sprintf("%1.0f",seq(0,1,by=0.1)*100)) + 
  scale_x_continuous(breaks=seq(0,0.3,by=0.05),labels = sprintf("%1.0f",seq(0,0.3,by=0.05)*100)) +
  scale_color_manual(values=cbpal[c(2,6)]) +
  labs(x = "malaria parasitemia (%)", y=expression(paste(italic("P. falciparum "),"seroprevalence (%)")), title=expression(paste("All ",italic("P. falciparum "),"antigens")),subtitle="MSP-1, AMA1, GLRUP-Ro, LSA1, CSP, HRP2",tag="a") +
  coord_cartesian(xlim=c(0,0.3), ylim=c(0,1))+
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold")
  )

plot_serop_vs_mp

# ggsave(filename = here("figures","mordor-ab-serop-vs-mp-overall.pdf"),plot=plot_serop_vs_mp,device="pdf",height=5,width=5)
```


## Long duration antigens
```{r infection v ab long antigens}

#-----------------------------
# count N and estimate 
# seroprevalence by community
# averaged over all phases
#
# exclude children <12 months
# due to maternal IgG contributions
#-----------------------------
mu_plong <- d3long %>%
  group_by(pathogen,antigenf,clusterid,arm) %>%
  filter(agem>=12) %>%
  mutate(nomiss_ab = ifelse(!is.na(seropos),1,0),
         nomiss_ts = ifelse(!is.na(malaria_thicksmear),1,0)) %>%
  summarize(nobs_ab = sum(nomiss_ab), 
            npos_ab = sum(seropos, na.rm=TRUE),
            nobs_ts = sum(nomiss_ts), 
            npos_ts = sum(malaria_thicksmear, na.rm=TRUE),
            seroprev = mean(seropos),
            pfprev = mean(malaria_thicksmear, na.rm=TRUE),
            .groups = "keep") 
#-----------------------------
# estimate exact binomial
# 95% CIs for community means
#-----------------------------
# estimate exact binomial CIs for seroprevalence
mu_plong <- mu_plong %>% 
  rowwise() %>% 
  mutate(seroprev_ci = list(enframe(binom.test(x=npos_ab, n=nobs_ab, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(seroprev_ci) %>% 
    spread(name, value) %>% 
    rename("seroprev_lb" = "1", "seroprev_ub" = "2") %>%
  # identify malaria antigens vs others
    mutate(group = ifelse(substr(pathogen,1,2)=="P.","malaria","bacteria\nand\nprotozoa"))

# estimate exact binomial CIs for prevalence
mu_plong <- mu_plong %>% 
  rowwise() %>% 
  mutate(pfprev_ci = list(enframe(binom.test(x=npos_ts, n=nobs_ts, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(pfprev_ci) %>% 
    spread(name, value) %>% 
    rename("pfprev_lb" = "1", "pfprev_ub" = "2")

```


```{r pf infection vs seroprev long scatter, message=FALSE, warning = FALSE}

# Loess fit, trimmed to avoid edge effects
lfitlong <- loess(seroprev ~pfprev, data=mu_plong)
mu_plong$plfit <- predict(lfitlong)
mu_plong$plfit[mu_plong$pfprev<0.023 | mu_plong$pfprev>0.2 ] <- NA

# spearman correlation coefficient
( spcorlong <- cor.test(mu_plong$pfprev,mu_plong$seroprev,method="spearman", exact = TRUE) )
spcorlong_print <- sprintf("%1.2f",spcorlong$estimate)


plot_serop_vs_mp_long <- ggplot(data=mu_plong, aes(x = pfprev, color = arm)) +
  geom_point(aes(y = seroprev),alpha=0.8) +
  # geom_smooth(method = "loess", se = FALSE, color="gray40",lwd=0.5) +
  geom_line(aes(y=plfit), color="gray40",lwd=0.5) +
  annotate("text",x=0.01,y=0.95,label=paste("rho ==",deparse(spcorlong_print)),parse=TRUE, hjust=0)+
  annotate("text",x=0.04,y=0.42,label="azithromycin",hjust=0,color=cbpal[6])+
  annotate("text",x=0.17,y=0.64,label="placebo",hjust=0,color=cbpal[2])+
  scale_y_continuous(breaks=seq(0,1,by=0.1),labels = sprintf("%1.0f",seq(0,1,by=0.1)*100)) + 
  scale_x_continuous(breaks=seq(0,0.3,by=0.05),labels = sprintf("%1.0f",seq(0,0.3,by=0.05)*100)) + 
  scale_color_manual(values=cbpal[c(2,6)]) +

  labs(x = "malaria parasitemia (%)", y=expression(paste(italic("P. falciparum "),"seroprevalence (%)")), title=expression(paste("Long-duration ",italic("P. falciparum "),"antigens")),subtitle="MSP-1, AMA1",tag="b") +
  coord_cartesian(xlim=c(0,0.3), ylim=c(0,1))+
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold")
  )

plot_serop_vs_mp_long

# ggsave(filename = here("figures","mordor-ab-serop-vs-mp-long.pdf"),plot=plot_serop_vs_mp_long,device="pdf",height=5,width=5)

```

## Short duration antigens
```{r infection v ab short antigens}

#-----------------------------
# count N and estimate 
# seroprevalence by community
# averaged over all phases
#
# exclude children <12 months
# due to maternal IgG contributions
#-----------------------------
mu_pshort <- d3short %>%
  group_by(pathogen,antigenf,clusterid,arm) %>%
  filter(agem>12) %>%
  mutate(nomiss_ab = ifelse(!is.na(seropos),1,0),
         nomiss_ts = ifelse(!is.na(malaria_thicksmear),1,0)) %>%
  summarize(nobs_ab = sum(nomiss_ab), 
            npos_ab = sum(seropos, na.rm=TRUE),
            nobs_ts = sum(nomiss_ts), 
            npos_ts = sum(malaria_thicksmear, na.rm=TRUE),
            seroprev = mean(seropos),
            pfprev = mean(malaria_thicksmear, na.rm=TRUE),
            .groups = "keep") 
#-----------------------------
# estimate exact binomial
# 95% CIs for community means
#-----------------------------
# estimate exact binomial CIs for seroprevalence
mu_pshort <- mu_pshort %>% 
  rowwise() %>% 
  mutate(seroprev_ci = list(enframe(binom.test(x=npos_ab, n=nobs_ab, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(seroprev_ci) %>% 
    spread(name, value) %>% 
    rename("seroprev_lb" = "1", "seroprev_ub" = "2") %>%
  # identify malaria antigens vs others
    mutate(group = ifelse(substr(pathogen,1,2)=="P.","malaria","bacteria\nand\nprotozoa"))

# estimate exact binomial CIs for prevalence
mu_pshort <- mu_pshort %>% 
  rowwise() %>% 
  mutate(pfprev_ci = list(enframe(binom.test(x=npos_ts, n=nobs_ts, alternative =  "two.sided", conf.level = 0.95)$conf.int))) %>% 
    unnest(pfprev_ci) %>% 
    spread(name, value) %>% 
    rename("pfprev_lb" = "1", "pfprev_ub" = "2")

```


```{r pf infection vs seroprev short scatter, message=FALSE, warning = FALSE}

# Loess fit, trimmed to avoid edge effects
lfitshort <- loess(seroprev ~pfprev, data=mu_pshort)
mu_pshort$plfit <- predict(lfitshort)
mu_pshort$plfit[mu_pshort$pfprev<0.02 | mu_pshort$pfprev>0.22 ] <- NA

# spearman correlation coefficient
(spcorshort <- cor.test(mu_pshort$pfprev,mu_pshort$seroprev,method="spearman", exact = TRUE) )
spcorshort_print <- sprintf("%1.2f",spcorshort$estimate)


plot_serop_vs_mp_short <- ggplot(data=mu_pshort, aes(x = pfprev, color = arm)) +
  geom_point(aes(y = seroprev),alpha=0.8) +
  # geom_smooth(method = "loess", se = FALSE, color="gray40",lwd=0.5) +
  geom_line(aes(y=plfit), color="gray40",lwd=0.5) +
  annotate("text",x=0.01,y=0.55,label=paste("rho ==",deparse(spcorshort_print)),parse=TRUE, hjust=0)+
  annotate("text",x=0.195,y=0.36,label="azithromycin",hjust=0,color=cbpal[6])+
  annotate("text",x=0.13,y=0.50,label="placebo",hjust=0,color=cbpal[2])+
  scale_y_continuous(breaks=seq(0,1,by=0.1),labels = sprintf("%1.0f",seq(0,1,by=0.1)*100)) + 
  scale_x_continuous(breaks=seq(0,0.3,by=0.05),labels = sprintf("%1.0f",seq(0,0.3,by=0.05)*100)) + 
  scale_color_manual(values=cbpal[c(2,6)]) +

  labs(x = "malaria parasitemia (%)", y=expression(paste(italic("P. falciparum "),"seroprevalence (%)")), title=expression(paste("Short-duration ",italic("P. falciparum "),"antigens")),subtitle="GLRUP-Ro, LSA1, CSP, HRP2",tag="c") +
  coord_cartesian(xlim=c(0,0.3), ylim=c(0,1))+
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold")
  )

plot_serop_vs_mp_short

```

## Composite figure

```{r infection vs seroprev composite figure, fig.width=6,fig.height=12, warning=FALSE}
plot_serop_vs_mp_comp <- grid.arrange(plot_serop_vs_mp,plot_serop_vs_mp_long,plot_serop_vs_mp_short,ncol=1,nrow=3)

ggsave(here("figures","mordor-ab-serop-vs-mp.pdf"),plot_serop_vs_mp_comp,device="pdf",height=15,width=5)


```



# Identify low baseline parasitemia communities

The figure below sorts communities by baseline malaria parasitemia, measured by thick smear microscopy. The horizontal line marks the threshold used to divide communities in the subgroup analysis, where approximately half of communities had baseline malaria parasitemia >5% versus lower.

```{r low pf prevalence at baseline}
#-------------------------------
# summarize baseline community 
# Pf prevalence by thicksmear
#-------------------------------
dmpbase <- d1 %>% 
  select(clusterid,arm,phase,childid,phase,malaria_thicksmear) %>%
  filter(phase==0) %>%
  group_by(clusterid,arm,childid,phase) %>%
  slice(1) %>%
  group_by(clusterid,arm,phase) %>%
  summarize(mpbase = mean(malaria_thicksmear, na.rm=TRUE), .groups = "keep") %>%
  ungroup() %>%
  arrange(mpbase) %>%
  mutate(mpbase_rank = row_number())

ggplot(data=dmpbase,aes(x=mpbase_rank,y=mpbase,color=arm)) +
  geom_hline(yintercept=0.07,color="gray40")+
  geom_point() +
  scale_y_continuous(breaks=seq(0,0.4,by=0.1),labels = sprintf("%1.0f",seq(0,0.4,by=0.1)*100)) +
  scale_x_continuous(breaks=seq(0,30,by=2)) +
  scale_color_manual(values=cbpal[c(2,6)]) +
  labs(y = "baseline malaria parasitemia (%)", x="baseline malaria parasitemia (%) rank") +
  theme_minimal()

```

# Subgroup analysis

This subgroup analysis was not prespecified, but was suggested by peer reviewers during review. Specifically, the query was whether serology might be a more sensitive endpoint for malaria effects in lower transmission communities, where the IgG seroprevalence is not approaching saturation, at or above 90%. 

In this exploratory analysis to test this hypothesis, we stratified communities by whether or not they had baseline malaria parasitemia at or below 5% (by thick smear microscopy). This cutoff was identified by inspection of the distribution of baseline parasitemia and without regard to any post-baseline measurements. The cutoff provided a natural break in the distribution with a roughly even split of communities above and below (figure above).

```{r pf subgroup analysis}
#-------------------------------
# join baseline Pf prevalence
# to the main serology analysis
# dataset and restrict to post
# baseline measurements
#-------------------------------
d4 <- d3 %>%
  left_join(dmpbase %>% select(clusterid, mpbase,mpbase_rank),by = "clusterid") %>%
  filter(phase>0) %>%
  # create an indicator of low Pf, defined as <10% prevalence at baseline
  mutate(mpbasef = factor(ifelse(mpbase<0.06,"baseline parasitemia ≤5%","baseline parasitemia >5%"),levels=c("baseline parasitemia ≤5%","baseline parasitemia >5%")))
```

## Age dependent means
```{r age-dependent seroprevalence by arm low baseline pf}

#-----------------------------
# create an indicator for 
# children living in intervention
# and control communities
#
# create a dummy indicator equal
# to 1 for all communities to zero
# out the marginal predictions
# over all community random effects
#
# create a factor variable
# for clusterid (for model)
#-----------------------------
d4 <- d4 %>%
  mutate(tr0 = ifelse(arm == "placebo",1,0),
         tr1 = ifelse(arm == "azithro",1,0),
         dummy = 1,
         clusteridf = factor(clusterid),
         phasef = factor(phase, levels=c("6","12","24","36"))
         )
d4low <- d4 %>% filter(mpbasef=="baseline parasitemia ≤5%")
d4high <- d4 %>% filter(mpbasef!="baseline parasitemia ≤5%")
#----------------------------------------
# fit a GAM model fitting separate
# splines by age for each arm
# estimate simultaneous 95% confidence 
# intervals
#----------------------------------------
fitp_age_low <- mgcv::gam(seropos ~ s(agem,bs="cr",by=tr1) + 
                        s(agem,bs="cr",by=tr0) + 
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link="cloglog"),
                       data=d4low)
  newd0 <- d4low %>%  mutate(dummy=0, tr0 = 1, tr1 = 0)
  newd1 <- d4low %>%  mutate(dummy=0, tr1 = 1, tr0 = 0)
  fit_tr0_ci <- gamCI(m=fitp_age_low,newdata=newd0,nreps=1000)
  fit_tr1_ci <- gamCI(m=fitp_age_low,newdata=newd1,nreps=1000)
  fit_ci <- bind_rows(fit_tr0_ci,fit_tr1_ci)

# convert linear predictor to prevalence
fitp_age2_low <- fit_ci %>%
  mutate(arm = ifelse(tr1==1,"azithro","placebo"),
         arm = factor(arm,levels = c("placebo","azithro"))) %>%
  mutate(fit = cloglogfn(fit),
         uprP = cloglogfn(uprP),
         lwrP = cloglogfn(lwrP),
         uprS = cloglogfn(uprS),
         lwrS = cloglogfn(lwrS),
         )  
```

```{r age-dependent seroprevalence by arm and high baseline pf}
#----------------------------------------
# fit a GAM model fitting separate
# splines by age for each arm
# estimate simultaneous 95% confidence 
# intervals
#----------------------------------------
fitp_age_high <- mgcv::gam(seropos ~ s(agem,bs="cr",by=tr1) + 
                        s(agem,bs="cr",by=tr0) + 
                         s(clusteridf,bs="re",by=dummy), 
                       family = binomial(link="cloglog"),
                       data=d4high)
  newd0 <- d4high %>%  mutate(dummy=0, tr0 = 1, tr1 = 0)
  newd1 <- d4high %>%  mutate(dummy=0, tr1 = 1, tr0 = 0)
  fit_tr0_ci <- gamCI(m=fitp_age_high,newdata=newd0,nreps=1000)
  fit_tr1_ci <- gamCI(m=fitp_age_high,newdata=newd1,nreps=1000)
  fit_ci <- bind_rows(fit_tr0_ci,fit_tr1_ci)

# convert linear predictor to prevalence
fitp_age2_high <- fit_ci %>%
  mutate(arm = ifelse(tr1==1,"azithro","placebo"),
         arm = factor(arm,levels = c("placebo","azithro"))) %>%
  mutate(fit = cloglogfn(fit),
         uprP = cloglogfn(uprP),
         lwrP = cloglogfn(lwrP),
         uprS = cloglogfn(uprS),
         lwrS = cloglogfn(lwrS),
         )  
```



```{r age-dependent seroprevalence by arm figure malaria, warning = FALSE}

#-----------------------------
# plot age-dependent seroprev
# by arm for malaria
#-----------------------------
fitp_malaria <- fitp_age2_low %>%
  bind_rows(fitp_age2_high) %>%
  mutate(antigenf = factor(antigenf))

#-----------------------------
# create a bounding box
# to shade ages used in
# the primary analysis
# and in age-dependent FOI
# estimation
#-----------------------------
d_shade_box <- data.frame(antigenf = levels(fitp_malaria$antigenf)) %>%
  mutate(minX = 12,
         maxX = 60,
         arm = "placebo",
         antigenf = factor(antigenf,levels(fitp_malaria$antigenf))
         )
d_text_labs <- data.frame(antigenf=levels(fitp_malaria$antigenf),
                          mpbasef=factor("baseline parasitemia ≤5%",levels=levels(fitp_malaria$mpbasef)),
                          x=c(36,36), 
                          y=c(0.95,0.75), 
                          arm=c("placebo","azithro"),
                          labs=c("placebo","azithromycin"), 
                          cols=cbpal[c(2,6)],
                          hjusts=c(1,0)) 

pcols <- cbpal[c(2,6)]

plot_age_seroprev_malaria <- ggplot(data = fitp_malaria , aes(x = agem, y = fit, group = arm, color = arm, fill = arm)) +
  facet_grid(cols=vars(mpbasef)) +
  # add shade box
  geom_rect(data=d_shade_box, inherit.aes = FALSE, aes(xmin = minX, xmax = maxX, ymin = 0, ymax = 1), alpha=0.1) +
  # approximate pointwise 95% CIs
  # geom_ribbon(aes(ymin = lwrP, ymax = uprP), color = NA, alpha = 0.3) +
  # spline fits
  geom_line() +
  # plot aesthetics
  geom_text(data=d_text_labs,aes(x=x,y=y,label=labs),hjust=d_text_labs$hjusts)+
  # annotate("text",x=36,y=0.95,label="placebo",color=pcols[1],hjust=1) +
  # annotate("text",x=36,y=0.75,label="azithromycin",color=pcols[2],hjust=0) +
  scale_y_continuous(breaks = seq(0,1,by=0.2), labels = sprintf("%1.0f",seq(0,1,by=0.2)*100)) +
  scale_x_continuous(breaks = seq(0,60,by=12), minor_breaks = seq(0,60,by=3)) +
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  coord_cartesian(ylim = c(0,1.04), xlim = c(0,60)) +
  labs(x = "Age, months", y = expression(paste(italic("P. falciparum")," IgG seroprevalence (%)"))) +
  theme_minimal(base_size = 14) +
  theme(
    legend.position = "none"
  )
plot_age_seroprev_malaria

ggsave(here("figures","mordor-ab-age-curves-malaria-parasitemia-stratified.png"),plot_age_seroprev_malaria, device = "png", width = 8, height = 4)
ggsave(here("figures","mordor-ab-age-curves-malaria-parasitemia-stratified.pdf"),plot_age_seroprev_malaria, device = "pdf", width = 8, height = 4)

```



## Effects on seroprevalence, by baseline parasitemia

Estimate average difference in seroprevalence stratified by baseline malaria parasitemia (≤5%, >5%). Limit the analysis to ages 12-59 months to avoid maternal IgG contributions.

```{r seroprev by arm and subgroup}
#-----------------------------
# limit to children older than
# 12 months
#-----------------------------
d5 <- d4  %>% filter(agem >= 12)

#-----------------------------
# tally number of observations
# by antigen (including composites)
#-----------------------------
n_serop <- d5 %>%
  group_by(antigenf,mpbasef,arm) %>%
  mutate(nomiss = ifelse(!is.na(seropos),1,0)) %>%
  summarize(nobs = sum(nomiss), .groups = "keep") %>%
  pivot_wider(id_cols = c(antigenf,mpbasef), names_from = "arm", values_from = nobs) %>%
  rename(placebo_n = placebo,
         azithro_n = azithro)

#-----------------------------
# estimate seroprevalence by arm
# and difference in seroprevalence
# within subgroups
#-----------------------------
mu_serop <- d5 %>%
  group_by(antigenf,mpbasef,arm) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c(antigenf,mpbasef), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo)

#-----------------------------
# merge means onto the Ns
#-----------------------------
mu_serop <- left_join(n_serop, mu_serop, by=c("antigenf","mpbasef"))
```


```{r bootstrap mean seropos diff}
#-----------------------------
# bootstrap resample communities
# with replacement to estimate
# 95% CIs for difference in seroprev
# stratified by baseline Pf prevalence
#-----------------------------
set.seed(1235)
nreps <- 1000
ncl <- length(unique(d5$clusterid))
bsamp <- matrix(sample(unique(d5$clusterid),size = ncl*nreps, replace = TRUE),nrow = ncl,ncol = nreps)
d_serop_boot <- d5 %>% ungroup() %>% select(clusterid,antigenf,mpbasef,arm,seropos)
mu_serop_boot <- foreach(bi = 1:nreps, .combine = rbind) %do% {
  
  if(bi %% 50 == 0) cat(bi,"\n") else cat(".")
  
  di <- left_join(data.frame(clusterid = bsamp[,bi], stringsAsFactors = FALSE),
                  d_serop_boot, by = "clusterid")
  di %>%
  group_by(antigenf,mpbasef,arm) %>%
  summarize(mean_serop = mean(seropos), .groups = "keep") %>%
  pivot_wider(id_cols = c(antigenf,mpbasef), names_from = "arm", values_from = mean_serop) %>%
  mutate(diff = azithro - placebo, 
         bootrep = bi)

}

```

```{r mean seroprev bootstrap cis}
#-----------------------------
# calculate percentile 95%
# CIs from the bootstrap
# distribution merge to estimates
#-----------------------------
mu_serop_ci <- mu_serop_boot %>%
  group_by(antigenf,mpbasef) %>%
  mutate(mean_p_lb = quantile(placebo, probs = 0.025,na.rm=TRUE),
         mean_p_ub = quantile(placebo, probs = 0.975,na.rm=TRUE),
         mean_a_lb = quantile(azithro, probs = 0.025,na.rm=TRUE),
         mean_a_ub = quantile(azithro, probs = 0.975,na.rm=TRUE),
         diff_bmean = mean(diff,na.rm=TRUE),
         diff_bse = sd(diff,na.rm=TRUE),
         diff_lb = quantile(diff, probs = 0.025,na.rm=TRUE),
         diff_ub = quantile(diff, probs = 0.975,na.rm=TRUE)) %>%
  select(antigenf,mpbasef,starts_with("mean_"), starts_with("diff_")) %>%
  slice(1)
  
mu_serop2 <- mu_serop %>%
  left_join(mu_serop_ci, by = c("antigenf","mpbasef"))
  
```


```{r seroprev means table}
mu_serop_table <- mu_serop2 %>%
  ungroup() %>%
  mutate(diff_ci = paste0(sprintf("%1.2f",diff)," (",sprintf("%1.2f",diff_lb),", ", sprintf("%1.2f",diff_ub),")")
         ) %>%
  select(mpbasef,azithro_n,placebo_n,azithro,placebo, diff_ci)

knitr::kable(mu_serop_table,
             caption = "Proportion of children in MORDOR Niger ages 12-59 months who were seropositive to P. falciparum antigens, stratified by community-level baseline malaria parasitemia prevalence. Confidence intervals estimated with a bootstrap resample of communities (1,000 iterations).",
             col.names = c("Baseline Malaria Parasitemia","Azithromycin", "Placebo","Azithromycin","Placebo","Difference (95% CI)"), 
             digits = c(0,0,0,2,2,0),
             align = "lccc") %>%
  kable_styling(bootstrap_options = c("striped","condensed"),full_width = TRUE) %>%
  add_header_above(c(" " = 1,"N measured" = 2,  "Proportion\nSeropositive" = 2,  " " = 1))

```

Estimate interaction p-values on the additive scale using an ANOVA on cluster level means.

```{r seroprev p-values for interaction phase}

#-----------------------------
# test for interaction on
# the additive scale using
# ANOVA F-test on cluster level means
# See Hayes and Moulten 2017, 
# section 10.8 (p241)
#-----------------------------
clmeans_serop <- d5 %>% 
    group_by(arm,mpbasef,clusterid) %>%
    summarize(seroprev = mean(seropos), .groups = "keep")

fiti <- lm(seroprev ~ arm*mpbasef, data=clmeans_serop)
anova(fiti)

```


```{r FOI estimates low malaria parasitemia}

d5low <- d5 %>%
  filter(mpbasef=="baseline parasitemia ≤5%")

set.seed(123)
gfit_low <- gam(seropos ~ s(agem, bs="cr") + arm +
              s(clusteridf, bs = "re",by=dummy), 
            data=d5low,
                family=binomial(link="cloglog"))
gsum_low <- summary(gfit_low)
    # estimate average seroconversion rate by arm
    # use a parametric bootstrap of the posterior VCOV matrix to get inference
    newd0 <- data.frame(agem=c(min(d5low$agem),max(d5low$agem)), arm = "placebo", dummy=0, clusteridf = "community-1")
    newd1 <- data.frame(agem=c(min(d5low$agem),max(d5low$agem)), arm = "azithro", dummy=0, clusteridf = "community-1")
    lambda0 <- avgFOI(gfit_low, newdata=newd0, a1=min(d5low$agem),a2=max(d5low$agem),nreps=10000)
    lambda1 <- avgFOI(gfit_low, newdata=newd1, a1=min(d5low$agem),a2=max(d5low$agem),nreps=10000)
    # estimate HR and 95% CI from the model from the model 
    hr <- as.numeric(exp(gfit_low$coefficients[2]))
    loghr_se <-as.numeric(gsum_low$se[2])
    # return the results
foi_ests_low <- data.frame(antigenf = unique(d5low$antigenf),
                           mpbasef = unique(d5low$mpbasef),
                      lambda0 = lambda0$mufoi,
                      lambda0_se = lambda0$mufoi_se,
                      lambda0_min95 = lambda0$mufoi_lb,
                      lambda0_max95 = lambda0$mufoi_ub,
                      lambda1 = lambda1$mufoi,
                      lambda1_se = lambda1$mufoi_se,
                      lambda1_min95 = lambda1$mufoi_lb,
                      lambda1_max95 = lambda1$mufoi_ub,
                      hr = hr, 
                      loghr_se = loghr_se,
                      hr_min95 = exp(log(hr) - 1.96*loghr_se),
                      hr_max95 = exp(log(hr) + 1.96*loghr_se)
                      )

foi_ests_tab_low <- foi_ests_low %>%
  mutate(lambda0c = paste0(sprintf("%1.2f",lambda0*12)," (",
                            sprintf("%1.2f",lambda0_min95*12),", ", 
                            sprintf("%1.2f",lambda0_max95*12),")"),
         lambda1c = paste0(sprintf("%1.2f",lambda1*12)," (", 
                            sprintf("%1.2f",lambda1_min95*12),", ", 
                            sprintf("%1.2f",lambda1_max95*12),")"),
         hrc = paste0(sprintf("%1.2f",hr)," (",
                       sprintf("%1.2f",hr_min95),", ",
                       sprintf("%1.2f",hr_max95),")")
         )

```

```{r FOI estimates high malaria parasitemia}

d5high <- d5 %>%
  filter(mpbasef=="baseline parasitemia >5%")

set.seed(123)
gfit_high <- gam(seropos ~ s(agem, bs="cr") + arm +
              s(clusteridf, bs = "re",by=dummy), 
            data=d5high,
                family=binomial(link="cloglog"))
gsum_high <- summary(gfit_high)
    # estimate average seroconversion rate by arm
    # use a parametric bootstrap of the posterior VCOV matrix to get inference
    newd0 <- data.frame(agem=c(min(d5high$agem),max(d5high$agem)), arm = "placebo", dummy=0, clusteridf = "community-4")
    newd1 <- data.frame(agem=c(min(d5high$agem),max(d5high$agem)), arm = "azithro", dummy=0, clusteridf = "community-4")
    lambda0 <- avgFOI(gfit_high, newdata=newd0, a1=min(d5high$agem),a2=max(d5high$agem),nreps=10000)
    lambda1 <- avgFOI(gfit_high, newdata=newd1, a1=min(d5high$agem),a2=max(d5high$agem),nreps=10000)
    # estimate HR and 95% CI from the model from the model 
    hr <- as.numeric(exp(gfit_high$coefficients[2]))
    loghr_se <-as.numeric(gsum_high$se[2])
    # return the results
foi_ests_high <- data.frame(antigenf = unique(d5high$antigenf),
                            mpbasef = unique(d5high$mpbasef),
                      lambda0 = lambda0$mufoi,
                      lambda0_se = lambda0$mufoi_se,
                      lambda0_min95 = lambda0$mufoi_lb,
                      lambda0_max95 = lambda0$mufoi_ub,
                      lambda1 = lambda1$mufoi,
                      lambda1_se = lambda1$mufoi_se,
                      lambda1_min95 = lambda1$mufoi_lb,
                      lambda1_max95 = lambda1$mufoi_ub,
                      hr = hr, 
                      loghr_se = loghr_se,
                      hr_min95 = exp(log(hr) - 1.96*loghr_se),
                      hr_max95 = exp(log(hr) + 1.96*loghr_se)
                      )


foi_ests_tab_high <- foi_ests_high %>%
  mutate(lambda0c = paste0(sprintf("%1.2f",lambda0*12)," (",
                            sprintf("%1.2f",lambda0_min95*12),", ", 
                            sprintf("%1.2f",lambda0_max95*12),")"),
         lambda1c = paste0(sprintf("%1.2f",lambda1*12)," (", 
                            sprintf("%1.2f",lambda1_min95*12),", ", 
                            sprintf("%1.2f",lambda1_max95*12),")"),
         hrc = paste0(sprintf("%1.2f",hr)," (",
                       sprintf("%1.2f",hr_min95),", ",
                       sprintf("%1.2f",hr_max95),")")
         )

```

```{r summary stratified scr table}
foi_ests_tab <- foi_ests_tab_low %>%
  bind_rows(foi_ests_tab_high)

knitr::kable(foi_ests_tab %>% select(mpbasef,lambda0c,lambda1c,hrc),digits=3,
             caption="Average P. falciparum seroconversion rate (SCR) per year estimated from age-structured seroprevalence, stratified by baseline parasitemia",
             align = "lccc",
             col.names = c("Malaria parasitemia","Placebo SCR (95% CI)","Azithromycin SCR (95% CI)", "HR (95% CI)"),
             row.names = FALSE)  %>%
  kable_styling(bootstrap_options = "striped",full_width = TRUE)

```

# Session Info
```{r session info}
sessionInfo()
```
